+++ Sources/epoch_config/Configs/CfgBuildingLootPos.hpp
@@ -1,0 +1,6098 @@
+class CfgBuildingLootPos
+{
+	class Default
+	{
+		fridgePos[] = {};
+		shelfPos[] = {};
+		pelicanPos[] = {};
+		wardrobePos[] = {};
+		bedPos[] = {};
+		couchPos[] = {};
+		cookerPos[] = {};
+		tablePos[] = {};
+		lockerPos[] = {};
+		toolRackPos[] = {};
+		shoeboxPos[] = {};
+		lootBiasPos = 25;
+		lootType = "normal";
+	};
+	// test
+	class Land_Kiosk_redburger_F : Default {
+		fridgePos[] = {
+				{ { 1.64453, 1.43945, -2.897 }, 89.9032 }
+		};
+		cookerPos[] = {
+				{ { 1.57813, -0.261719, -2.89732 }, 90.9244 }
+		};
+		filingPos[] = {
+				{ { -1.65625, 1.67773, -2.89687 }, 272.463 }
+		};
+		shoeboxPos[] = {
+				{ { -0.275391, -1.92188, -2.07339 }, 358 }
+		};
+		freezerPos[] = {
+				{ { -1.51367, 0.0078125, -2.86339 }, 271.8 }
+		};
+	};
+	class Land_Kiosk_blueking_F : Default {
+		fridgePos[] = {
+				{ { -1.56055, 1.67383, -1.79198 }, 0.076973 }
+		};
+		cookerPos[] = {
+				{ { -1.62109, -0.296875, -1.79261 }, 269.732 }
+		};
+		filingPos[] = {
+				{ { 1.82031, 0.972656, -1.75876 }, 91.8 },
+				{ { 1.82031, 1.55664, -1.75876 }, 91.8 }
+		};
+		shoeboxPos[] = {
+				{ { 1.86914, -1.18555, -0.645341 }, 90 }
+		};
+		freezerPos[] = {
+				{ { 1.54297, -0.0410156, -1.75876 }, 91.8 }
+		};
+	};
+	class Land_Kiosk_gyros_F : Default {
+		fridgePos[] = {
+				{ { 1.64258, 0.197266, -2.0025 }, 89.8823 }
+		};
+		cookerPos[] = {
+				{ { 1.5, 1.62305, -2.0029 }, 89.3177 }
+		};
+		filingPos[] = {
+				{ { -1.52344, 1.62109, -2.00235 }, 296.665 }
+		};
+		pelicanPos[] = {
+				{ { 2.91016, -0.0703125, 0.187521 }, 263.525 }
+		};
+		shoeboxPos[] = {
+				{ { -1.89844, -1.08008, -1.17894 }, 91.8 }
+		};
+		freezerPos[] = {
+				{ { -1.57031, 0.199219, -1.96894 }, 269.8 }
+		};
+	};
+	class Land_Kiosk_papers_F : Default {
+		shelfPos[] = {
+				{ { 1.36523, 1.79492, -2.0024 }, 1.00576 }
+		};
+		chairPos[] = {
+				{ { -1.47852, 1.27148, -1.96894 }, 306.8 }
+		};
+		filingPos[] = {
+				{ { 1.69336, 0.0625, -2.0024 }, 86.0276 },
+				{ { -1.68164, -0.347656, -2.00241 }, 266.904 },
+				{ { 1.71484, -0.404297, -2.0024 }, 90.176 }
+		};
+		shoeboxPos[] = {
+				{ { 1.81055, -1.86133, -0.808939 }, 338.8 }
+		};
+	};
+	class Land_Research_HQ_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 0.380859, 4.1875, -3.29264 }, 2.04878 },
+				{ { -1.03125, -6.27734, -3.17222 }, 181.183 },
+				{ { 8.52148, 3.01367, -3.17222 }, 90.053 }
+		};
+		fridgePos[] = {
+				{ { 5.52148, 6.21094, -3.17237 }, 0.0829086 }
+		};
+		bedPos[] = {
+				{ { 0.450195, 0.876953, -3.29266 }, 0 }
+		};
+		couchPos[] = {
+				{ { 2.89746, 0.804688, -3.29266 }, 181.475 }
+		};
+		cookerPos[] = {
+				{ { 6.66992, 6.10547, -3.17264 }, 358.477 }
+		};
+		chairPos[] = {
+				{ { -3.59766, 3.67578, -3.29266 }, 325.551 }
+		};
+		filingPos[] = {
+				{ { -1.17676, 4.13086, -3.29262 }, 358.166 },
+				{ { -2.75781, -6.06445, -3.1722 }, 177.048 },
+				{ { -3.52148, -5.88281, -3.17221 }, 202.305 },
+				{ { 6.77246, -3.36914, -0.692615 }, 84.9868 },
+				{ { 8.46094, 3.97266, -3.17206 }, 89.6625 }
+		};
+		pelicanPos[] = {
+				{ { -0.431641, 0.957031, -0.767684 }, 213.601 },
+				{ { 4.77832, -4.33398, -3.29265 }, 149.079 }
+		};
+		lockerPos[] = {
+				{ { -3.9668, -4.05664, -3.29265 }, 268.178 },
+				{ { 7.18848, 0.742188, -3.29265 }, 89.2752 }
+		};
+		palletPos[] = {
+				{ { 0.328125, -3.94141, -0.740771 }, 174 },
+				{ { 2.7959, -6.25977, -3.94077 }, 260 }
+		};
+		freezerPos[] = {
+				{ { 2.64355, 5.34766, -3.14077 }, 271 }
+		};
+		cabinetPos[] = {
+				{ { -4.21973, -1.09131, -1.87077 }, 0 },
+				{ { 2.64795, -5.21143, -1.87077 }, 270 }
+		};
+	};
+
+
+
+	class Land_Research_house_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 0.408203, 3.9668, 0.00387025 }, 0.47818 },
+				{ { -2.37891, 3.47461, 0.0038693 }, 272.048 }
+		};
+		fridgePos[] = {
+				{ { 2.17383, 1.02539, -0.132167 }, 91.4155 }
+		};
+		filingPos[] = {
+				{ { -2.18359, 2.4082, 0.00388861 }, 270 },
+				{ { -0.384766, 3.89453, 0.00388622 }, 10.991 }
+		};
+		pelicanPos[] = {
+				{ { -1.88477, 0.826172, -0.132179 }, 233.568 }
+		};
+		lockerPos[] = {
+				{ { 2.18555, 3.36523, 0.00385499 }, 88.1972 }
+		};
+		palletPos[] = {
+				{ { -1.45508, -1.11914, -0.671438 }, 174 }
+		};
+		cabinetPos[] = {
+				{ { -0.443359, -0.0527344, 1.29856 }, 270 }
+		};
+	};
+	class Land_spp_Tower_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 2.79297, -1.28906, -14.9882 }, 88.9351 },
+				{ { -2.72266, -2.93164, -24.3965 }, 179.88 }
+		};
+		bedPos[] = {
+				{ { -1.89648, -0.994141, 10.1607 }, 93.8027 }
+		};
+		couchPos[] = {
+				{ { 1.39258, -2.8418, -14.9885 }, 89.9995 }
+		};
+		chairPos[] = {
+				{ { 1.55859, -1.35156, 10.1608 }, 177.348 },
+				{ { -1.16016, -3.20703, -14.9884 }, 187.768 }
+		};
+		filingPos[] = {
+				{ { -3.67969, -2.19141, -24.3019 }, 180.033 },
+				{ { 4.3125, 0.748047, 10.1609 }, 87.6986 },
+				{ { -4.07617, -3.69531, 23.799 }, 181.064 }
+		};
+		pelicanPos[] = {
+				{ { -1.03711, -2.03906, 25.6749 }, 180.506 },
+				{ { 4.40625, -2.83008, 23.7991 }, 113.987 }
+		};
+		tablePos[] = {
+				{ { -0.00390625, 0.507813, -14.9885 }, 0.00285339 }
+		};
+		lockerPos[] = {
+				{ { -2.67773, 0.0273438, -14.9884 }, 270.033 }
+		};
+		toolRackPos[] = {
+				{ { 4.57813, -0.574219, 10.1609 }, 89.1923 }
+		};
+		shoeboxPos[] = {
+				{ { 2.64648, -2.03516, 25.6749 }, 195.354 },
+				{ { -2.6582, -3.65625, -14.9885 }, 195.527 }
+		};
+		palletPos[] = {
+				{ { 1.52539, 0.0351563, 23.8031 }, 272 },
+				{ { 3.38477, 0.015625, 23.8031 }, 179 }
+		};
+		freezerPos[] = {
+				{ { 3.29102, -1.50195, 10.1633 }, 179 }
+		};
+		cabinetPos[] = {
+				{ { -3.93164, 0.230469, -22.7967 }, 0 }
+		};
+	};
+	class Land_MilOffices_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 5.89844, 0.773438, -2.89318 }, 269.647 },
+				{ { 6.08203, -0.435547, -2.89318 }, 0.742516 },
+				{ { -4.00781, -5.64648, -2.89318 }, 180.694 },
+				{ { 7.26367, -2.26563, -2.89318 }, 267.774 },
+				{ { 8.74219, -5.70313, -2.89318 }, 178.843 },
+				{ { 5.89063, 9.54297, -2.89318 }, 270.185 },
+				{ { -5.83789, 11.2949, -2.89318 }, 359.547 },
+				{ { -13.8574, -5.7207, -2.89318 }, 179.686 },
+				{ { -16.1719, -0.859375, -2.89318 }, 358.847 },
+				{ { -16.9512, 0.576172, -2.89318 }, 268.745 },
+				{ { -16.3457, 11.2871, -2.89318 }, 0.686188 }
+		};
+		fridgePos[] = {
+				{ { 6.43945, -3.44531, -2.89318 }, 89.5034 },
+				{ { 6.60547, 4.67578, -2.89318 }, 180.413 },
+				{ { -10.707, -5.51563, -2.89318 }, 178.591 }
+		};
+		bedPos[] = {
+				{ { 13.5332, 5.33789, -2.8932 }, 95.0654 }
+		};
+		couchPos[] = {
+				{ { 4.08984, 8.96289, -2.8932 }, 4.74224 },
+				{ { 15.9063, 9.83789, -2.8932 }, 357.568 }
+		};
+		wardrobePos[] = {
+				{ { 16.7813, 6.91211, -2.89308 }, 89.005 }
+		};
+		cookerPos[] = {
+				{ { -12.2871, -3.27539, -2.89359 }, 269.464 }
+		};
+		chairPos[] = {
+				{ { -6.48242, -1.26172, -2.89333 }, 30.6329 },
+				{ { -6.36328, -5.20898, -2.89321 }, 138.614 },
+				{ { -9.26953, -5.20703, -2.89319 }, 221.625 },
+				{ { 2.14844, 10.8848, -2.8932 }, 357.801 },
+				{ { 16.2461, 0.330078, -2.89319 }, 58.6309 },
+				{ { 12.3359, 10.6348, -2.8932 }, 349.887 },
+				{ { 16.2637, -3.21289, -2.89318 }, 108.767 }
+		};
+		filingPos[] = {
+				{ { -0.710938, -0.902344, -2.89316 }, 273.543 },
+				{ { -1.625, -5.49219, -2.89316 }, 172.228 },
+				{ { -2.56445, 6.4082, -2.89316 }, 86.596 },
+				{ { 7.35742, -0.585938, -2.89316 }, 269.601 },
+				{ { 6.50391, -5.62305, -2.89316 }, 103.818 },
+				{ { -1.80859, 11.1055, -2.89316 }, 353.219 },
+				{ { 10.082, -5.57422, -2.89316 }, 171.96 },
+				{ { -11.4727, -5.625, -2.89316 }, 178.889 },
+				{ { 7.61523, 11.1699, -2.89316 }, 357.944 },
+				{ { 8.19141, 11.0684, -2.89316 }, 328.836 },
+				{ { -14.7676, -0.96875, -2.89316 }, 359.13 },
+				{ { -10.7695, 11.1172, -2.89316 }, 348.322 },
+				{ { 16.793, 3.68164, -2.89316 }, 58.8697 },
+				{ { -16.8691, -5.51758, -2.89316 }, 190.272 }
+		};
+		pelicanPos[] = {
+				{ { -15.4648, -5.25586, -2.89319 }, 177.065 },
+				{ { 16.1719, 5.08984, -2.89319 }, 160.192 }
+		};
+		tablePos[] = {
+				{ { 0.990234, -5.35742, -2.8932 }, 181.381 },
+				{ { -7.78516, -3.08789, -2.8932 }, 0.00653839 },
+				{ { 16.6719, -1.39063, -2.8932 }, 90.476 },
+				{ { 14.0918, 9.58008, -2.8932 }, 90.0005 }
+		};
+		lockerPos[] = {
+				{ { 7.49805, -4.18359, -2.89319 }, 271.895 },
+				{ { 10.3281, 4.60352, -2.89319 }, 179.863 },
+				{ { -16.7461, -3.25195, -2.89319 }, 266.767 }
+		};
+		toolRackPos[] = {
+				{ { 9.29297, -0.222656, -2.89323 }, 359.898 },
+				{ { -13.0098, -3.95703, -2.89322 }, 89.093 }
+		};
+		palletPos[] = {
+				{ { -15.3184, -3.625, -2.85274 }, 272 }
+		};
+		freezerPos[] = {
+				{ { -12.2461, -4.50391, -2.85274 }, 272 }
+		};
+		cabinetPos[] = {
+				{ { 5.3125, 4.95215, -1.47264 }, 180 },
+				{ { 5.27734, -5.80957, -1.47264 }, 270 },
+				{ { -12.1123, -2.60938, -1.47264 }, 270 }
+		};
+	};
+	class Land_Airport_Tower_F : Default {
+		shelfPos[] = {
+				{ { 1.00195, 4.2207, -10.4679 }, 90 },
+				{ { 2.09375, -5.83984, -1.46719 }, 136.235 }
+		};
+		fridgePos[] = {
+				{ { -3.19922, -5.58789, -1.46735 }, 223.518 }
+		};
+		chairPos[] = {
+				{ { 0.740234, -0.0332031, -1.46723 }, 3.7747 },
+				{ { 2.19141, -1.23438, -1.46718 }, 54.1489 }
+		};
+		filingPos[] = {
+				{ { -2.14844, 1.53906, -7.46764 }, 267.225 },
+				{ { -0.503906, -3.69141, 2.22974 }, 268.915 },
+				{ { -1.09375, -3.69727, 2.2298 }, 90.7877 },
+				{ { -3.73633, -5.13086, -1.46725 }, 223.885 },
+				{ { -2.69922, -6.11133, -1.46731 }, 226.111 }
+		};
+		pelicanPos[] = {
+				{ { -0.482422, -6.20898, 3.00391 }, 187.096 },
+				{ { -0.744141, 6.49414, 1.52821 }, 348.84 }
+		};
+		lockerPos[] = {
+				{ { -3.26172, -0.542969, -1.4672 }, 314.971 }
+		};
+		palletPos[] = {
+				{ { -4.09766, 1.40625, -10.6461 }, 272 }
+		};
+		cabinetPos[] = {
+				{ { 2.48779, -0.672852, 2.53949 }, 135 },
+				{ { -2.37842, 1.64404, -9.16051 }, 0 },
+				{ { -3.85498, -0.772461, 2.53949 }, 45 }
+		};
+	};
+	class Land_Slum_House01_F : Default {
+		chairPos[] = {
+				{ { 1.20313, 1.6582, -1.07285 }, 36.9507 }
+		};
+		filingPos[] = {
+				{ { 0.351563, 2.05469, -1.08115 }, 8.25228 }
+		};
+		pelicanPos[] = {
+				{ { 1.4668, 2.32031, 1.60821 }, 184.525 }
+		};
+		toolRackPos[] = {
+				{ { -1.56641, 0.789063, -1.08133 }, 270.436 }
+		};
+		palletPos[] = {
+				{ { 0.169922, 0.214844, -1.08145 }, 272 }
+		};
+	};
+	class Land_Slum_House02_F : Default {
+		shelfPos[] = {
+				{ { 2.52148, 3.48438, -0.824517 }, 359.959 }
+		};
+		couchPos[] = {
+				{ { 2.14844, 0.236328, -0.824545 }, 0.0984116 }
+		};
+		chairPos[] = {
+				{ { 1.15234, 3.20508, -0.824523 }, 352.052 }
+		};
+		tablePos[] = {
+				{ { -0.15625, 0.738281, -0.824542 }, 268.176 }
+		};
+	};
+	class Land_Slum_House03_F : Default {
+		bedPos[] = {
+				{ { 0.166016, 0.230469, -1.07675 }, 95 }
+		};
+		chairPos[] = {
+				{ { -1.99414, 2.25586, -1.07675 }, 335 }
+		};
+		filingPos[] = {
+				{ { 3.98633, -0.507813, -1.07675 }, 95 }
+		};
+		toolRackPos[] = {
+				{ { 0.773438, 2.98633, -1.07675 }, 0 }
+		};
+		palletPos[] = {
+				{ { 3.60938, 2.14453, -1.07675 }, 272 }
+		};
+	};
+	class Land_cargo_house_slum_F : Default {
+		shelfPos[] = {
+				{ { -0.891602, -1.20117, -0.38667 }, 181.006 }
+		};
+		bedPos[] = {
+				{ { -3.16504, -0.253906, -0.386693 }, 276.075 }
+		};
+		chairPos[] = {
+				{ { 0.650391, -0.734375, -0.383496 }, 90.6667 }
+		};
+		filingPos[] = {
+				{ { -4.45898, -1.12109, -0.386626 }, 269.224 }
+		};
+	};
+	class Land_FuelStation_Build_F : Default {
+		shelfPos[] = {
+				{ { 0.415039, 1.52734, -1.36232 }, 0.956482 }
+		};
+		filingPos[] = {
+				{ { -2.31543, -1.1582, -1.3623 }, 259.317 }
+		};
+		pelicanPos[] = {
+				{ { -2.79199, 0.658203, 2.11388 }, 86.4283 }
+		};
+		shoeboxPos[] = {
+				{ { -1.99316, 0.386719, -0.361005 }, 112 }
+		};
+		freezerPos[] = {
+				{ { 2.10938, -0.701172, -1.391 }, 90 }
+		};
+	};
+	class Land_Factory_Main_F : Default {
+		shelfPos[] = {
+				{ { -3.72852, 3.85742, -6.35809 }, 178.041 },
+				{ { 5.62012, -9.38477, 2.34198 }, 89.9969 },
+				{ { -6.9209, 10.8223, -6.32809 }, 95.4476 },
+				{ { -10.208, 14.8887, -6.32809 }, 1.97624 }
+		};
+		fridgePos[] = {
+				{ { -1.76074, 8.24805, -6.35806 }, 94.2797 }
+		};
+		couchPos[] = {
+				{ { -2.44629, 5.73047, -6.35812 }, 2.06058 },
+				{ { -3.50684, 16.8828, -6.41408 }, 90.0061 }
+		};
+		cookerPos[] = {
+				{ { -5.74609, 5.52734, -6.35851 }, 183.545 }
+		};
+		chairPos[] = {
+				{ { -5.46777, 4.26367, -6.35815 }, 231.258 },
+				{ { -7.42676, 14.3906, -6.3281 }, 42.1397 }
+		};
+		filingPos[] = {
+				{ { -5.79004, 6.72266, -6.35807 }, 270.867 },
+				{ { -4.03516, -9.26758, 2.34797 }, 0.889618 },
+				{ { 5.33691, -11.1738, 2.34195 }, 75.1523 },
+				{ { 5.16602, -12.2598, 2.34194 }, 135.024 },
+				{ { -10.5898, 12.2129, -6.32807 }, 270 },
+				{ { -10.6104, 12.6699, -6.32807 }, 267.548 }
+		};
+		pelicanPos[] = {
+				{ { 2.81738, -9.37305, 2.3419 }, 2.64822 }
+		};
+		toolRackPos[] = {
+				{ { -3.79004, 9.14063, -6.35813 }, 0.489624 },
+				{ { -6.61133, 12.8359, -6.32813 }, 89.3704 }
+		};
+		palletPos[] = {
+				{ { -3.94238, -16.6523, -6.65465 }, 94 },
+				{ { 4.20313, 16.6426, -6.35465 }, 95 }
+		};
+	};
+	class Land_Airport_right_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -4.19922, -2.1543, -1.76375 }, 359.705 },
+				{ { -4.57422, 2.20508, -1.76375 }, 270.932 },
+				{ { 5.27734, -5.37891, -1.76368 }, 357.591 },
+				{ { -15.0703, -7.48633, -6.76375 }, 89.4887 },
+				{ { -14.4297, -11.0059, -6.76375 }, 270.174 },
+				{ { -15.1445, -10.3418, -6.76375 }, 89.1635 },
+				{ { -23.3359, -3.69922, -6.76375 }, 358.136 },
+				{ { -21.1836, 14.7402, -6.77192 }, 359.554 }
+		};
+		fridgePos[] = {
+				{ { -2.48633, 3.61719, -1.76375 }, 0.951996 },
+				{ { -15.4355, 13.5039, -6.76375 }, 90.2176 }
+		};
+		bedPos[] = {
+				{ { -3.82422, -10.3633, -1.76377 }, 184.477 }
+		};
+		couchPos[] = {
+				{ { 4.75195, 2.29883, -1.76378 }, 4.22821 },
+				{ { 0.705078, -6.63281, -6.76378 }, 269.802 },
+				{ { -16.1621, 7.51172, -6.76378 }, 358.918 },
+				{ { -25.6094, 1.10938, -6.76377 }, 183.238 }
+		};
+		wardrobePos[] = {
+				{ { 0.699219, -5.58203, -1.76363 }, 1.96878 }
+		};
+		cookerPos[] = {
+				{ { -0.195313, 3.53906, -1.74532 }, 0 }
+		};
+		chairPos[] = {
+				{ { -5.11133, 3.41406, -6.76377 }, 358.691 },
+				{ { -5.23047, 5.16602, -6.76375 }, 180.383 },
+				{ { 5.21094, -6.4082, -6.76376 }, 49.2284 },
+				{ { 5.05273, -13.168, -1.76375 }, 29.6497 },
+				{ { -4.29688, -14.416, -6.73171 }, 205.019 },
+				{ { 5.29102, -14.1387, -6.76376 }, 130.737 },
+				{ { 5.28906, 14.3145, -6.76377 }, 41.909 },
+				{ { -15.7715, -2.20508, -6.76376 }, 122.506 },
+				{ { -16.5605, -3.95313, -6.76376 }, 1.55649 },
+				{ { -23.498, -2.37891, -6.76377 }, 177.809 },
+				{ { -26.168, 5.07813, -6.76375 }, 276.716 }
+		};
+		filingPos[] = {
+				{ { -4.57031, -1.3418, -1.76373 }, 183.677 },
+				{ { 5.76953, -4.5918, -6.76376 }, 164.562 },
+				{ { -15.1465, -9.57227, -6.76368 }, 88.7803 },
+				{ { -24.3516, -3.79688, -6.76373 }, 359.584 }
+		};
+		pelicanPos[] = {
+				{ { 7.55078, -11.002, -6.78875 }, 217.643 },
+				{ { 4.91406, -16.5039, -3.16119 }, 5.77048 },
+				{ { -25.8477, -2.27734, -6.76375 }, 202.245 }
+		};
+		tablePos[] = {
+				{ { 2.9707, 2.35352, -1.74532 }, 278 },
+				{ { -4.41406, -6.25391, -1.76377 }, 271.909 },
+				{ { 5.57227, -10.1172, -6.76377 }, 90.0333 },
+				{ { -23.7207, 1.08594, -6.76377 }, 89.9196 }
+		};
+		lockerPos[] = {
+				{ { 3.54688, -1.14844, -6.76377 }, 179.474 },
+				{ { 5.26563, -1.10938, -6.76372 }, 179.41 },
+				{ { -14.2656, -1.27734, -6.76377 }, 269.09 },
+				{ { -14.1992, 6.67188, -6.76373 }, 269.966 },
+				{ { -14.2227, 8.41016, -6.76367 }, 268.618 },
+				{ { -26.5117, 9.03516, -6.76376 }, 269.173 },
+				{ { -26.5234, 10.9512, -6.76368 }, 270.17 },
+				{ { -26.5176, 12.9082, -6.76377 }, 270.361 }
+		};
+		toolRackPos[] = {
+				{ { 1.51953, -1.72852, -1.76373 }, 179.724 },
+				{ { -14.9082, -12.3223, -6.76378 }, 89.5845 }
+		};
+		shoeboxPos[] = {
+				{ { -8.14258, -16.0605, -3.27531 }, 278 }
+		};
+		palletPos[] = {
+				{ { 3.84766, -3.8418, -6.74531 }, 350 },
+				{ { -6.58008, 0.285156, -1.86532 }, 278 },
+				{ { -6.44922, -1.64844, -1.86532 }, 264 },
+				{ { -23.873, -11.5293, -6.74531 }, 350 },
+				{ { -23.8867, -13.6484, -6.74531 }, 86 }
+		};
+		freezerPos[] = {
+				{ { -1.31055, 3.54297, -1.74532 }, 0 }
+		};
+	};
+	class Land_Airport_center_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 11.7051, 7.16016, -2.74629 }, 359.37 }
+		};
+		couchPos[] = {
+				{ { -4.50391, -7.57422, 1.62245 }, 88.2815 },
+				{ { 11.8594, -11.0684, -2.74632 }, 2.03362 }
+		};
+		chairPos[] = {
+				{ { -12.3184, 6.87109, -2.74437 }, 280.834 },
+				{ { -12.2793, -14.4258, -2.7463 }, 272.565 }
+		};
+		filingPos[] = {
+				{ { 12.8184, 7.07422, -2.74627 }, 358.116 },
+				{ { 12.7363, -16.0586, -2.74625 }, 90.9752 }
+		};
+		pelicanPos[] = {
+				{ { -1.08398, -10.5664, 1.59369 }, 341.766 },
+				{ { -12.0918, -15.8086, -2.74628 }, 213.62 }
+		};
+		tablePos[] = {
+				{ { 12.5625, -14.0469, -2.74631 }, 92.4684 }
+		};
+		lockerPos[] = {
+				{ { 12.6758, -6.87891, -2.7463 }, 90.0581 },
+				{ { -12.5391, -7.84961, -2.7463 }, 270.451 },
+				{ { -12.541, -9.83398, -2.7463 }, 268.378 },
+				{ { -12.4824, -11.8047, -2.7463 }, 269.442 }
+		};
+	};
+	class Land_Airport_left_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 3.91211, -1.41797, -1.76377 }, 184.071 },
+				{ { -1.94336, -5.52148, -1.76367 }, 0.193413 },
+				{ { -5.31641, -4.6875, -6.76368 }, 179.512 },
+				{ { 6.14258, 13.9648, -6.76367 }, 268.974 },
+				{ { 15.1094, -8.04102, -6.76375 }, 270.202 },
+				{ { 15.6563, -14.7227, -6.76375 }, 180.679 },
+				{ { 16.9883, 14.7461, -6.76375 }, 359.573 }
+		};
+		fridgePos[] = {
+				{ { 1.42773, 3.67578, -1.76375 }, 359.297 },
+				{ { 15.7793, -2.51563, -6.76374 }, 178.661 }
+		};
+		bedPos[] = {
+				{ { -4.95898, -6.73047, -1.76377 }, 179.773 }
+		};
+		couchPos[] = {
+				{ { -4.08008, 2.0957, -1.76377 }, 241.929 },
+				{ { 3.72461, -6.98047, -1.76378 }, 4.04132 },
+				{ { 13.4414, 7.67383, -6.76378 }, 3.19656 },
+				{ { 25.5898, 12.5527, -6.76377 }, 0.251305 }
+		};
+		wardrobePos[] = {
+				{ { -1.35547, 3.14844, -1.7637 }, 62.9544 }
+		};
+		cookerPos[] = {
+				{ { 2.95117, 3.63477, -1.76408 }, 0.722954 },
+				{ { 16.668, -2.55078, -6.76418 }, 179.019 }
+		};
+		chairPos[] = {
+				{ { -0.699219, -1.14844, -1.76373 }, 179.434 },
+				{ { -2.9707, -1.18555, -1.76369 }, 181.435 },
+				{ { 4.33398, 4.2832, -6.76373 }, 89.2875 },
+				{ { -3.68945, -5.95117, -6.75249 }, 352.602 },
+				{ { 5.99023, 4.28711, -6.76371 }, 269.261 },
+				{ { -5.23438, -13.1641, -1.76377 }, 338.538 },
+				{ { 4.23047, -14.3047, -6.76375 }, 167.634 },
+				{ { -5.30859, 14.2852, -6.76372 }, 297.672 },
+				{ { 23.127, 4.48242, -6.76377 }, 209.073 },
+				{ { 22.8613, 7.83984, -6.76376 }, 313.711 },
+				{ { 26.1641, 4.21094, -6.76375 }, 135.598 },
+				{ { 26.2402, 8.27734, -6.76374 }, 47.2424 }
+		};
+		filingPos[] = {
+				{ { -0.0703125, 3.63086, -1.76371 }, 1.43237 },
+				{ { 15.2598, -11.4375, -6.76373 }, 268.46 },
+				{ { 15.2832, -11.8906, -6.76373 }, 270.436 },
+				{ { 15.1895, 14.6602, -6.76373 }, 344.677 },
+				{ { 24.4355, -2.6543, -6.76369 }, 180.398 },
+				{ { 26.6348, -2.56055, -6.76373 }, 172.362 }
+		};
+		pelicanPos[] = {
+				{ { 6.28711, 0.486328, -1.80376 }, 285.45 },
+				{ { 9.81641, -16.4297, -3.16111 }, 3.09702 }
+		};
+		tablePos[] = {
+				{ { 4.40625, 1.19727, -1.76377 }, 89.5157 },
+				{ { -5.57227, -7.76367, -6.76377 }, 270.379 },
+				{ { 26.3789, 6.25, -6.76377 }, 90.3904 }
+		};
+		lockerPos[] = {
+				{ { 3.56445, -1.11523, -6.76379 }, 179.89 },
+				{ { -5.07617, -1.12891, -6.76377 }, 178.63 },
+				{ { -0.271484, -5.82422, -6.7637 }, 0.334862 },
+				{ { 1.68945, -5.81836, -6.7637 }, 0.41259 },
+				{ { 15.3574, 5.5957, -6.76373 }, 268.91 },
+				{ { 15.2949, 7.52344, -6.76377 }, 270.217 }
+		};
+		toolRackPos[] = {
+				{ { 14.8926, -9.83789, -6.76369 }, 269.885 },
+				{ { 23.1777, -3.47266, -6.76379 }, 0.366142 }
+		};
+		palletPos[] = {
+				{ { 23.9648, -7.17188, -6.77531 }, 357 },
+				{ { 23.6523, -13.7891, -6.77531 }, 90 }
+		};
+		freezerPos[] = {
+				{ { 0.832031, -1.19531, -1.75531 }, 182 },
+				{ { 24.3145, -10.5938, -6.77531 }, 90 }
+		};
+	};
+	class Land_Hangar_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -14.3125, 1.24023, -5.40771 }, 270.363 },
+				{ { -14.2754, 6.36328, -5.4077 }, 269.988 },
+				{ { -14.5117, -19.8359, -5.4077 }, 268.904 },
+				{ { 8.80469, 23.3691, -5.4077 }, 359.987 },
+				{ { 13.6602, -21.0078, -5.40769 }, 180.14 }
+		};
+		couchPos[] = {
+				{ { 11.9082, 22.3652, -5.40773 }, 267.453 }
+		};
+		chairPos[] = {
+				{ { -14.3223, -5.03906, -5.40451 }, 270.026 }
+		};
+		filingPos[] = {
+				{ { -14.5059, -2.35352, -5.40699 }, 272.098 },
+				{ { -14.4941, -3.14453, -5.40626 }, 278.808 },
+				{ { -14.457, -3.89453, -5.40769 }, 265.43 },
+				{ { 14.4023, 19.6953, -5.40768 }, 91.5292 },
+				{ { 14.4004, 20.4277, -5.40769 }, 86.6895 },
+				{ { 9.92969, 23.1992, -5.40769 }, 352.998 }
+		};
+		pelicanPos[] = {
+				{ { 14.5156, -8.02148, -5.40772 }, 90 },
+				{ { -14.6602, 12.4863, -5.40772 }, 270 }
+		};
+		shoeboxPos[] = {
+				{ { 14.9688, -20.9824, -5.375 }, 86 }
+		};
+		palletPos[] = {
+				{ { 13.8945, 12.375, -5.375 }, 86 },
+				{ { -12.043, 22.7324, -5.375 }, 86 },
+				{ { -14.209, 22.5332, -5.375 }, 86 }
+		};
+		freezerPos[] = {
+				{ { 14.6035, 21.8516, -5.375 }, 86 }
+		};
+	};
+	class Land_CarService_F : Default {
+		shelfPos[] = {
+				{ { -5.08594, -0.996094, -1.28249 }, 271.675 },
+				{ { 4.69141, 5.75781, -1.28248 }, 88.5845 },
+				{ { -0.40625, 7.68945, -1.28235 }, 89.3095 },
+				{ { 0.767578, 8.21289, -1.28248 }, 358.72 }
+		};
+		fridgePos[] = {
+				{ { 1.85352, 8.18359, -1.28246 }, 358.017 }
+		};
+		couchPos[] = {
+				{ { 3.74219, 0.666016, -1.2825 }, 2.03192 }
+		};
+		chairPos[] = {
+				{ { 4.28711, 3.18164, -1.28249 }, 83.2968 }
+		};
+		filingPos[] = {
+				{ { 4.66992, 4.72852, -1.28246 }, 90.2187 },
+				{ { 4.56055, 8.21484, -1.28246 }, 359.503 }
+		};
+		pelicanPos[] = {
+				{ { -2.64063, 3.3418, -1.36499 }, 279.695 }
+		};
+		tablePos[] = {
+				{ { 0.695313, 2.63086, -1.2825 }, 269.367 }
+		};
+		lockerPos[] = {
+				{ { -4.9668, 7.55469, -1.28248 }, 268.961 }
+		};
+		toolRackPos[] = {
+				{ { -5.35742, 4.21484, -1.28251 }, 268.905 },
+				{ { -2.66602, 8.53711, -1.28252 }, 359.227 }
+		};
+		palletPos[] = {
+				{ { 1.45898, 10.209, -10 }, 271 }
+		};
+		freezerPos[] = {
+				{ { 0.634766, 0.9375, -1.29904 }, 271 }
+		};
+	};
+	class Land_Cargo_Tower_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 0.552734, -1.625, 5.05352 }, 178.136 },
+				{ { -5.64844, -0.777344, -0.146542 }, 268.528 },
+				{ { 3.00781, 5.08984, -0.14654 }, 0.118889 },
+				{ { 5.29297, -4.09766, 2.45352 }, 181.726 },
+				{ { 5.87109, 3.71094, -0.146542 }, 90.215 }
+		};
+		fridgePos[] = {
+				{ { -4.76172, 6.00391, 2.58955 }, 0.860298 }
+		};
+		bedPos[] = {
+				{ { -1.06836, -2.64844, 2.45344 }, 1.62634 },
+				{ { -1.98633, -5.37109, 2.58947 }, 90.633 }
+		};
+		couchPos[] = {
+				{ { -2.73828, -4.30469, 0.0657005 }, 92.6866 },
+				{ { 3.06641, 4.15234, 2.45343 }, 270.898 }
+		};
+		cookerPos[] = {
+				{ { -3.72461, 5.85938, 2.58907 }, 354.227 }
+		};
+		chairPos[] = {
+				{ { -5.20117, 4.55859, -0.146549 }, 342.563 }
+		};
+		filingPos[] = {
+				{ { -4.07422, 0.931641, 5.05348 }, 270.343 },
+				{ { 1.67773, -3.90039, -0.146523 }, 178.258 },
+				{ { 2.23047, -3.91016, -0.146523 }, 176.187 },
+				{ { 2.77344, -3.92188, -0.146523 }, 177.87 },
+				{ { -0.625, 6.09961, 2.58951 }, 5.7925 },
+				{ { -1.2207, 6.14258, 2.58951 }, 348.149 },
+				{ { 5.6582, 4.91406, -0.14642 }, 44.1073 }
+		};
+		pelicanPos[] = {
+				{ { -0.595703, 1.38867, -4.32224 }, 95.4843 },
+				{ { -2.75977, -0.216797, 4.97829 }, 358.108 },
+				{ { -6.13672, 4.19727, 2.66941 }, 266.308 }
+		};
+		tablePos[] = {
+				{ { 6.55664, -0.714844, 2.64038 }, 89.4977 }
+		};
+		lockerPos[] = {
+				{ { 5.69727, 1.50391, -0.146555 }, 90.7961 },
+				{ { -5.09961, -5.55469, 2.5895 }, 267.631 }
+		};
+		toolRackPos[] = {
+				{ { -5.85156, -0.675781, 2.45344 }, 267.998 }
+		};
+		palletPos[] = {
+				{ { 4.88086, 3.89453, 4.99228 }, 0 }
+		};
+		freezerPos[] = {
+				{ { -2.78516, 4.82227, 0.0422649 }, 0 },
+				{ { -2.44922, 6.07422, 2.59228 }, 0 }
+		};
+		cabinetPos[] = {
+				{ { -0.0439453, 1.71631, 1.22235 }, 180 },
+				{ { 5.98682, 1.67432, 3.82235 }, 180 }
+		};
+	};
+	class Land_Cargo_Tower_V1_No1_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V1_No2_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V1_No3_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V1_No4_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V1_No5_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V1_No6_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V1_No7_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V2_F : Land_Cargo_Tower_V1_F { };
+	class Land_Cargo_Tower_V3_F : Land_Cargo_Tower_V1_F { };
+
+	class Land_Cargo_Patrol_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 2.61914, 0.96875, -0.789934 }, 90 }
+		};
+		filingPos[] = {
+				{ { 0.0117188, -0.283203, -0.585847 }, 1.42631 }
+		};
+		pelicanPos[] = {
+				{ { 2.00977, -1.52344, -0.585956 }, 133.857 }
+		};
+		palletPos[] = {
+				{ { -2, -1.18555, -0.546541 }, 0 }
+		};
+	};
+	class Land_Cargo_Patrol_V2_F : Land_Cargo_Patrol_V1_F { };
+	class Land_Cargo_Patrol_V3_F : Land_Cargo_Patrol_V1_F { };
+
+	class Land_Medevac_house_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -2.61328, 1.0918, -0.132166 }, 268.449 }
+		};
+		fridgePos[] = {
+				{ { 1.60156, 3.75977, 0.00387168 }, 355.287 }
+		};
+		filingPos[] = {
+				{ { 0.804688, 3.71875, 0.0038867 }, 7.82585 },
+				{ { -0.345703, 3.80273, 0.00388789 }, 352.697 },
+				{ { 0.230469, 3.85742, 0.00388765 }, 358.782 }
+		};
+		pelicanPos[] = {
+				{ { -1.60938, 3.26563, 0.00385642 }, 310.616 }
+		};
+		palletPos[] = {
+				{ { -0.992188, -1.05469, -10 }, 0 }
+		};
+		freezerPos[] = {
+				{ { 1.92969, 0.728516, -0.141539 }, 93 }
+		};
+		cabinetPos[] = {
+				{ { -0.466309, -0.0532227, 1.20856 }, 270 }
+		};
+	};
+	class Land_Medevac_HQ_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 2.64648, -5.00391, -3.29263 }, 179.273 },
+				{ { 2.29688, 5.35156, -3.17222 }, 270.045 },
+				{ { -3.0918, -6.20117, -3.17222 }, 179.371 },
+				{ { 7.07422, 5.46094, -3.17222 }, 89.2793 }
+		};
+		fridgePos[] = {
+				{ { 0.632813, -6.04492, -3.17222 }, 175.967 }
+		};
+		bedPos[] = {
+				{ { -2.65625, 3.33594, -3.29266 }, 272.831 }
+		};
+		couchPos[] = {
+				{ { 0.324219, 1.95313, -3.29266 }, 358.773 },
+				{ { 2.9082, 1.72461, -3.29266 }, 179.738 }
+		};
+		cookerPos[] = {
+				{ { -0.855469, -6.01172, -3.17265 }, 179.15 }
+		};
+		filingPos[] = {
+				{ { 8.4043, 2.6875, -3.1722 }, 89.7649 },
+				{ { 8.46094, 3.30469, -3.17216 }, 91.4351 },
+				{ { 8.44336, 3.92969, -3.1722 }, 86.4131 }
+		};
+		pelicanPos[] = {
+				{ { -1.37891, 1.34375, -0.767792 }, 314.355 },
+				{ { 8.10742, -1.68359, -3.17223 }, 92.3688 }
+		};
+		lockerPos[] = {
+				{ { 4.35938, 6.0918, -3.17224 }, 356.684 }
+		};
+		palletPos[] = {
+				{ { 0.730469, -4.40039, -0.740871 }, 88 },
+				{ { 0.443359, 5.44727, -10 }, 88 },
+				{ { 8.66406, 0.873047, -10 }, 93 }
+		};
+		freezerPos[] = {
+				{ { 3.97852, -4.58594, -3.27087 }, 182 }
+		};
+		cabinetPos[] = {
+				{ { -4.21851, -3.41992, -1.96077 }, 0 },
+				{ { 7.48828, 0.931152, -1.96077 }, 180 }
+		};
+	};
+	class Land_Cargo_HQ_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 0.302734, 4.1875, -3.2987 }, 357.597 },
+				{ { -4.12695, -4.01367, -3.2987 }, 270.059 },
+				{ { 2.70313, -5.11328, -3.2987 }, 177.669 }
+		};
+		fridgePos[] = {
+				{ { 2.61328, 5.1543, -3.17829 }, 267.807 }
+		};
+		bedPos[] = {
+				{ { -2.7832, 3.32227, -3.29873 }, 275.975 }
+		};
+		couchPos[] = {
+				{ { 3.13867, 1.7832, -3.29873 }, 191.466 }
+		};
+		cookerPos[] = {
+				{ { 6.47852, 6.01563, -3.17872 }, 359.818 }
+		};
+		filingPos[] = {
+				{ { 1.13867, 4.03906, -3.29869 }, 0.783356 },
+				{ { 8.12305, 2.75586, -3.17827 }, 92.0529 },
+				{ { 8.17969, 3.33203, -3.17827 }, 94.1181 },
+				{ { 8.21094, 3.93945, -3.17827 }, 88.4099 }
+		};
+		pelicanPos[] = {
+				{ { 1.14453, -4.05273, -0.773856 }, 93.1656 },
+				{ { 7.8418, -1.64844, -0.784355 }, 90.2845 }
+		};
+		tablePos[] = {
+				{ { -4.81055, 0.771484, -3.1762 }, 270.808 }
+		};
+		lockerPos[] = {
+				{ { 0.353516, -6.19727, -3.17831 }, 179.945 },
+				{ { -3.07031, -6.14453, -3.1783 }, 178.945 }
+		};
+		palletPos[] = {
+				{ { -4.98828, 2.61328, -10 }, 182 },
+				{ { 8.20508, 0.830078, -10 }, 182 }
+		};
+		freezerPos[] = {
+				{ { 4.70508, 6.0332, -3.14694 }, 1.99997 }
+		};
+		cabinetPos[] = {
+				{ { 7.48877, 0.866943, -1.96684 }, 180 }
+		};
+	};
+	class Land_Cargo_HQ_V2_F : Land_Cargo_HQ_V1_F { };
+	class Land_Cargo_HQ_V3_F : Land_Cargo_HQ_V1_F { };
+
+	class Land_Cargo_House_V1_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 2.3457, 3.13867, 0.0138698 }, 90 }
+		};
+		fridgePos[] = {
+				{ { -2.57031, 1.02148, -0.122166 }, 267.298 }
+		};
+		chairPos[] = {
+				{ { 1.95898, 1.5293, -0.122179 }, 85.5595 },
+				{ { -1.5293, 2.88867, 0.0138509 }, 322.238 }
+		};
+		filingPos[] = {
+				{ { 2.32617, 0.175781, -0.122146 }, 94.7773 }
+		};
+		pelicanPos[] = {
+				{ { -1.85352, 3.45898, 0.0138566 }, 317.925 }
+		};
+		palletPos[] = {
+				{ { -0.875, -0.892578, -10 }, 1.99997 }
+		};
+		freezerPos[] = {
+				{ { 0.539063, 3.60547, 0.048461 }, 1.99997 }
+		};
+		cabinetPos[] = {
+				{ { -0.491699, -0.0534668, 1.21856 }, 270 }
+		};
+	};
+	class Land_Cargo_House_V2_F : Land_Cargo_House_V1_F {};
+	class Land_Cargo_House_V3_F : Land_Cargo_House_V1_F {};
+
+	class Land_dp_smallFactory_F : Default {
+		shelfPos[] = {
+				{ { -1.50586, 4.79297, -1.87129 }, 180.159 }
+		};
+		pelicanPos[] = {
+				{ { 0.976563, 3.16797, 2.601 }, 97.492 }
+		};
+		palletPos[] = {
+				{ { 10.5059, 3.17969, 1.18859 }, 1.99997 }
+		};
+		cabinetPos[] = {
+				{ { 5.49805, 4.76367, -0.461309 }, 270 }
+		};
+	};
+	class Land_i_Garage_V2_F : Default {
+		shelfPos[] = {
+				{ { 4.90039, -1.19922, -0.123899 }, 88.7339 }
+		};
+		fridgePos[] = {
+				{ { 4.73438, 0.322266, -0.123898 }, 87.5145 }
+		};
+		chairPos[] = {
+				{ { 0.949219, -2.05469, -0.124453 }, 180.041 },
+				{ { 0.75, 2.23828, -0.120063 }, 0.105377 }
+		};
+		filingPos[] = {
+				{ { -0.519531, -2.26758, -0.123836 }, 181.535 },
+				{ { -1.18945, -2.25, -0.123765 }, 172.485 },
+				{ { 4.83203, -2.30273, -0.123735 }, 99.5615 }
+		};
+		tablePos[] = {
+				{ { -2.72266, -2.13672, -0.132633 }, 179.857 }
+		};
+		toolRackPos[] = {
+				{ { 2.37305, 2.75977, -0.124017 }, 0.339203 },
+				{ { 2.76563, -2.57422, -0.124016 }, 179.612 }
+		};
+		palletPos[] = {
+				{ { 4.86719, -2.50781, 3.25354 }, 1.99997 }
+		};
+		freezerPos[] = {
+				{ { -0.976563, 2.29883, -0.116461 }, 1.99997 }
+		};
+	};
+	class Land_u_Barracks_V2_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -2.01953, 3.59766, 1.40866 }, 86.5322 },
+				{ { -1.18555, 4.4082, 1.40866 }, 267.099 },
+				{ { 5.37695, -2.42773, -1.92432 }, 268.209 },
+				{ { -6.13281, 2.58594, -1.92432 }, 177.59 },
+				{ { 5.46289, 3.92578, 1.40866 }, 274.739 },
+				{ { 7.93555, -0.273438, 1.40866 }, 358.598 },
+				{ { 8.07617, -0.306641, -1.92432 }, 358.744 },
+				{ { 8.76172, 2.60547, -1.92432 }, 177.537 },
+				{ { -9.86914, 3.00391, -1.92432 }, 269.405 },
+				{ { -10.7129, 3.96875, -1.92432 }, 83.8923 },
+				{ { -12.1953, -0.359375, 1.40871 }, 357.971 },
+				{ { -14.7266, 2.57813, 1.40876 }, 181.364 },
+				{ { -16.127, -0.359375, -1.92432 }, 357.877 }
+		};
+		fridgePos[] = {
+				{ { 5.38867, -2.61719, 1.40864 }, 268.987 },
+				{ { -10.6797, -1.86133, 1.40864 }, 89.0169 },
+				{ { 10.7656, 4.75586, -1.92426 }, 91.9329 }
+		};
+		bedPos[] = {
+				{ { 0.00390625, 3.40234, -1.92435 }, 91.4153 },
+				{ { 3.58984, -1.56055, 1.40863 }, 1.31381 },
+				{ { 9.80469, 3.27344, 1.40863 }, 97.6807 },
+				{ { -11.7617, -1.10547, -1.92435 }, 269.367 }
+		};
+		couchPos[] = {
+				{ { 2.72656, -1.28516, -1.92434 }, 273.867 },
+				{ { -12.2988, 4.4668, 1.40864 }, 273.1 },
+				{ { -14.6855, 3.48633, -1.92434 }, 92.4456 }
+		};
+		wardrobePos[] = {
+				{ { -1.10742, -2.48438, 1.40877 }, 271.139 },
+				{ { 4.26367, 4.1543, -1.92423 }, 90.3193 }
+		};
+		cookerPos[] = {
+				{ { 5.52148, -1.29688, 1.40827 }, 269.9 },
+				{ { 10.6504, 3.61328, -1.92474 }, 90.6279 },
+				{ { -10.916, -2.91406, 1.40832 }, 90.6295 }
+		};
+		chairPos[] = {
+				{ { -0.787109, -2.87109, -1.92433 }, 243.978 },
+				{ { -0.953125, 5.07422, -1.92361 }, 283.609 },
+				{ { -4.85156, -2.75781, -1.92434 }, 47.0285 },
+				{ { 10.4219, 5.0625, 1.40864 }, 65.2562 },
+				{ { -16.418, 2.94922, 1.40865 }, 214.598 },
+				{ { -16.4785, -2.82031, -1.91074 }, 233.102 }
+		};
+		filingPos[] = {
+				{ { -1.01953, -0.453125, -1.92431 }, 357.404 },
+				{ { -2.16602, 2.66797, -1.92431 }, 110.166 },
+				{ { -2.17773, 4.83203, -1.9243 }, 85.2777 },
+				{ { -2.19727, 5.30664, 1.40868 }, 84.274 },
+				{ { -2.1875, 5.33203, -1.9243 }, 86.7386 },
+				{ { 4.37109, 4.39844, 1.40868 }, 71.7087 },
+				{ { 5.33984, -3.24219, -1.9243 }, 266.113 },
+				{ { 5.41602, 3.54102, -1.9243 }, 263.452 },
+				{ { 4.29102, 4.91797, 1.40868 }, 85.4132 },
+				{ { 5.37891, 4.10938, -1.9243 }, 275.738 },
+				{ { 4.30859, 5.38086, 1.40868 }, 86.1467 },
+				{ { 5.42188, 4.70898, -1.9243 }, 265.797 },
+				{ { -8.33398, 2.60352, 1.40874 }, 182.724 },
+				{ { -8.9707, 2.60156, 1.40873 }, 180.321 },
+				{ { -9.7832, 2.66797, 1.40867 }, 195.667 },
+				{ { 10.8789, -0.318359, 1.40868 }, 73.9709 },
+				{ { 10.9199, -2.47656, -1.9243 }, 90.1058 },
+				{ { 10.9004, -3.0625, -1.9243 }, 90.2496 },
+				{ { -10.8066, 5.09375, -1.9243 }, 83.8553 },
+				{ { -16.6758, -2.74023, 1.40868 }, 269.525 },
+				{ { -16.7051, -3.23438, 1.40868 }, 265.732 },
+				{ { -16.6211, 5.41211, 1.40871 }, 281.602 }
+		};
+		pelicanPos[] = {
+				{ { -9.26563, -4.74609, 1.40864 }, 207.949 },
+				{ { 10.4531, -2.80273, 1.40864 }, 136.408 },
+				{ { -16.666, 6.80859, 1.40864 }, 268.101 }
+		};
+		tablePos[] = {
+				{ { 1.79883, -0.546875, 1.40863 }, 0.30365 },
+				{ { 7.95117, -3.04688, 1.40863 }, 181.397 },
+				{ { -16.5781, 4.66797, -1.92435 }, 269.307 }
+		};
+		lockerPos[] = {
+				{ { 0.539063, 2.65234, 1.40864 }, 180.316 },
+				{ { 6.29492, -0.507813, -1.92434 }, 359.612 },
+				{ { -6.15625, 2.65625, 1.40874 }, 178.513 },
+				{ { -7.85352, 2.64844, -1.92434 }, 178.304 }
+		};
+		freezerPos[] = {
+				{ { -5.82227, -4.86328, 1.49066 }, 185 },
+				{ { -6.21875, 5.08398, 1.49066 }, 1.99997 },
+				{ { -16.0039, -0.580078, 1.49066 }, 1.99997 }
+		};
+		cabinetPos[] = {
+				{ { -7.6416, -5.42383, -0.409235 }, 270 },
+				{ { -17.0132, 1.60693, 2.89076 }, 0 }
+		};
+	};
+	class Land_i_Barracks_V2_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -3.91406, 4.36816, 0.579132 }, 0.755371 },
+				{ { -7.16211, -2.42676, 3.91308 }, 90 },
+				{ { -7.08789, -3.69434, 0.579177 }, 88.7136 },
+				{ { 7.91797, 2.16504, 0.579097 }, 93.2098 },
+				{ { 8.86719, 2.98145, 0.579103 }, 265.996 },
+				{ { 12.2988, 1.54297, 3.91308 }, 180.227 }
+		};
+		fridgePos[] = {
+				{ { -13.209, -3.86133, 0.579104 }, 267.951 },
+				{ { -13.1953, -3.96875, 3.91308 }, 266.429 }
+		};
+		bedPos[] = {
+				{ { 3.25391, 2.82422, 0.579081 }, 181.072 },
+				{ { 7.13086, -3.10645, 0.579082 }, 359.996 },
+				{ { 10.2598, -2.19531, 3.91306 }, 274.345 },
+				{ { -12.0977, 2.37695, 3.91306 }, 91.0109 },
+				{ { -12.6543, 3.00977, 0.579082 }, 184.452 }
+		};
+		couchPos[] = {
+				{ { 6.04297, -2.55762, 3.91306 }, 280.83 },
+				{ { 10.2754, -2.4502, 0.579081 }, 280.799 }
+		};
+		cookerPos[] = {
+				{ { -12.6621, -1.54199, 0.578749 }, 358.466 },
+				{ { -13.2383, -2.48926, 3.91281 }, 270.982 }
+		};
+		chairPos[] = {
+				{ { 2.58594, -3.91309, 0.579165 }, 261.223 },
+				{ { -1.79297, -5.75391, 3.91291 }, 133.686 },
+				{ { -5.79688, -5.7041, 3.91307 }, 219.674 },
+				{ { -7.625, 3.94141, 0.579091 }, 54.7852 },
+				{ { -7.74414, 3.70801, 3.91307 }, 37.1618 },
+				{ { 9.36328, -3.69824, 3.91306 }, 252.24 },
+				{ { 13.834, -3.76172, 0.579099 }, 137.181 }
+		};
+		filingPos[] = {
+				{ { 2.26172, -1.52051, 3.91319 }, 357.145 },
+				{ { 1.2832, 3.91797, 3.9131 }, 85.9261 },
+				{ { 1.24609, 4.37402, 3.9131 }, 83.3282 },
+				{ { 4.24805, 1.63574, 3.9131 }, 173.705 },
+				{ { 4.69336, 1.63379, 3.9131 }, 177.094 },
+				{ { -6.30273, 3.34766, 0.579124 }, 266.774 },
+				{ { -6.23438, 3.84766, 0.579123 }, 274.501 },
+				{ { -7.2832, -4.13379, 3.91315 }, 178.61 },
+				{ { 7.93555, 3.07422, 0.57919 }, 82.7785 },
+				{ { 9.14648, 1.65332, 3.91318 }, 184.404 },
+				{ { -13.1055, 4.26563, 3.91319 }, 288.672 },
+				{ { 14.2305, 1.56543, 3.9131 }, 90.9699 }
+		};
+		pelicanPos[] = {
+				{ { -2.50391, -5.04199, 7.4924 }, 169.602 },
+				{ { -12.8242, 5.72266, 3.82744 }, 317.876 }
+		};
+		tablePos[] = {
+				{ { -2.72852, 1.75391, 3.91306 }, 178.339 },
+				{ { -3.75781, -6.08105, 3.91306 }, 180.247 },
+				{ { -9.46484, 4.12988, 3.91306 }, 359.779 },
+				{ { -10.1289, 4.10156, 0.579082 }, 0.00921631 }
+		};
+		lockerPos[] = {
+				{ { 1.2793, 3.78418, 0.579087 }, 86.8882 },
+				{ { -2.58789, -3.3584, 0.579131 }, 359.544 },
+				{ { -6.10352, 2.22168, 3.9131 }, 269.383 },
+				{ { -6.22266, 2.14551, 0.57918 }, 271.326 },
+				{ { 7.8125, 3.70313, 3.91307 }, 88.2042 },
+				{ { 14.0723, -3.22754, 3.91322 }, 88.3252 },
+				{ { 14.2012, 3.70703, 0.579188 }, 88.1507 }
+		};
+		toolRackPos[] = {
+				{ { -2.93359, 1.23438, 0.579066 }, 180.737 }
+		};
+		freezerPos[] = {
+				{ { -8.96094, -1.70605, 0.508462 }, 360 },
+				{ { -9.27344, -1.7334, 3.85846 }, 360 }
+		};
+		cabinetPos[] = {
+				{ { -6.55322, -2.76904, 5.30856 }, 0 }
+		};
+	};
+	class Land_i_Barracks_V1_F : Land_i_Barracks_V2_F {};
+
+	class Land_d_Windmill01_F : Default {
+		pelicanPos[] = {
+				{ { -0.888672, 0.150391, 0.965859 }, 132.301 },
+				{ { -2.92773, 2.72852, 0.965859 }, 132.301 }
+		};
+		shoeboxPos[] = {
+				{ { -1.11328, 1.94141, -2.32494 }, 0 }
+		};
+		palletPos[] = {
+				{ { -1.2168, -0.789063, -2.31882 }, 360 }
+		};
+	};
+	class Land_u_Shed_Ind_F : Default {
+		shelfPos[] = {
+				{ { -5.90234, -1.64453, -1.45137 }, 179.958 },
+				{ { -0.0195313, 8.39648, -1.43504 }, 6.50078 }
+		};
+		chairPos[] = {
+				{ { -6.70703, 1.93359, -1.43049 }, 357.957 }
+		};
+		filingPos[] = {
+				{ { -4.83203, 2.18359, -1.43498 }, 3.3181 },
+				{ { 6.72266, -1.33984, -1.43502 }, 208.603 }
+		};
+		pelicanPos[] = {
+				{ { -4.81055, 2.53125, 1.51679 }, 180.197 }
+		};
+		palletPos[] = {
+				{ { -2.47266, 7.94922, -1.51769 }, 360 },
+				{ { 8.92383, 3.15039, -1.51769 }, 91 },
+				{ { 14.834, -0.927734, -1.51769 }, 360 }
+		};
+		freezerPos[] = {
+				{ { -4.94141, 0.177734, -1.51769 }, 91 }
+		};
+	};
+	class Land_i_Shed_Ind_F : Default {
+		shelfPos[] = {
+				{ { 3.75781, -1.39844, -1.43597 }, 178.598 },
+				{ { -6.29688, -1.64844, -1.43597 }, 180.042 },
+				{ { -1.85938, 8.64844, -1.43597 }, 1.254 },
+				{ { 15.8086, 5.01367, -1.43592 }, 90.0259 }
+		};
+		fridgePos[] = {
+				{ { -4.0918, 2.25391, -1.43614 }, 269.878 }
+		};
+		couchPos[] = {
+				{ { 0.0527344, 3.64453, -1.43621 }, 214.408 }
+		};
+		cookerPos[] = {
+				{ { -5.86523, 2.00391, -1.43661 }, 358.838 }
+		};
+		chairPos[] = {
+				{ { 15.4688, 7.75586, -1.43558 }, 90.0616 }
+		};
+		filingPos[] = {
+				{ { -4.89648, -1.49609, -1.43617 }, 178.115 },
+				{ { -0.6875, 8.5293, -1.43617 }, 358.177 },
+				{ { 15.6309, 6.44336, -1.43617 }, 86.6001 }
+		};
+		lockerPos[] = {
+				{ { 8.71875, -1.44141, -1.43617 }, 179.064 }
+		};
+		toolRackPos[] = {
+				{ { -1.60742, -1.83594, -1.43602 }, 181.024 },
+				{ { 8.53906, 8.82422, -1.43602 }, 0.0482483 }
+		};
+		shoeboxPos[] = {
+				{ { -6.76953, 2.15039, -1.43621 }, 356.704 }
+		};
+		palletPos[] = {
+				{ { -3.32031, -0.679688, -1.50884 }, 0 },
+				{ { 2.86133, 7.82031, -1.50884 }, 0 },
+				{ { 4.82617, 7.72266, -1.50884 }, 91 }
+		};
+		freezerPos[] = {
+				{ { -5.0293, 0.304688, -1.47884 }, 91 }
+		};
+	};
+	class Land_BagBunker_Large_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 0.0820313, -0.914063, -0.766759 }, 269.689 },
+				{ { -3.85156, -1.18555, -0.76676 }, 272.369 }
+		};
+		couchPos[] = {
+				{ { 2.64648, -3.39258, -0.766787 }, 87.3199 }
+		};
+		filingPos[] = {
+				{ { 1.02148, 1.80078, -0.766742 }, 359.841 },
+				{ { -1.57031, 1.58008, -0.766741 }, 359.563 }
+		};
+		pelicanPos[] = {
+				{ { 0.382813, 0.970703, -0.766773 }, 269.105 },
+				{ { -0.769531, -1.35352, -0.766773 }, 73.6849 }
+		};
+		tablePos[] = {
+				{ { -0.646484, 0.492188, -0.766783 }, 90.0316 }
+		};
+		palletPos[] = {
+				{ { 3.13672, -1.19922, -0.69166 }, 0 },
+				{ { -3.17773, -3.54102, -0.69166 }, 0 }
+		};
+	};
+	class Land_BagBunker_Tower_F : Default {
+		lootType = "mil";
+		pelicanPos[] = {
+				{ { -1.95508, -0.0195313, -0.90682 }, 269.117 },
+				{ { -0.703125, -2.67383, 0.543192 }, 212.824 }
+		};
+		shoeboxPos[] = {
+				{ { 2.37891, -1.93555, -0.873341 }, 0 }
+		};
+		palletPos[] = {
+				{ { -0.732422, 2.07617, -2.17334 }, 348 },
+				{ { -0.771484, -2.21094, -2.17334 }, 0 }
+		};
+	};
+	class Land_Offices_01_V1_F : Default {
+		shelfPos[] = {
+				{ { 1.18555, 3.91016, 0.712739 }, 359.463 },
+				{ { -1.66797, -5.50195, -7.10147 }, 270.268 },
+				{ { 7.38477, 1.79688, 4.63913 }, 89.9385 },
+				{ { 7.48047, -4.49414, -3.18539 }, 87.3787 },
+				{ { 8.3457, 2.93164, -7.08514 }, 270.832 },
+				{ { 12.4883, 3.5293, -7.08517 }, 89.032 },
+				{ { -16.5742, 6.37695, -7.08533 }, 269.995 }
+		};
+		fridgePos[] = {
+				{ { 8.60547, 1.05273, -7.08536 }, 261.917 }
+		};
+		couchPos[] = {
+				{ { 6.63086, 2.61133, -7.08539 }, 359.378 }
+		};
+		cookerPos[] = {
+				{ { 12.3555, 2.48438, -7.0863 }, 88.735 }
+		};
+		chairPos[] = {
+				{ { 7.04297, -3.23633, 0.71985 }, 54.2617 },
+				{ { 6.82813, -6.80273, 0.698915 }, 116.603 },
+				{ { 12.1465, 7.50391, -7.08539 }, 21.1554 },
+				{ { -15.248, 7.0293, 9.81961 }, 126.25 }
+		};
+		filingPos[] = {
+				{ { -0.162109, 3.8457, 0.712403 }, 355.347 },
+				{ { -1.60547, -3.98047, -7.10168 }, 270.008 },
+				{ { 5.1582, 0.861328, 4.62234 }, 186.369 },
+				{ { 4.81445, 3.93164, -3.18698 }, 86.2954 },
+				{ { 5.69531, 3.77148, -3.18649 }, 2.55798 },
+				{ { 7.20898, 3.4082, 4.63931 }, 72.0425 },
+				{ { 7.27148, 3.89648, 4.63947 }, 84.6359 },
+				{ { 7.46289, -5.73047, -3.18564 }, 89.5913 },
+				{ { 9.43164, -0.173828, 4.63879 }, 355.477 },
+				{ { 9.89063, -0.171875, 4.63889 }, 357.143 },
+				{ { 8.54883, 7.85352, -7.08539 }, 354.434 },
+				{ { 9.14648, 7.85742, -7.08536 }, 342.434 },
+				{ { -16.377, 4.77148, -7.08542 }, 249.62 },
+				{ { -16.3945, 5.33789, -7.08539 }, 272.711 }
+		};
+		pelicanPos[] = {
+				{ { -7.70508, 3.27539, 9.80328 }, 165.226 },
+				{ { 11.9551, -0.947266, 4.63934 }, 22.6672 },
+				{ { -11.2246, -9.87695, 9.81973 }, 358.384 }
+		};
+		tablePos[] = {
+				{ { 7.39453, -4.98242, 0.717866 }, 89.9971 }
+		};
+		lockerPos[] = {
+				{ { 10.3047, 7.74805, -7.08533 }, 359.889 }
+		};
+		toolRackPos[] = {
+				{ { 5.50781, 4.16016, 4.63971 }, 0.0171576 }
+		};
+		shoeboxPos[] = {
+				{ { 3.23438, -0.361328, 4.62201 }, 314.676 },
+				{ { -13.8223, 3.53125, 9.81958 }, 272.283 }
+		};
+		palletPos[] = {
+				{ { -5.23438, 1.21875, 9.83307 }, 348 },
+				{ { 6.88477, -4.91602, 9.83307 }, 348 },
+				{ { -7.91797, -7.52734, 9.83307 }, 348 },
+				{ { 10.6992, -4.91602, 4.62271 }, 0 },
+				{ { 11.6484, 6.89063, 9.83307 }, 348 }
+		};
+		freezerPos[] = {
+				{ { 12.332, 1.18555, -7.08542 }, 90 }
+		};
+		cabinetPos[] = {
+				{ { 4.25293, 0.552734, 5.96314 }, 270 },
+				{ { 10.2202, 0.577637, -5.76687 }, 270 }
+		};
+	};
+	class Land_d_Stone_Shed_V1_F : Default {
+		bedPos[] = {
+				{ { 2.83203, 1.8418, -0.0155029 }, 269.824 }
+		};
+		palletPos[] = {
+				{ { 1.93164, -3.32227, -0.204259 }, 348 }
+		};
+	};
+	class Land_i_House_Small_01_V1_F : Default {
+		shelfPos[] = {
+				{ { 0.712891, 2.01953, -1.06801 }, 179.636 }
+		};
+		fridgePos[] = {
+				{ { 2.66992, 4.35742, -1.06878 }, 359.99 }
+		};
+		bedPos[] = {
+				{ { -3.06641, 0.388672, -1.06846 }, 273.681 }
+		};
+		couchPos[] = {
+				{ { -1.00195, -3.09766, -1.06751 }, 91.678 }
+		};
+		wardrobePos[] = {
+				{ { -4.11914, -1.85938, -1.08958 }, 269.028 }
+		};
+		cookerPos[] = {
+				{ { 3.91992, 4.31055, -1.06914 }, 359.993 }
+		};
+		chairPos[] = {
+				{ { 0.800781, -1.12109, -1.06701 }, 74.3133 },
+				{ { -3.79492, 2.57031, -1.06882 }, 227.565 },
+				{ { -3.71875, -3.625, -1.05222 }, 229.261 }
+		};
+		tablePos[] = {
+				{ { -1.875, 2.26172, -1.06836 }, 180.248 }
+		};
+	};
+	class Land_i_House_Small_01_V2_F : Land_i_House_Small_01_V1_F {};
+	class Land_i_House_Small_01_V3_F : Land_i_House_Small_01_V1_F {};
+
+	class Land_u_House_Big_01_V1_F : Default {
+		shelfPos[] = {
+				{ { 1.56055, 0.982422, -2.59135 }, 359.087 },
+				{ { 3.06836, -2.5918, 0.82871 }, 183.405 },
+				{ { 1.39258, -6.70508, -2.59135 }, 182.249 },
+				{ { 3.20703, 6.94922, -2.59135 }, 3.01586 }
+		};
+		fridgePos[] = {
+				{ { 0.0546875, 6.88477, -2.59135 }, 357.87 }
+		};
+		bedPos[] = {
+				{ { -3.03711, -5.87109, 0.828626 }, 91.7906 },
+				{ { 3.26758, -5.99805, 0.828626 }, 95.6288 },
+				{ { 3.30273, 6.26758, 0.828626 }, 272.1 }
+		};
+		couchPos[] = {
+				{ { 0.482422, -1.69141, 0.828588 }, 90.019 },
+				{ { -0.523438, -2.12305, -2.59137 }, 184.381 },
+				{ { -3.01172, 5.27344, -2.59138 }, 177.246 },
+				{ { -2.76172, 6.18164, 0.828629 }, 269.359 }
+		};
+		wardrobePos[] = {
+				{ { 4.35352, -3.9375, 0.828695 }, 91.5199 },
+				{ { 4.29102, 4.18945, 0.828744 }, 90.0946 }
+		};
+		cookerPos[] = {
+				{ { 1.33398, 6.96289, -2.59177 }, 359.496 }
+		};
+		chairPos[] = {
+				{ { -2.83594, -0.570313, -2.68137 }, 75.9432 },
+				{ { -2.95898, -2.29102, -2.68137 }, 82.8292 },
+				{ { 0.705078, 3.7832, 0.828643 }, 180.005 },
+				{ { -3.65039, 2.18164, 0.828628 }, 313.022 },
+				{ { 4.05078, 3.9043, -2.5792 }, 111.554 }
+		};
+		tablePos[] = {
+				{ { 1.75977, 0.580078, 0.828625 }, 359.479 },
+				{ { 0.232422, -6.57813, 0.828625 }, 180.157 },
+				{ { 4.20898, -5.11133, -2.59138 }, 90.4264 }
+		};
+		toolRackPos[] = {
+				{ { -2.86523, 1.09375, -2.5914 }, 179.843 }
+		};
+		freezerPos[] = {
+				{ { 1.55859, 3.71289, -2.61788 }, 177 }
+		};
+	};
+	class Land_i_House_Big_01_V1_F : Default {
+		shelfPos[] = {
+				{ { -3.33594, 2.69336, 0.828649 }, 0.965714 },
+				{ { 2.96875, 3.44531, -2.59135 }, 182.75 },
+				{ { 3.70898, -2.63867, 0.828707 }, 180.077 },
+				{ { -4.09961, 2.03516, -2.59135 }, 269.739 }
+		};
+		fridgePos[] = {
+				{ { 4.11914, 3.69922, -2.59135 }, 168.186 },
+				{ { -3.92578, 6.73633, 0.828675 }, 290.478 }
+		};
+		bedPos[] = {
+				{ { 0.835938, -4.21289, 0.828626 }, 273.228 },
+				{ { 3.64844, -4.54688, 0.828626 }, 1.43391 },
+				{ { 3.29297, 6.20508, 0.828626 }, 273.903 }
+		};
+		couchPos[] = {
+				{ { -3.18359, -1.30273, 0.828621 }, 181.307 },
+				{ { 3.11328, 4.48633, 0.828624 }, 89.1209 },
+				{ { -2.99609, 5.30859, -2.59138 }, 178.667 },
+				{ { 3.05664, -5.85938, -2.59137 }, 90.8967 }
+		};
+		wardrobePos[] = {
+				{ { -3.63281, 3.60742, 0.828697 }, 179.668 },
+				{ { 4.30273, -3.64844, -2.59128 }, 90 },
+				{ { -2.66016, -6.70117, 0.828702 }, 181.241 }
+		};
+		cookerPos[] = {
+				{ { 1.91016, 3.75781, -2.59176 }, 181.387 }
+		};
+		chairPos[] = {
+				{ { 1.875, 0.488281, -2.59136 }, 12.1387 },
+				{ { 2.37305, 0.478516, 0.85181 }, 84.246 },
+				{ { -3, -0.240234, -2.68137 }, 57.9483 },
+				{ { -2.76953, -2.30664, -2.68137 }, 80.3317 }
+		};
+		tablePos[] = {
+				{ { -1.17188, -1.95703, -2.59138 }, 270.249 },
+				{ { 0.105469, 6.84375, 0.828627 }, 0.00192261 }
+		};
+		toolRackPos[] = {
+				{ { -2.8125, 1.09375, -2.59139 }, 180.224 }
+		};
+		freezerPos[] = {
+				{ { 0.59375, 6.80273, -2.61788 }, 3 }
+		};
+	};
+	class Land_i_House_Big_01_V2_F : Land_i_House_Big_01_V1_F {};
+	class Land_i_House_Big_01_V3_F : Land_i_House_Big_01_V1_F {};
+
+	class Land_Hospital_side1_F : Default {
+		lootBiasPos = 50;
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -5.40039, 2.9082, -7.91957 }, 2.14935 },
+				{ { -6.25, -7.61719, -7.92645 }, 270.854 },
+				{ { 2.96484, 10.916, 7.44774 }, 359.647 },
+				{ { 9.71289, -6.21289, -7.92565 }, 87.4192 }
+		};
+		fridgePos[] = {
+				{ { -2.66797, -11.2949, -7.92739 }, 179.351 }
+		};
+		bedPos[] = {
+				{ { -0.912109, -3.25195, -7.92176 }, 91.7708 },
+				{ { 4.10547, -3.19141, -7.92099 }, 89.6996 },
+				{ { -0.625, -5.79688, -7.92326 }, 266.29 }
+		};
+		couchPos[] = {
+				{ { -0.386719, 1.90234, -7.91998 }, 271.421 }
+		};
+		cookerPos[] = {
+				{ { -3.70898, -11.1738, -7.92829 }, 180.136 }
+		};
+		chairPos[] = {
+				{ { 7.83008, 2.53906, -7.92018 }, 340.336 },
+				{ { 9.28516, 2.51367, -7.92017 }, 27.1802 }
+		};
+		pelicanPos[] = {
+				{ { -4.66016, 7.12695, -0.617266 }, 177.338 }
+		};
+		lockerPos[] = {
+				{ { -6.14063, 0.242188, -7.92175 }, 268.608 },
+				{ { 9.47656, 0.0429688, -7.92081 }, 89.4555 }
+		};
+		toolRackPos[] = {
+				{ { -0.167969, 6.7832, 7.44031 }, 0.197876 },
+				{ { 9.93164, 5.09375, -7.92028 }, 90.4358 }
+		};
+		shoeboxPos[] = {
+				{ { 9.5293, -5.36328, 7.43984 }, 95.593 },
+				{ { 0.126953, 11.4395, 2.38099 }, 90 }
+		};
+		palletPos[] = {
+				{ { 5.7207, 5.69727, 7.451 }, 90 },
+				{ { -2.02539, 10.0586, -7.89901 }, 90 },
+				{ { 8.43945, -6.91016, 7.451 }, 90 }
+		};
+		freezerPos[] = {
+				{ { 9.43945, -8.53906, -7.90901 }, 90 }
+		};
+		cabinetPos[] = {
+				{ { -6.39502, -5.93945, -6.56894 }, 0 },
+				{ { 9.81006, -4.22998, -6.56894 }, 180 }
+		};
+	};
+	class Land_Hospital_side2_F : Default {
+		lootBiasPos = 100;
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 4.34766, -10.8809, -8.13902 }, 179.22 },
+				{ { -5.95703, -10.9258, -8.13085 }, 178.43 },
+				{ { 15.5547, -1.91211, -8.13092 }, 359.336 }
+		};
+		fridgePos[] = {
+				{ { -8.25586, -8.94922, -8.13093 }, 267.631 }
+		};
+		couchPos[] = {
+				{ { -7.19336, 1.38867, -8.13095 }, 179.136 }
+		};
+		cookerPos[] = {
+				{ { -8.17969, -10.1816, -8.13134 }, 269.372 }
+		};
+		pelicanPos[] = {
+				{ { -7.36133, 3.00391, 11.219 }, 282.51 },
+				{ { -10.5195, -4.00586, 7.23841 }, 237.441 }
+		};
+		lockerPos[] = {
+				{ { -1.35938, -10.7207, -8.13079 }, 180.383 },
+				{ { 8.24609, -10.7422, -8.13077 }, 178.528 },
+				{ { 11.8438, -10.709, -8.13078 }, 179.892 }
+		};
+		toolRackPos[] = {
+				{ { -8.71094, -3.51367, -8.13078 }, 270 }
+		};
+		palletPos[] = {
+				{ { 1.34766, -0.410156, -8.14045 }, 90 },
+				{ { -5.58203, -0.400391, -8.14045 }, 90 },
+				{ { 6.58008, 3.79297, -8.14045 }, 90 }
+		};
+		freezerPos[] = {
+				{ { -4.11914, -10.6406, -8.14045 }, 180 }
+		};
+		cabinetPos[] = {
+				{ { 0.117188, -1.76221, -6.77038 }, 90 },
+				{ { 1.33496, -11.0347, -6.77038 }, 270 },
+				{ { 14.6304, -11.0347, -6.77038 }, 270 }
+		};
+	};
+	class Land_Hospital_main_F : Default {
+		lootBiasPos = 50;
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 14.3516, -2.75781, -8.04017 }, 90.0092 },
+				{ { 14.4102, -10.8027, -8.04031 }, 88.6201 },
+				{ { 6.71875, -18.9082, 7.32756 }, 88.904 }
+		};
+		chairPos[] = {
+				{ { -10.5098, -12.3008, -8.04206 }, 272.027 },
+				{ { 4.49805, 19.1055, -8.03925 }, 359.681 },
+				{ { 6.66016, 19.1504, -8.0383 }, 356.209 },
+				{ { 0.691406, -20.6602, -8.03756 }, 178.007 },
+				{ { 9.58594, 19.1191, -8.03699 }, 359.114 },
+				{ { 5.51953, -20.6738, -8.03836 }, 179.692 }
+		};
+		pelicanPos[] = {
+				{ { 5.98242, -6.33008, 7.32735 }, 22.9554 },
+				{ { 10.6816, 0.878906, -7.01696 }, 300.456 }
+		};
+		tablePos[] = {
+				{ { 3.35938, -20.6563, -8.04205 }, 179.332 }
+		};
+		lockerPos[] = {
+				{ { 14.1875, 10.0137, -8.03923 }, 88.9373 },
+				{ { 14.1895, 12.209, -8.03929 }, 89.9901 },
+				{ { 14.2285, 14.4941, -8.03952 }, 89.9901 }
+		};
+		shoeboxPos[] = {
+				{ { -1.01367, -4.61133, -1.04287 }, 188.531 },
+				{ { 12.9023, -7.36914, 7.32734 }, 210.758 }
+		};
+		palletPos[] = {
+				{ { -10.0547, -6.80469, -8.05151 }, 180 },
+				{ { 1.62305, 18.627, -8.05151 }, 180 }
+		};
+		cabinetPos[] = {
+				{ { 14.5366, -0.879883, -6.88591 }, 180 },
+				{ { 14.5056, -9.5127, -6.88591 }, 180 }
+		};
+	};
+	class Land_d_Shop_02_V1_F : Default {
+		shelfPos[] = {
+				{ { -1.25195, -3.0293, -2.79755 }, 250.062 },
+				{ { 1.2793, -4.13867, -2.79765 }, 178.402 },
+				{ { -5.04688, 2.14063, 1.0899 }, 273.409 }
+		};
+		fridgePos[] = {
+				{ { -1.77148, 4.04688, -2.79756 }, 99.5686 }
+		};
+		couchPos[] = {
+				{ { 2.12891, -2.2168, 1.09474 }, 94.155 }
+		};
+		chairPos[] = {
+				{ { 4.9707, -3.16992, 1.09518 }, 180.623 },
+				{ { -4.48633, 4.00586, -2.79765 }, 325.405 },
+				{ { -4.54883, 4.10352, 1.08889 }, 330.581 }
+		};
+		pelicanPos[] = {
+				{ { 0.566406, 1.81445, 4.50349 }, 353.316 }
+		};
+		palletPos[] = {
+				{ { -4.0918, -2.88086, -2.45702 }, 238 },
+				{ { 4.7832, 4.58984, -2.52702 }, 238 }
+		};
+	};
+	class Land_d_House_Small_01_V1_F : Default {
+		shelfPos[] = {
+				{ { 6.82813, 6.42285, -0.861197 }, 354.675 }
+		};
+		fridgePos[] = {
+				{ { 8.44141, 2.11035, -0.859908 }, 88.0808 }
+		};
+		couchPos[] = {
+				{ { 4.75, -0.15918, -0.859535 }, 359.982 }
+		};
+		chairPos[] = {
+				{ { 3.43359, 4.36914, -0.860575 }, 180.154 },
+				{ { 8.06641, 5.74316, -0.861 }, 27.7226 }
+		};
+		pelicanPos[] = {
+				{ { 1.42969, -1.45898, 2.73048 }, 98.759 }
+		};
+		palletPos[] = {
+				{ { 0.865234, -1.24414, -0.196001 }, 238 }
+		};
+	};
+	class Land_Church_01_V1_F : Default {
+		wardrobePos[] = {
+				{ { 4.5957, 5.27051, -6.49416 }, 180.222 }
+		};
+		chairPos[] = {
+				{ { 5.29492, -5.31055, -6.49177 }, 10.1886 }
+		};
+		pelicanPos[] = {
+				{ { -5.17969, 5.49414, -5.52077 }, 284.503 }
+		};
+		shoeboxPos[] = {
+				{ { -5.70313, -4.57422, -5.49586 }, 271.096 }
+		};
+		palletPos[] = {
+				{ { -4.17969, -5.43848, -5.50872 }, 238 }
+		};
+	};
+	class Land_Castle_01_tower_F : Default {
+		shelfPos[] = {
+				{ { 2.58398, 2.0791, -9.50637 }, 85.6591 }
+		};
+		couchPos[] = {
+				{ { -1.63281, -1.07617, -9.50639 }, 183.524 }
+		};
+		chairPos[] = {
+				{ { -2.01953, -1.68555, 7.91612 }, 225.648 }
+		};
+		pelicanPos[] = {
+				{ { 2.05469, -1.40527, 7.91612 }, 117.265 }
+		};
+		palletPos[] = {
+				{ { -1.76758, 2.14648, -9.5529 }, 238 }
+		};
+	};
+	class Land_dp_bigTank_F : Default {
+		pelicanPos[] = {
+				{ { 0.326172, -0.174805, 3.57153 }, 253.033 }
+		};
+		palletPos[] = {
+				{ { -2.50781, -3.73828, 3.60099 }, 180 },
+				{ { -2.02734, 4.60449, 3.60099 }, 238 }
+		};
+	};
+	class Land_i_House_Small_02_V1_F : Default {
+		fridgePos[] = {
+				{ { -3.27148, 0.0527344, -0.734623 }, 270.037 }
+		};
+		bedPos[] = {
+				{ { 6.20313, -2.11328, -0.726701 }, 91.5496 }
+		};
+		couchPos[] = {
+				{ { -0.511719, -1.35156, -0.735134 }, 1.5394 },
+				{ { 6.25391, 0.796875, -0.726635 }, 2.02849 }
+		};
+		wardrobePos[] = {
+				{ { 1.42578, -2.14453, -0.732787 }, 274.294 }
+		};
+		cookerPos[] = {
+				{ { -3.22266, 2.26367, -0.727057 }, 270.633 }
+		};
+		chairPos[] = {
+				{ { 1.82813, -0.0859375, -0.732331 }, 294.633 },
+				{ { 3.58398, -2.61719, -0.730065 }, 188.429 }
+		};
+		tablePos[] = {
+				{ { 4.15039, 2.44727, -0.726453 }, 0.13829 }
+		};
+	};
+	class Land_i_House_Small_02_V2_F : Land_i_House_Small_02_V1_F { };
+	class Land_i_House_Small_02_V3_F : Land_i_House_Small_02_V1_F { };
+
+	class Land_Chapel_V1_F : Default {
+		wardrobePos[] = {
+				{ { 6.00195, -3.92969, -2.83485 }, 179.372 },
+				{ { 6.51758, 3.87891, -2.83486 }, 0.951279 }
+		};
+		chairPos[] = {
+				{ { -5.87305, -3.24414, -2.83491 }, 100.36 }
+		};
+		pelicanPos[] = {
+				{ { 6.86914, 0.0175781, -2.63492 }, 267.468 }
+		};
+		tablePos[] = {
+				{ { -4.00977, -1.36523, -2.83493 }, 269.847 }
+		};
+		shoeboxPos[] = {
+				{ { 1.01758, -1.21484, -2.83492 }, 277.811 },
+				{ { 3.37695, 3.05078, -2.83492 }, 276.733 },
+				{ { -1.56445, -4.63281, -1.51178 }, 1.34718 }
+		};
+		palletPos[] = {
+				{ { -6.60352, -0.580078, -2.83143 }, 107 }
+		};
+	};
+	class Land_Chapel_V2_F : Land_Chapel_V1_F {};
+
+	class Land_Chapel_Small_V1_F : Default {
+		chairPos[] = {
+				{ { 3.26953, 0.988281, -0.894284 }, 95.6997 }
+		};
+		pelicanPos[] = {
+				{ { 1.88379, -1, -0.54896 }, 269.755 }
+		};
+		shoeboxPos[] = {
+				{ { -0.423828, -1.18945, -0.9293 }, 272.97 }
+		};
+		palletPos[] = {
+				{ { 5.34961, 1.12695, -0.990805 }, 107 }
+		};
+	};
+	class Land_Chapel_Small_V2_F : Default {
+		chairPos[] = {
+				{ { 1.81445, 1.00391, -0.894289 }, 263.639 },
+				{ { 1.86621, -1.23633, -0.894289 }, 264.855 }
+		};
+		pelicanPos[] = {
+				{ { -0.228516, -0.994141, -0.54997 }, 264.521 }
+		};
+		tablePos[] = {
+				{ { 3.75586, -0.105469, -0.894377 }, 89.9709 }
+		};
+		shoeboxPos[] = {
+				{ { -0.333984, -0.396484, -0.9293 }, 269.367 }
+		};
+	};
+	class Land_Unfinished_Building_02_F : Default {
+		shelfPos[] = {
+				{ { -3.17188, -1.88086, 0.949918 }, 269.1 },
+				{ { 5.81641, 1.0918, -2.52319 }, 89.9583 },
+				{ { 5.91406, 0.357422, 0.950038 }, 90.0202 },
+				{ { -7.89746, 0.162109, 0.949969 }, 268.502 }
+		};
+		bedPos[] = {
+				{ { -4.95898, 3.81641, 0.949724 }, 2.84769 }
+		};
+		couchPos[] = {
+				{ { -0.923828, -4.65039, 0.949726 }, 89.1011 },
+				{ { -6.69727, 1.0293, -2.52443 }, 191.658 }
+		};
+		pelicanPos[] = {
+				{ { 0.539063, 3.33398, 0.949805 }, 48.6076 }
+		};
+		lockerPos[] = {
+				{ { 2.0166, 4.06055, 0.949846 }, 269.094 }
+		};
+		toolRackPos[] = {
+				{ { 1.55176, -1.90625, 0.949803 }, 90 }
+		};
+		shoeboxPos[] = {
+				{ { -2.40234, 3.50586, -2.29579 }, 316.814 }
+		};
+		palletPos[] = {
+				{ { 3.5625, -2.18359, -2.38678 }, 107 },
+				{ { -4.375, 3.11133, -2.43678 }, 107 },
+				{ { -7.12988, -2.27539, -2.38678 }, 107 }
+		};
+	};
+	class Land_Unfinished_Building_01_F : Default {
+		shelfPos[] = {
+				{ { 1.72168, -2.44336, 1.17171 }, 270.818 },
+				{ { -4.88184, 2.35938, 1.17171 }, 269.023 },
+				{ { -3.98438, 6.07227, -2.33344 }, 359.657 }
+		};
+		pelicanPos[] = {
+				{ { -0.557617, 5.36523, 4.46878 }, 7.08017 },
+				{ { -1.00684, 5.38867, -2.33351 }, 80.4785 }
+		};
+		toolRackPos[] = {
+				{ { -5.13184, 4.25, 1.1717 }, 270.006 }
+		};
+		shoeboxPos[] = {
+				{ { -0.368164, 5.74023, -2.31357 }, 26.9766 }
+		};
+		palletPos[] = {
+				{ { -2.25879, -1.67578, 1.17496 }, 107 },
+				{ { -3.6416, -1.50586, -2.30504 }, 107 }
+		};
+	};
+	class Land_i_Garage_V1_F : Default {
+		shelfPos[] = {
+				{ { -1.55078, 2.56055, -0.123878 }, 2.39618 },
+				{ { 1.91309, 2.57422, -0.123676 }, 359.724 }
+		};
+		fridgePos[] = {
+				{ { 2.73438, -2.17188, -0.123873 }, 182.516 },
+				{ { 4.35547, -1.73828, -0.123592 }, 152.206 }
+		};
+		cookerPos[] = {
+				{ { 0.046875, -2.07031, -0.124327 }, 178.265 },
+				{ { 4.73438, -0.0839844, -0.124328 }, 90 }
+		};
+		chairPos[] = {
+				{ { -3.29883, 2.07813, -0.115259 }, 30.7671 },
+				{ { -3.37695, -2.0293, -0.108249 }, 179.713 }
+		};
+		toolRackPos[] = {
+				{ { 0.271484, 2.77539, -0.124063 }, 0.0454407 }
+		};
+		freezerPos[] = {
+				{ { 1.19922, -1.98047, -0.116426 }, 176 }
+		};
+	};
+	class Land_i_Shop_01_V1_F : Default {
+		shelfPos[] = {
+				{ { -3.48438, -1.17578, -2.788 }, 268.197 },
+				{ { 3.14453, 4.45508, -2.78799 }, 180.136 }
+		};
+		fridgePos[] = {
+				{ { 1.59961, 4.04297, -2.78796 }, 85.053 }
+		};
+		bedPos[] = {
+				{ { -2.6543, 2.58594, 1.08346 }, 176.063 }
+		};
+		couchPos[] = {
+				{ { -2.60742, 2.36133, -2.78802 }, 178.997 }
+		};
+		wardrobePos[] = {
+				{ { 1.71094, 1.96094, 1.10037 }, 90.3614 }
+		};
+		cookerPos[] = {
+				{ { 1.47656, 0.724609, -2.7976 }, 91.1 }
+		};
+		chairPos[] = {
+				{ { -3.06836, -1.5293, 1.08361 }, 216.965 },
+				{ { -3.23242, 6.06445, 1.08373 }, 271.268 },
+				{ { -3.09766, 6.14063, -2.78284 }, 312.75 }
+		};
+		pelicanPos[] = {
+				{ { -2.66016, -3.25195, 1.07017 }, 300.566 }
+		};
+		freezerPos[] = {
+				{ { 1.54688, 2.13867, -2.7976 }, 91.1 }
+		};
+	};
+	class Land_i_Shop_01_V2_F : Land_i_Shop_01_V1_F {};
+	class Land_i_Shop_01_V3_F : Land_i_Shop_01_V1_F {};
+
+	class Land_u_Shop_01_V1_F : Default {
+		shelfPos[] = {
+				{ { 1.78516, -0.0390625, -2.90725 }, 88.4493 },
+				{ { -3.54688, -0.136719, -2.90725 }, 269.105 },
+				{ { -3.5332, -1.28711, -2.90719 }, 270 }
+		};
+		fridgePos[] = {
+				{ { 1.63086, 4.2793, -2.90721 }, 88.9036 }
+		};
+		bedPos[] = {
+				{ { -2.75195, -0.902344, 0.964377 }, 184.931 }
+		};
+		couchPos[] = {
+				{ { 0.546875, 1.35742, 0.963423 }, 8.58237 },
+				{ { -2.38672, 2.46484, -2.90728 }, 183.604 }
+		};
+		wardrobePos[] = {
+				{ { -3.41992, 2.08984, 0.964587 }, 270.137 },
+				{ { -3.36719, 5.67578, -2.90725 }, 269.105 }
+		};
+		cookerPos[] = {
+				{ { 3.19727, 4.55469, -2.90767 }, 182.053 }
+		};
+		chairPos[] = {
+				{ { 0.945313, 4.17188, 0.962904 }, 81.2836 },
+				{ { -2.9707, 3.94727, 0.964307 }, 304.317 }
+		};
+		pelicanPos[] = {
+				{ { -2.71094, 6.19141, 0.964416 }, 326.125 }
+		};
+		toolRackPos[] = {
+				{ { 2.03125, 1.91406, -2.90729 }, 90 }
+		};
+		palletPos[] = {
+				{ { 0.800781, -1.02539, 0.953138 }, 91.1 },
+				{ { 0.273438, 8.36328, -3.16686 }, 91.1 }
+		};
+	};
+	class Land_u_Shop_02_V1_F : Default {
+		shelfPos[] = {
+				{ { -0.126953, 2.28711, -2.7004 }, 358.842 },
+				{ { -4.41406, 4.35938, -2.69497 }, 355.221 }
+		};
+		bedPos[] = {
+				{ { -4.14648, -3.36914, 1.21215 }, 101.364 }
+		};
+		couchPos[] = {
+				{ { -0.917969, 1.4707, 1.21215 }, 275.214 },
+				{ { -0.701172, -2.85156, -2.69499 }, 179.577 }
+		};
+		wardrobePos[] = {
+				{ { -5.29297, 1.41602, 1.21222 }, 274.211 }
+		};
+		cookerPos[] = {
+				{ { -2, 3.81641, -2.69536 }, 88.9063 }
+		};
+		chairPos[] = {
+				{ { 2.67773, -3.54297, 1.21222 }, 123.36 },
+				{ { -3.26953, -3.9375, -2.69715 }, 150.356 }
+		};
+		pelicanPos[] = {
+				{ { 2.83984, 3.80273, 1.21219 }, 68.5479 }
+		};
+		lockerPos[] = {
+				{ { -5.19336, 1.49219, -2.69484 }, 269.367 }
+		};
+		toolRackPos[] = {
+				{ { -2.47852, -1.13086, -2.69459 }, 89.5903 }
+		};
+		palletPos[] = {
+				{ { 2.65234, -3.43359, -2.68103 }, 91.1 }
+		};
+	};
+	class Land_i_Shop_02_V1_F : Default {
+		shelfPos[] = {
+				{ { 3.26367, -1.38867, 1.21217 }, 90.8952 },
+				{ { -2.08203, -2.95703, -2.69497 }, 269.105 },
+				{ { -1.02539, -4.35742, -2.69497 }, 180.056 },
+				{ { -5.05078, 4.06055, -2.69497 }, 316.341 }
+		};
+		fridgePos[] = {
+				{ { -1.98047, 3.92383, -2.6833 }, 90.3202 },
+				{ { -5.27734, 3.7793, 1.21222 }, 279.71 }
+		};
+		bedPos[] = {
+				{ { -0.222656, 1.53906, 1.21215 }, 273.796 }
+		};
+		couchPos[] = {
+				{ { 0.4375, 1.25391, -2.70038 }, 267.738 },
+				{ { 0.759766, -3.41016, 1.21215 }, 92.1877 }
+		};
+		wardrobePos[] = {
+				{ { 2.8125, 4.05664, 1.21287 }, 9.59067 }
+		};
+		cookerPos[] = {
+				{ { -5.24609, 1.91992, -2.69537 }, 270 }
+		};
+		chairPos[] = {
+				{ { -2.39258, -3.73633, 1.21216 }, 148.243 },
+				{ { -4.91406, -1.41211, 1.21217 }, 307.249 },
+				{ { 4.62891, 3.41992, 1.21202 }, 300.318 },
+				{ { -4.88477, -3.74023, 1.21216 }, 223.381 }
+		};
+		pelicanPos[] = {
+				{ { 4.62695, -3.32031, 1.21216 }, 96.5193 }
+		};
+		tablePos[] = {
+				{ { -1.75391, -1.13086, -2.69498 }, 258.801 },
+				{ { -3.5957, -2.40234, 1.21215 }, 178.44 }
+		};
+		lockerPos[] = {
+				{ { -2.96484, -3.17773, -2.6967 }, 86.8359 }
+		};
+		toolRackPos[] = {
+				{ { -2.52344, -0.677734, -2.69483 }, 89.3848 }
+		};
+		freezerPos[] = {
+				{ { 0.943359, -4.0625, -2.69103 }, 179.201 }
+		};
+	};
+	class Land_i_Shop_02_V2_F : Land_i_Shop_02_V1_F {};
+	class Land_i_Shop_02_V3_F : Land_i_Shop_02_V1_F {};
+
+	class Land_d_Shop_01_V1_F : Default {
+		shelfPos[] = {
+				{ { 0.605469, -1.54102, -2.90717 }, 299.385 },
+				{ { 4.18555, 0.601563, -2.90725 }, 61.8762 }
+		};
+		fridgePos[] = {
+				{ { 4.00195, 2.49414, -2.90721 }, 89.2228 }
+		};
+		bedPos[] = {
+				{ { 3.50781, -0.328125, 0.96386 }, 7.76801 }
+		};
+		couchPos[] = {
+				{ { 0.0292969, -2.46289, 0.964392 }, 181.746 }
+		};
+		cookerPos[] = {
+				{ { 4.14648, 2.53125, 0.962579 }, 89.109 }
+		};
+		chairPos[] = {
+				{ { -0.193359, 1.15234, -2.90727 }, 319.613 },
+				{ { 3.76758, -3.32617, 0.964494 }, 152.858 }
+		};
+		pelicanPos[] = {
+				{ { 1.72461, 1.38086, 4.25002 }, 148.152 }
+		};
+		lockerPos[] = {
+				{ { 4.16016, 3.7207, -2.90727 }, 177.719 }
+		};
+		toolRackPos[] = {
+				{ { 4.625, -1.60938, -2.90729 }, 90.0519 }
+		};
+		palletPos[] = {
+				{ { -0.757813, 4.76367, -2.69686 }, 179.201 },
+				{ { -0.251953, -5.05664, -3.17686 }, 179.201 }
+		};
+	};
+	class Land_i_Addon_02_V1_F : Default {
+		fridgePos[] = {
+				{ { 3.17969, 3.81641, 0.0858996 }, 89.972 }
+		};
+		bedPos[] = {
+				{ { -0.154297, 3.38086, 0.085825 }, 276.397 }
+		};
+		wardrobePos[] = {
+				{ { -2.44922, 3.91602, 0.0865476 }, 359.106 }
+		};
+		chairPos[] = {
+				{ { 0.693359, 0.376953, 0.0858068 }, 160.478 }
+		};
+		tablePos[] = {
+				{ { -2.86133, 1.26953, 0.0858243 }, 269.367 }
+		};
+		palletPos[] = {
+				{ { 0.777344, -0.1875, 2.85581 }, 187.201 }
+		};
+	};
+	class Land_u_Addon_02_V1_F : Land_i_Addon_02_V1_F {};
+
+	class Land_d_Addon_02_V1_F : Default {
+		couchPos[] = {
+				{ { -0.962891, 2.74805, 0.41301 }, 261.447 }
+		};
+		pelicanPos[] = {
+				{ { 1.94531, 2.58301, 0.439629 }, 350.421 }
+		};
+		palletPos[] = {
+				{ { -2.35156, 0.995117, 0.423958 }, 187.201 },
+				{ { 3.19336, 1.2168, 0.413958 }, 187.201 }
+		};
+	};
+	class Land_d_Stone_HouseSmall_V1_F : Default {
+		shelfPos[] = {
+				{ { 7.36328, 2.99121, -0.59085 }, 87.8075 }
+		};
+		fridgePos[] = {
+				{ { 3.63281, -1.93262, -0.616286 }, 268.761 }
+		};
+		bedPos[] = {
+				{ { 10.1387, 2.39941, -0.585113 }, 271.572 }
+		};
+		couchPos[] = {
+				{ { -0.128906, -0.958008, -0.578321 }, 80.7286 }
+		};
+		tablePos[] = {
+				{ { 1.56055, 2.4541, -0.599906 }, 3.8147 },
+				{ { 7.77539, -0.0078125, -0.597849 }, 341.959 }
+		};
+		palletPos[] = {
+				{ { -6.20508, 1.12988, -0.453797 }, 187.201 }
+		};
+	};
+	class Land_i_Stone_HouseSmall_V3_F : Default {
+		shelfPos[] = {
+				{ { 0.423828, 0.0566406, -0.65061 }, 91.2415 },
+				{ { 8.48047, -0.707031, -0.614687 }, 178.388 }
+		};
+		fridgePos[] = {
+				{ { 1.63477, -0.37207, -0.645178 }, 237.969 }
+		};
+		bedPos[] = {
+				{ { -7.63867, 3.78809, -0.637772 }, 284.46 }
+		};
+		couchPos[] = {
+				{ { -3.54297, 3.81445, -0.635458 }, 279.006 },
+				{ { 6.89453, 3.41211, -0.618565 }, 306.201 }
+		};
+		wardrobePos[] = {
+				{ { -5.28516, -0.365234, -0.67601 }, 152.535 }
+		};
+		cookerPos[] = {
+				{ { 3.74414, -0.546875, -0.636206 }, 178.279 }
+		};
+		chairPos[] = {
+				{ { -0.378906, 4.13086, -0.630233 }, 34.7144 }
+		};
+		tablePos[] = {
+				{ { -2.85547, 1.70898, -0.659347 }, 164.896 },
+				{ { 5.85547, 2.18652, -0.626413 }, 196.919 }
+		};
+		palletPos[] = {
+				{ { 9.33398, -2.30273, -10 }, 187.201 }
+		};
+	};
+	class Land_i_Stone_HouseSmall_V2_F : Default {
+		shelfPos[] = {
+				{ { 4.76758, -0.796875, -0.631255 }, 179.539 },
+				{ { -9.4043, 1.01563, -0.670586 }, 267.808 }
+		};
+		fridgePos[] = {
+				{ { 1.48047, -0.0341797, -0.645909 }, 258.003 }
+		};
+		bedPos[] = {
+				{ { -1.64648, 4.09277, -0.631337 }, 292.711 }
+		};
+		couchPos[] = {
+				{ { -5.07031, 0.208008, -0.67044 }, 95.3583 }
+		};
+		wardrobePos[] = {
+				{ { -0.00585938, -0.03125, -0.648986 }, 126.318 }
+		};
+		cookerPos[] = {
+				{ { 5.89258, 4.7959, -0.619837 }, 359.875 }
+		};
+		chairPos[] = {
+				{ { 8.50977, 0.861328, -0.614511 }, 139.91 },
+				{ { 8.43945, 3.40625, -0.614825 }, 57.7087 }
+		};
+		pelicanPos[] = {
+				{ { -8.76563, 4.50488, -0.630112 }, 339.768 }
+		};
+		tablePos[] = {
+				{ { -5.22852, 2.18652, -0.655036 }, 2.32285 }
+		};
+	};
+	class Land_i_Stone_HouseSmall_V1_F : Land_i_Stone_HouseSmall_V2_F {};
+
+	class Land_i_Stone_Shed_V3_F : Default {
+		bedPos[] = {
+				{ { -2.78711, 2.84375, -0.126943 }, 174.67 }
+		};
+		wardrobePos[] = {
+				{ { 0.425781, 4.04688, -0.139276 }, 1.30153 }
+		};
+		cookerPos[] = {
+				{ { -1.9375, -0.209961, -0.127349 }, 176.707 }
+		};
+		tablePos[] = {
+				{ { 2.2207, 2.56445, -0.126943 }, 75.8372 }
+		};
+		palletPos[] = {
+				{ { -0.837891, 1.98828, 2.30655 }, 187.201 }
+		};
+	};
+	class Land_i_Stone_Shed_V2_F : Default {
+		fridgePos[] = {
+				{ { 2.71875, 3.87891, -0.126784 }, 68.5485 }
+		};
+		bedPos[] = {
+				{ { -2.76758, 2.82813, -0.126943 }, 180.094 }
+		};
+		couchPos[] = {
+				{ { -1.91602, 0.928711, 0.00655437 }, 98.2006 }
+		};
+		pelicanPos[] = {
+				{ { -1.00195, 2.06738, 2.21351 }, 305.061 }
+		};
+		tablePos[] = {
+				{ { 0.671875, 1.21875, -0.126942 }, 176.837 }
+		};
+	};
+	class Land_i_Stone_Shed_V1_F : Default {
+		bedPos[] = {
+				{ { -2.9209, 0.884766, -0.126943 }, 183.73 }
+		};
+		couchPos[] = {
+				{ { -0.408203, 3.18359, -0.126948 }, 274.981 }
+		};
+		chairPos[] = {
+				{ { -3.11133, 3.29883, -0.126935 }, 304.61 }
+		};
+		pelicanPos[] = {
+				{ { -0.580078, 1.9082, 2.23559 }, 286.89 }
+		};
+		tablePos[] = {
+				{ { 0.279297, -0.111328, -0.126944 }, 178.227 }
+		};
+		toolRackPos[] = {
+				{ { 3.20313, 3.35938, -0.126961 }, 88.6889 }
+		};
+	};
+	class Land_d_Stone_HouseBig_V1_F : Default {
+		shelfPos[] = {
+				{ { -1.88477, 1.23047, -1.6673 }, 89.1051 }
+		};
+		fridgePos[] = {
+				{ { -2.71875, -3.03711, -1.66745 }, 188.243 }
+		};
+		bedPos[] = {
+				{ { -2.0498, -2.62891, 1.19249 }, 98.1494 }
+		};
+		couchPos[] = {
+				{ { 2.18457, -1.48438, -1.61955 }, 5.4863 }
+		};
+		wardrobePos[] = {
+				{ { -1.54102, 1.11914, 1.19367 }, 340.004 }
+		};
+		chairPos[] = {
+				{ { -0.714844, -3.05078, -1.6675 }, 178.582 }
+		};
+		pelicanPos[] = {
+				{ { 4.28516, -0.453125, 1.20476 }, 87.2772 }
+		};
+		tablePos[] = {
+				{ { -2.75977, -0.654297, 1.19248 }, 274.611 }
+		};
+		shoeboxPos[] = {
+				{ { -0.728516, 0.9375, -0.647514 }, 33.2239 }
+		};
+		palletPos[] = {
+				{ { 1.16309, -2.74805, 1.18598 }, 98.2006 }
+		};
+	};
+	class Land_i_Stone_HouseBig_V3_F : Default {
+		shelfPos[] = {
+				{ { -1.8916, 1.73047, -1.68199 }, 270.888 },
+				{ { 0.605469, 3.55469, 1.19272 }, 359.924 },
+				{ { 4.22168, -1.08594, 1.19272 }, 175.587 }
+		};
+		fridgePos[] = {
+				{ { -1.59277, -0.886719, -1.68134 }, 246.969 }
+		};
+		bedPos[] = {
+				{ { -0.640625, -0.460938, 1.19249 }, 91.5507 }
+		};
+		couchPos[] = {
+				{ { 3.75488, 1.30664, -1.66749 }, 9.91046 }
+		};
+		wardrobePos[] = {
+				{ { -1.41113, 2.97461, 1.19534 }, 300.8 }
+		};
+		chairPos[] = {
+				{ { -1.52539, 1.28516, 1.1925 }, 266.593 }
+		};
+		tablePos[] = {
+				{ { 0.642578, -0.855469, -1.67561 }, 187.85 },
+				{ { 2.20898, -0.970703, 1.19249 }, 176.776 }
+		};
+		shoeboxPos[] = {
+				{ { 0.287109, 3.42773, -0.594017 }, 98.2006 }
+		};
+	};
+	class Land_i_Stone_HouseBig_V2_F : Land_i_Stone_HouseBig_V3_F { };
+	class Land_i_Stone_HouseBig_V1_F : Land_i_Stone_HouseBig_V3_F { };
+
+	class Land_d_House_Big_02_V1_F : Default {
+		shelfPos[] = {
+				{ { 0.669922, -1.39453, -3.06634 }, 90 },
+				{ { 2.7373, 4.13477, 0.463613 }, 90 },
+				{ { -5.60547, 4.73633, -3.0663 }, 338.985 }
+		};
+		fridgePos[] = {
+				{ { 2.41797, 0.193359, -3.06639 }, 156.125 }
+		};
+		bedPos[] = {
+				{ { 0.832031, 1.24023, 0.463591 }, 73.8195 }
+		};
+		couchPos[] = {
+				{ { 1.16406, 3.83398, -3.06641 }, 285.269 }
+		};
+		wardrobePos[] = {
+				{ { -5.1709, 0.628906, 0.46367 }, 210.054 }
+		};
+		cookerPos[] = {
+				{ { -5.51367, -5.44727, -2.91562 }, 283.817 }
+		};
+		chairPos[] = {
+				{ { -5.57617, 0.324219, -3.06635 }, 249.612 },
+				{ { -5.49219, -2.3457, 0.523658 }, 265.17 }
+		};
+		pelicanPos[] = {
+				{ { 1.33398, 6.15625, 0.40361 }, 34.738 }
+		};
+		tablePos[] = {
+				{ { -1.68457, 4.60742, 0.463591 }, 357.124 }
+		};
+		palletPos[] = {
+				{ { 1.80078, -5.45703, -2.37291 }, 98.2006 }
+		};
+	};
+	class Land_u_House_Big_02_V1_F : Default {
+		shelfPos[] = {
+				{ { -3.04102, -1.03125, 0.962879 }, 0 },
+				{ { -3.25195, 0.0253906, -2.40712 }, 185.201 }
+		};
+		fridgePos[] = {
+				{ { 2.1748, -1.3457, -2.40712 }, 85 }
+		};
+		bedPos[] = {
+				{ { 3.27832, 0.75, 0.939383 }, 94.647 }
+		};
+		couchPos[] = {
+				{ { 3.25391, 0.902344, -2.57062 }, 90.0017 },
+				{ { -3.19824, 2.6543, 0.939378 }, 187.634 }
+		};
+		wardrobePos[] = {
+				{ { 1.70801, 4.85156, 0.939409 }, 359.848 }
+		};
+		cookerPos[] = {
+				{ { -1.25, -3.69336, -2.40712 }, 185.201 }
+		};
+		chairPos[] = {
+				{ { -3.16992, 0.291016, 0.939465 }, 245.407 },
+				{ { -3.71875, -3.23047, -2.40712 }, 215.201 },
+				{ { 3.91016, 4.31836, -2.40712 }, 49.2006 }
+		};
+		pelicanPos[] = {
+				{ { 3.94434, 4.5, 0.939445 }, 20.8601 }
+		};
+		tablePos[] = {
+				{ { -1.43262, 2.4707, 0.939383 }, 85.2193 },
+				{ { -3.81055, 3.12305, -2.40712 }, 272.201 }
+		};
+	};
+	class Land_i_House_Big_02_V1_F : Default {
+		shelfPos[] = {
+				{ { -4.27734, 0.900391, 0.757645 }, 273.29 },
+				{ { 4.46875, 4.30664, -2.64969 }, 89.2152 }
+		};
+		fridgePos[] = {
+				{ { 1.91797, -1.44141, -2.64966 }, 87.5481 }
+		};
+		bedPos[] = {
+				{ { 2.97656, 0.740234, 0.757623 }, 89.4966 }
+		};
+		couchPos[] = {
+				{ { 2.94531, 0.919922, -2.64972 }, 92.2641 },
+				{ { 0.375, 4.0293, 0.757618 }, 270.255 }
+		};
+		wardrobePos[] = {
+				{ { 4.24219, 4.36523, 0.757747 }, 92.8048 }
+		};
+		cookerPos[] = {
+				{ { -1.25, -3.58008, -2.63541 }, 184.944 }
+		};
+		chairPos[] = {
+				{ { -3.75586, 0.412109, -2.6497 }, 231.099 },
+				{ { -3.82227, -1.31445, 0.757625 }, 308.225 },
+				{ { 2.06641, 4.59961, -2.6497 }, 7.39194 }
+		};
+		pelicanPos[] = {
+				{ { 3.12305, 6.2168, 0.697647 }, 31.1078 }
+		};
+		tablePos[] = {
+				{ { -4.03906, 2.14258, -2.64971 }, 268.404 }
+		};
+		freezerPos[] = {
+				{ { -3.42383, -1.25195, -2.63541 }, 360 }
+		};
+	};
+	class Land_i_House_Big_02_V2_F : Land_i_House_Big_02_V1_F {};
+	class Land_i_House_Big_02_V3_F : Land_i_House_Big_02_V1_F {};
+
+	class Land_i_House_Small_03_V1_F : Default {
+		shelfPos[] = {
+				{ { 0.273438, 1.06055, -0.398047 }, 178.213 },
+				{ { 5.03125, 4.3418, -0.398025 }, 55.5188 }
+		};
+		fridgePos[] = {
+				{ { 1.30078, 3.96289, -0.398048 }, 85.9457 }
+		};
+		bedPos[] = {
+				{ { -4.09961, 3.68945, -0.39807 }, 268.584 }
+		};
+		couchPos[] = {
+				{ { -4.36719, -0.742188, -0.398075 }, 183.015 }
+		};
+		wardrobePos[] = {
+				{ { -2.23438, -1.9375, -0.397993 }, 183.121 },
+				{ { 5.0918, 1.56055, -0.397933 }, 90.0051 }
+		};
+		cookerPos[] = {
+				{ { 1.32813, 2.72852, -0.398477 }, 89.3721 }
+		};
+		chairPos[] = {
+				{ { -1.32031, -3.61523, -0.694491 }, 357.783 },
+				{ { -2.01563, 4.09961, -0.398058 }, 39.0805 },
+				{ { 2.7168, -5.76758, -0.398058 }, 216.364 },
+				{ { 4.70508, -5.81641, -0.398057 }, 145.992 }
+		};
+		tablePos[] = {
+				{ { 4.98828, -2.87695, -0.398071 }, 89.3266 }
+		};
+		palletPos[] = {
+				{ { -2.04102, -5.83398, -10 }, 360 }
+		};
+	};
+	class Land_u_House_Small_02_V1_F : Default {
+		shelfPos[] = {
+				{ { -0.0351563, -3.38086, -0.744917 }, 179.437 }
+		};
+		fridgePos[] = {
+				{ { 0.369141, -0.267578, -0.743873 }, 91.2655 }
+		};
+		bedPos[] = {
+				{ { 6.56055, 1.06641, -0.735898 }, 9.02117 }
+		};
+		couchPos[] = {
+				{ { 5.23047, -2.19336, -0.738171 }, 98.8483 }
+		};
+		wardrobePos[] = {
+				{ { 1.43945, -2.625, -0.74296 }, 270.844 }
+		};
+		cookerPos[] = {
+				{ { -3.2168, 0.0839844, -0.745483 }, 269.137 }
+		};
+		tablePos[] = {
+				{ { 4.68359, 2.11523, -0.736894 }, 356.553 }
+		};
+	};
+	class Land_d_House_Small_02_V1_F : Default {
+		fridgePos[] = {
+				{ { 7.47266, -0.1875, -0.999054 }, 79.7224 }
+		};
+		bedPos[] = {
+				{ { 2.45508, -0.201172, -0.981503 }, 187.343 },
+				{ { 6.89844, 1.89258, -1.00027 }, 4.89485 }
+		};
+		chairPos[] = {
+				{ { 4.09766, -2.04297, -0.9844 }, 210.762 },
+				{ { 7.12305, -2.12109, -0.994862 }, 135.992 }
+		};
+		palletPos[] = {
+				{ { -0.767578, -0.916016, -10 }, 276 }
+		};
+	};
+	class Land_Lighthouse_small_F : Default {
+		shelfPos[] = {
+				{ { -1.35547, -2.18555, -3.42188 }, 191.463 }
+		};
+		chairPos[] = {
+				{ { -1.66309, 1.66602, -3.4219 }, 311.523 }
+		};
+		pelicanPos[] = {
+				{ { -1.16602, 1.19336, -0.0622997 }, 276 }
+		};
+		toolRackPos[] = {
+				{ { 5.70996, 0.291016, -3.42193 }, 90 }
+		};
+		palletPos[] = {
+				{ { 4.57031, -1.70898, -3.3823 }, 276 }
+		};
+	};
+	class Land_WIP_F : Default {
+		shelfPos[] = {
+				{ { -0.957031, -4.42969, 12.2029 }, 358.087 },
+				{ { -0.919922, -5.26563, 0.396693 }, 178.896 },
+				{ { -2.20313, 5.43164, 0.323569 }, 1.17821 },
+				{ { 2.03516, -6.10938, 0.397176 }, 14.9686 },
+				{ { 5.12598, -5.45313, 8.32229 }, 181.318 },
+				{ { -6.88477, 5.69922, 8.33006 }, 356.041 },
+				{ { -6.33203, 9.34375, 4.2993 }, 267.673 },
+				{ { -5.47656, -11.3945, 0.348352 }, 178.537 },
+				{ { -13.2139, 1.17383, 0.390545 }, 90.2068 },
+				{ { -13.2334, -2.94336, 4.2993 }, 86.318 },
+				{ { 14.3809, 6.43945, 4.29953 }, 181.218 },
+				{ { -9.65625, -12.791, 12.2983 }, 180.234 },
+				{ { -12.9834, -10.2734, 0.319984 }, 89.2757 }
+		};
+		fridgePos[] = {
+				{ { -3.39063, -7.63477, 0.37021 }, 91.2212 }
+		};
+		bedPos[] = {
+				{ { 6.81055, 9.57617, 0.352686 }, 180.055 }
+		};
+		couchPos[] = {
+				{ { -0.21875, 7.48438, 0.331759 }, 112.451 },
+				{ { 4.41992, -9.9707, 4.29707 }, 53.0242 },
+				{ { -10.5244, 4.81836, 12.2371 }, 79.3503 },
+				{ { 17.5283, -2.80859, 0.327501 }, 353.002 },
+				{ { -16.7871, 9.91406, 0.370766 }, 207.103 }
+		};
+		chairPos[] = {
+				{ { -8.98828, 4.98828, 0.37769 }, 307.949 },
+				{ { 10.1426, -6.32813, 0.338495 }, 11.358 },
+				{ { -14, 6.58594, 4.29928 }, 138.034 },
+				{ { -15.9365, 7.62305, 8.31643 }, 107.749 },
+				{ { -14.8984, 9.5293, 8.31799 }, 137.305 },
+				{ { -9.3418, -17.8223, 0.310942 }, 136.534 }
+		};
+		filingPos[] = {
+				{ { 5.73438, 6.36328, 4.29928 }, 183.682 }
+		};
+		pelicanPos[] = {
+				{ { 9.72461, -10.5352, 8.31388 }, 118.227 },
+				{ { -7.59766, -14.3633, 12.2979 }, 67.7859 },
+				{ { 23.1895, 1.08008, 4.30315 }, 277.918 }
+		};
+		tablePos[] = {
+				{ { -17.8965, 10.6133, 8.31348 }, 303.712 }
+		};
+		lockerPos[] = {
+				{ { -7.36816, -9.47852, 4.29681 }, 90.0775 },
+				{ { -9.44043, -11.1738, 4.29663 }, 177.772 },
+				{ { -8.14551, -14.9766, 8.31173 }, 180.659 }
+		};
+		toolRackPos[] = {
+				{ { 1.60059, -5.43945, 0.405369 }, 180.797 },
+				{ { 2.0918, 5.63867, 0.351912 }, 357.597 },
+				{ { -2.88086, -12.0996, -0.00143886 }, 360 },
+				{ { 14.0498, -5.80859, 4.29928 }, 359.535 },
+				{ { -10.501, -15.4023, 8.312 }, 179.968 }
+		};
+		shoeboxPos[] = {
+				{ { 9.66016, 6.35352, 4.31021 }, 190.576 },
+				{ { -3.89844, -11.2754, 0.345911 }, 173.847 },
+				{ { 28.7275, -2.26172, 0.302157 }, 142.217 }
+		};
+		palletPos[] = {
+				{ { -2.81641, 3.73828, 12.2086 }, 15 },
+				{ { -3.99902, -3.44336, 12.2086 }, 15 },
+				{ { 0.981445, -6.29688, 8.33855 }, 15 },
+				{ { 3.71094, 6.82617, 4.32855 }, 182 },
+				{ { -3.94727, -10.5742, 8.33855 }, 15 },
+				{ { -5.66602, -9.97656, 4.32855 }, 345.6 },
+				{ { -8.00586, 10.6641, 4.32855 }, 182 },
+				{ { -13.8838, -1.19727, 0.338561 }, 360 },
+				{ { -15.0645, 1.19141, 8.33855 }, 15 },
+				{ { 12.4063, 10.8125, 0.368561 }, 265.5 },
+				{ { -17.6748, -1.5, 0.368561 }, 265.5 },
+				{ { -17.7871, -4.42773, 4.32855 }, 182 },
+				{ { -15.6992, -10.6797, 0.338561 }, 360 },
+				{ { 23.7109, -1.23633, 8.23855 }, 289.5 }
+		};
+	};
+	class Land_i_Addon_03mid_V1_F : Default {
+		couchPos[] = {
+				{ { -3.33203, 0.30957, -0.0338135 }, 176.588 }
+		};
+		pelicanPos[] = {
+				{ { 3.95313, 1.0918, -0.0338745 }, 101.571 }
+		};
+		palletPos[] = {
+				{ { 0.703125, -0.210938, -0.000281334 }, 15 }
+		};
+	};
+	class Land_i_Addon_03_V1_F : Default {
+		chairPos[] = {
+				{ { 2.5, -0.929688, -0.0765305 }, 346.611 },
+				{ { -2.61914, -1.03125, -0.0765266 }, 17.8128 }
+		};
+		palletPos[] = {
+				{ { 3.56641, 2.50391, -10 }, 344 },
+				{ { -3.58594, 2.75, -10 }, 15 }
+		};
+	};
+	class Land_i_Addon_04_V1_F : Default {
+		chairPos[] = {
+				{ { -3.78711, -1.16211, 0.050499 }, 21.1258 }
+		};
+		pelicanPos[] = {
+				{ { -2.04883, 0.263672, 0.0532227 }, 215.663 }
+		};
+		palletPos[] = {
+				{ { -0.130859, 2.37695, -10 }, 344 }
+		};
+	};
+	class Land_LightHouse_F : Default {
+		chairPos[] = {
+				{ { 1.64063, -3.0918, -11.1061 }, 121.303 }
+		};
+		pelicanPos[] = {
+				{ { -0.318359, -3.61328, 8.12647 }, 220.41 },
+				{ { 0.697266, -4.32227, -11.1665 }, 248.695 }
+		};
+		shoeboxPos[] = {
+				{ { -0.214844, -4.24609, -11.1689 }, 225.43 },
+				{ { 1.87109, -4.14453, 8.13151 }, 344 }
+		};
+	};
+	class Land_GH_Gazebo_F : Default {
+		shelfPos[] = {
+				{ { -5.49609, -1.95898, -1.76052 }, 180.195 }
+		};
+		bedPos[] = {
+				{ { 6.02734, 1.77734, -1.76616 }, 1.01784 }
+		};
+		couchPos[] = {
+				{ { -5.99219, 1.83398, -1.76616 }, 175.953 }
+		};
+		wardrobePos[] = {
+				{ { 6.67773, -0.783203, -1.7661 }, 88.4633 }
+		};
+		pelicanPos[] = {
+				{ { -0.517578, 0.345703, -1.76613 }, 47.9599 }
+		};
+		tablePos[] = {
+				{ { -6.71094, -0.613281, -1.76616 }, 269.48 }
+		};
+		palletPos[] = {
+				{ { 1.63672, 1.53516, -1.76332 }, 358 },
+				{ { -2.40039, 2.01953, -1.76332 }, 290 }
+		};
+	};
+	class Land_GH_MainBuilding_right_F : Default {
+		shelfPos[] = {
+				{ { 1.30273, -7.28516, -1.06903 }, 104.132 }
+		};
+		fridgePos[] = {
+				{ { 3.30859, 1.99219, -5.50915 }, 196.553 },
+				{ { 0.769531, -6.8125, 3.37646 }, 185.624 },
+				{ { 1.35352, -11.4238, 3.83718 }, 14.2214 },
+				{ { -14.1602, 5.25781, -1.06901 }, 244.418 }
+		};
+		bedPos[] = {
+				{ { -1.76758, 1.86133, 3.37644 }, 290.999 },
+				{ { -4.86328, 1.69727, 3.37644 }, 18.4142 },
+				{ { -1.08398, 5.9043, -1.06907 }, 197.348 },
+				{ { 6.74609, -1.57813, 3.37643 }, 24.4042 },
+				{ { -3.48242, 6.48047, -1.06907 }, 21.0007 },
+				{ { 7.32031, 1.24609, -1.06908 }, 25.1119 },
+				{ { -2.52734, 10.3438, -5.50922 }, 20.4328 },
+				{ { 8.82813, 7.0293, -5.50922 }, 26.5858 },
+				{ { -10.5645, 8.5625, -1.06908 }, 24.9801 },
+				{ { -13.9043, 3.83008, 3.37644 }, 196.357 },
+				{ { -12.2109, 10.7422, -5.50922 }, 201.181 }
+		};
+		couchPos[] = {
+				{ { 4.1582, 8.32813, -5.50922 }, 23.8942 }
+		};
+		wardrobePos[] = {
+				{ { 1.61328, -2.05078, -1.06904 }, 192.819 },
+				{ { 1.00977, 5.89063, -5.50915 }, 195.313 },
+				{ { 5.52734, -4.44531, 3.37651 }, 170.013 },
+				{ { -9.79492, 0.988281, -1.06905 }, 197.093 },
+				{ { -10.9043, -3.53125, 3.37651 }, 199.701 },
+				{ { -8.48242, 13.2871, -5.50919 }, 104.652 }
+		};
+		cookerPos[] = {
+				{ { 2.11328, 2.37109, -5.50961 }, 195.987 }
+		};
+		chairPos[] = {
+				{ { -1.2207, -2.72656, 3.37645 }, 193.938 },
+				{ { 3.36523, 2.84375, -1.06907 }, 101.74 },
+				{ { -10.2559, 1.17578, 3.37646 }, 283.356 },
+				{ { -9.08008, 5.47266, -1.06906 }, 268.731 },
+				{ { -0.0585938, 11.2246, -5.50921 }, 304.857 },
+				{ { -11.377, 2.25977, 3.37684 }, 104.34 }
+		};
+		filingPos[] = {
+				{ { -4.86328, -8.57031, -1.069 }, 282.224 }
+		};
+		pelicanPos[] = {
+				{ { -0.679688, -0.609375, -1.06897 }, 139.998 },
+				{ { -8.87891, 5.31641, -5.50921 }, 202.579 },
+				{ { -15.1152, -1.07031, 3.37656 }, 307.599 },
+				{ { -15.6387, -10.627, -5.5092 }, 130.018 }
+		};
+		tablePos[] = {
+				{ { -0.0839844, 1.53516, -1.06908 }, 192.028 },
+				{ { 2.57422, -1.21094, 3.37644 }, 104.506 },
+				{ { -6.80078, 3.25781, -1.06908 }, 194.092 },
+				{ { -7.66602, -1.30859, 3.37644 }, 194.642 },
+				{ { -0.845703, 9.5293, -5.50922 }, 284.72 },
+				{ { -14.7227, 0.589844, 3.37643 }, 196.169 }
+		};
+		lockerPos[] = {
+				{ { -4.8125, -5.11523, -1.06906 }, 15.9695 }
+		};
+		palletPos[] = {
+				{ { 8.36523, 7.57813, -1.0171 }, 37.2845 },
+				{ { -4.83398, 11.3008, -1.0171 }, 357 },
+				{ { -10.7734, 12.9609, -1.0171 }, 96 }
+		};
+		freezerPos[] = {
+				{ { -6.79492, -6.25781, -1.0171 }, 283 }
+		};
+	};
+	class Land_GH_MainBuilding_middle_F : Default {
+		shelfPos[] = {
+				{ { 2.69238, 3.21484, -8.37759 }, 179.858 },
+				{ { -5.88867, 0.726563, -3.91598 }, 105.411 },
+				{ { 5.24609, 11.0605, -3.91652 }, 270.656 },
+				{ { -15.541, -6.12305, 0.521439 }, 14.5104 },
+				{ { 14.248, -9.52734, 0.521439 }, 256.088 }
+		};
+		fridgePos[] = {
+				{ { -15.1309, -2.81836, 0.521437 }, 13.9473 }
+		};
+		couchPos[] = {
+				{ { -13.6484, -0.876953, 4.97235 }, 195.253 }
+		};
+		wardrobePos[] = {
+				{ { 14.1123, -3.39453, 0.521494 }, 346.293 }
+		};
+		chairPos[] = {
+				{ { 6.45313, 0.845703, -3.91599 }, 247.596 },
+				{ { -12.459, -3.85938, 0.521418 }, 73.8369 },
+				{ { -5.72656, 11.7988, -3.90909 }, 37.8367 },
+				{ { 12.8564, 5.21484, 0.521416 }, 112.013 },
+				{ { -0.03125, -16.3359, 4.97174 }, 184.279 },
+				{ { -6.17188, -16.8965, 0.521433 }, 185.345 },
+				{ { -3.02246, 21.9258, -9.01524 }, 195.022 }
+		};
+		filingPos[] = {
+				{ { -4.35156, 3.39844, -8.37744 }, 205.858 },
+				{ { -16.0371, -5.33398, 0.521456 }, 268.695 }
+		};
+		pelicanPos[] = {
+				{ { -1.16016, 8.58008, -9.13322 }, 288.066 }
+		};
+		tablePos[] = {
+				{ { 14.1885, 0.773438, 4.97233 }, 74.1932 }
+		};
+		lockerPos[] = {
+				{ { -16.1133, -9.58789, 0.521435 }, 196.292 }
+		};
+		toolRackPos[] = {
+				{ { 4.72754, 4.68555, -8.37762 }, 90.6886 }
+		};
+		palletPos[] = {
+				{ { -6.27344, 10.8496, 0.576112 }, 251 },
+				{ { -14.5596, -8.70117, 0.576112 }, 251 }
+		};
+		freezerPos[] = {
+				{ { 13.5488, -5.76953, 0.576112 }, 251 }
+		};
+	};
+	class Land_GH_MainBuilding_left_F : Default {
+		shelfPos[] = {
+				{ { -1.02734, -7.3623, -1.06904 }, 255.179 },
+				{ { -8.17188, 1.39258, -1.06893 }, 255.243 }
+		};
+		fridgePos[] = {
+				{ { 0.386719, -6.67578, 3.37653 }, 165.302 },
+				{ { -6.97656, -0.431641, -1.06901 }, 172.301 },
+				{ { 11.0391, -3.58008, 3.37646 }, 160.127 }
+		};
+		bedPos[] = {
+				{ { -1.83594, -0.640625, 3.37644 }, 165.641 },
+				{ { -1.66797, -1.12305, -1.06908 }, 167.488 },
+				{ { -4.0625, 8.01367, -5.50922 }, 163.408 },
+				{ { 0.40625, 9.14844, -5.50922 }, 345.918 },
+				{ { 9.70313, 1.55859, 3.37644 }, 345.36 },
+				{ { 8.6582, 6.32422, -5.50922 }, 165.048 },
+				{ { 8.26172, 7.47852, -1.06907 }, 348.908 },
+				{ { -8.69336, 7.13867, -5.50922 }, 165.97 },
+				{ { 14.4043, 2.89258, 3.37643 }, 345.42 }
+		};
+		couchPos[] = {
+				{ { 1.23242, 5.67188, -1.06907 }, 350.466 },
+				{ { 2.88281, 10.3672, -5.50922 }, 167.118 },
+				{ { -2.56641, -12.8418, 3.83704 }, 261.041 },
+				{ { 12.6992, 8.72461, -1.06908 }, 347.645 }
+		};
+		wardrobePos[] = {
+				{ { -2.58008, 2.1709, -5.5092 }, 167.528 },
+				{ { 10.0078, 0.827148, -1.06901 }, 166.411 }
+		};
+		cookerPos[] = {
+				{ { -0.742188, -6.74805, 3.37604 }, 172.375 }
+		};
+		chairPos[] = {
+				{ { 2.68555, 2.31445, 3.37668 }, 40.9297 },
+				{ { -3.30078, 4.20117, -1.06905 }, 248.905 },
+				{ { -1.18945, 6.18945, -5.5092 }, 164.369 },
+				{ { -5.23242, 4.45508, -1.05902 }, 44.6127 },
+				{ { -6.14648, -4.5498, 3.37645 }, 208.166 },
+				{ { -4.16016, -12.752, -1.06894 }, 256.238 },
+				{ { 7.50781, 12.041, -5.50921 }, 65.2004 },
+				{ { 14.0137, 5.32227, -1.06907 }, 108.336 },
+				{ { 15.3516, 0.828125, 3.37644 }, 89.9532 },
+				{ { 9.25, 12.7598, -5.50921 }, 264.026 },
+				{ { 12.9297, 9.67383, -5.50699 }, 117.822 }
+		};
+		filingPos[] = {
+				{ { 1.83984, -8.47363, -1.069 }, 160.823 },
+				{ { 15.752, -11.3574, -5.50917 }, 164.969 }
+		};
+		pelicanPos[] = {
+				{ { 8.55664, -2.79004, 3.37645 }, 36.1098 }
+		};
+		tablePos[] = {
+				{ { 0.234375, 1.52344, -1.06908 }, 163.886 },
+				{ { 1.57422, -2.80469, 3.37644 }, 166.456 },
+				{ { -3.63672, -1.45117, 3.37644 }, 75.2335 },
+				{ { 4.10352, 2.30273, 3.37644 }, 254.907 },
+				{ { 6.85156, 3.26758, -1.06908 }, 163.381 },
+				{ { 5.79492, 7.67578, -5.50922 }, 167.512 },
+				{ { -8.66992, 4.4082, -5.50922 }, 256.113 },
+				{ { -4.61719, -11.2539, -1.06898 }, 257.112 },
+				{ { 12.7383, 11.5723, -5.50922 }, 74.4651 },
+				{ { -4.82617, -19.6475, 3.83704 }, 168.125 }
+		};
+		lockerPos[] = {
+				{ { 7.10742, -5.75781, -1.06905 }, 75.8703 }
+		};
+		palletPos[] = {
+				{ { -1.95898, 8.42969, -1.0371 }, 77 },
+				{ { -8.23242, 7.49414, -1.0371 }, 321 },
+				{ { 5.07422, 10.7188, -1.0371 }, 54 }
+		};
+	};
+	class Land_GH_House_2_F : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -3.61914, -1.67969, -0.301109 }, 359.443 },
+				{ { 4.32227, -6.37305, -0.301155 }, 91.0395 }
+		};
+		fridgePos[] = {
+				{ { -4.25, -5.02637, -0.301128 }, 269.6 }
+		};
+		bedPos[] = {
+				{ { -1.24805, -4.87793, -0.301178 }, 5.78317 }
+		};
+		couchPos[] = {
+				{ { 3.05664, -3.25195, -0.301193 }, 2.05902 }
+		};
+		chairPos[] = {
+				{ { 0.806641, -6.70605, -0.301163 }, 224.27 }
+		};
+		filingPos[] = {
+				{ { -2.39844, 1.31348, -0.301136 }, 11.8617 }
+		};
+		pelicanPos[] = {
+				{ { 3.53711, 0.832031, -0.30117 }, 35.7628 }
+		};
+		tablePos[] = {
+				{ { 0.587891, -3.72168, -0.301182 }, 271.612 }
+		};
+		palletPos[] = {
+				{ { 2.25586, -8.62695, -0.307613 }, 54 }
+		};
+	};
+	class Land_GH_House_1_F : Default {
+		lootType = "mil";
+		fridgePos[] = { { { -4.17383, -2.11719, 1.37387 }, 356.381 } };
+		shelfPos[] = { { { 3.35352, 1.26758, -2.12615 }, 0.862475 }, { { -3.58203, -1.93359, -2.12615 }, 1.86357 } };
+		pelicanPos[] = { { { -3.33398, 0.990234, -2.12616 }, 359.887 } };
+		wardrobePos[] = { { { -4.21094, -4.39258, -2.12611 }, 272.378 } };
+		bedPos[] = { { { 1.22852, -4.24219, -2.12617 }, 181.084 }, { { 3.54102, -5.65625, 1.37383 }, 4.33171 } };
+		couchPos[] = { { { -1.45703, -5.70703, -2.12618 }, 0.461905 }, { { -3.22266, -5.88867, 1.37383 }, 180.141 } };
+		chairPos[] = { { { 0.833984, -6.78516, 1.37384 }, 217.232 }, { { -0.796875, -7.06055, 1.38113 }, 139.376 }, { { -3.92773, -6.72656, -2.12617 }, 216.64 }, { { 3.77539, -7.01172, -2.12617 }, 144.922 } };
+		filingPos[] = { { { 2.1543, -1.97656, 1.37398 }, 1.26754 } };
+		tablePos[] = { { { -0.554688, -4.24219, 1.35334 }, 88.7164 }, { { 4.14453, -2.66797, 1.37383 }, 87.9678 }, { { 3.95898, -3.00781, -2.12617 }, 85.0581 } };
+	};
+	class Land_Stadium_p9_F : Default {
+		shelfPos[] = {
+				{ { -0.947266, -2.85645, 0.260189 }, 177.78 },
+				{ { -3.80273, -1.29102, 3.35416 }, 268.338 }
+		};
+		fridgePos[] = {
+				{ { -3.57031, -2.56543, 0.270462 }, 268.533 }
+		};
+		couchPos[] = {
+				{ { -1.27344, 2.86328, 3.35296 }, 92.8845 }
+		};
+		cookerPos[] = {
+				{ { -3.52734, -1.58203, 0.259689 }, 273.789 }
+		};
+		chairPos[] = {
+				{ { 1.47656, 3.48535, 3.35328 }, 160.404 }
+		};
+		filingPos[] = {
+				{ { -0.632813, 1.39258, 3.35342 }, 1.10306 },
+				{ { -1.58594, 1.4082, 3.35342 }, 1.39755 }
+		};
+		pelicanPos[] = {
+				{ { -2.21484, 2.33984, 7.08003 }, 44.2351 }
+		};
+		toolRackPos[] = {
+				{ { 1.81055, -1.7666, 3.35528 }, 89.8722 }
+		};
+	};
+
+
+	//AIA Terrain Pack - Chernarus
+
+	// found missing since 0.3 - start
+	class Land_HouseB_Tenement : Default {
+		lootBiasPos = 20;
+		fridgePos[] = { { { 3.94238, 10.9263, -20.8109 }, 271.337 } };
+		shelfPos[] = { { { -12.9045, 4.21875, -20.8107 }, 269.047 } };
+		pelicanPos[] = { { { -1.07813, 4.78125, -20.8109 }, 174.223 } };
+		couchPos[] = { { { 2.45361, -4.50049, -20.8027 }, 276.133 } };
+		chairPos[] = { { { -3.4668, -1.78467, -20.8109 }, 190.415 } };
+		filingPos[] = { { { -6.31909, 6.52441, -20.8109 }, 1.58827 } };
+		tablePos[] = { { { -0.552246, -1.95654, -20.8043 }, 181.181 } };
+		shoeboxPos[] = { { { -12.9771, -2.2373, -20.8109 }, 269.642 } };
+		freezerPos[] = { { { 5.33984, 14.1865, -20.8109 }, 0.00699411 } };
+	};
+	class Land_HouseV2_01A : Default {
+		lootBiasPos = 20;
+		shelfPos[] = { { { -0.443115, -3.33936, -5.62667 }, 90.1773 } };
+		shoeboxPos[] = { { { -3.60864, -3.06836, -5.62945 }, 278.015 } };
+	};
+	class Land_Barn_W_02 : Default {
+		lootBiasPos = 20;
+		fridgePos[] = { { { 4.66504, 5.95215, -2.34131 }, 73.4104 } };
+		shelfPos[] = { { { 3.54199, -6.98242, -2.3354 }, 180.843 } };
+		pelicanPos[] = { { { -4.61279, 1.50342, 1.06024 }, 189.27 } };
+		couchPos[] = { { { 3.75586, -2.53906, -2.33824 }, 4.32959 } };
+		filingPos[] = { { { -6.13037, 7.19189, -2.33458 }, 357.942 } };
+		toolRackPos[] = { { { -6.70459, -3.06445, -2.33691 }, 271.665 } };
+		shoeboxPos[] = { { { -0.72168, 7.42285, 1.0585 }, 359.807 } };
+		palletPos[] = { { { -5.00684, -5.97168, -2.33609 }, 251.999 } };
+		freezerPos[] = { { { 3.05176, 6.89844, -2.34122 }, 1.42345 } };
+	};
+	class Land_Shed_W01 : Default {
+		chairPos[] = { { { -1.21777, 0.42627, -1.44093 }, 66.292 } };
+		toolRackPos[] = { { { 2.47363, -0.0288086, -1.31694 }, 269.634 } };
+	};
+	class Land_A_Castle_Bergfrit : Default {
+		shelfPos[] = { { { -0.993164, 2.89014, -2.8378 }, 356.361 }, { { 1.80957, 4.87695, 8.74091 }, 358.335 } };
+		pelicanPos[] = { { { -3.03516, -5.00195, 16.8296 }, 195.125 } };
+		couchPos[] = { { { 2.07031, -1.73779, -2.85995 }, 359.982 } };
+		chairPos[] = { { { 1.5791, -0.515137, 8.74054 }, 62.6844 }, { { 1.63281, -2.84229, 8.74051 }, 127.602 } };
+		tablePos[] = { { { -1.22559, -1.10742, 8.74051 }, 270.108 } };
+		shoeboxPos[] = { { { 3.47461, 3.8999, 8.80176 }, 0.625621 } };
+	};
+	class Land_Shed_wooden : Default {
+		shelfPos[] = { { { 0.172363, -1.19922, -1.31766 }, 189.191 } };
+		filingPos[] = { { { -1.19043, 1.20996, -1.31787 }, 273.29 } };
+		shoeboxPos[] = { { { -0.77002, -1.16211, -1.31796 }, 0 } };
+		palletPos[] = { { { 1.97168, 0.560547, -1.31796 }, 357.8 } };
+		freezerPos[] = { { { 1.86523, -0.863281, -1.31796 }, 178.905 } };
+	};
+	class Land_HouseV_1L2 : Default {
+		shelfPos[] = { { { 2.33252, -5.83691, -3.1355 }, 267.989 } };
+		pelicanPos[] = { { { -1.02002, 4.17871, -0.403534 }, 51.0824 } };
+		bedPos[] = { { { 2.53662, 4.57813, -2.63803 }, 95.5054 } };
+		filingPos[] = { { { 3.8457, 2.2998, -2.71201 }, 266.031 } };
+		toolRackPos[] = { { { 4.28125, -3.77441, -2.97101 }, 270.154 } };
+		shoeboxPos[] = { { { 0.916016, 7.69043, -0.397675 }, 175.886 } };
+		palletPos[] = { { { -4.99902, 2.96875, -3.10464 }, 89.0781 } };
+		freezerPos[] = { { { -0.606934, 4.00098, -2.83835 }, 180.414 } };
+	};
+	class Land_HouseV_1I1 : Default {
+		fridgePos[] = { { { -5.5127, 4.8623, -3.33844 }, 117.119 } };
+		chairPos[] = { { { -0.822754, -2.1123, -2.8017 }, 270.285 } };
+		filingPos[] = { { { 0.899414, -1.97656, -2.83847 }, 43.1203 } };
+		shoeboxPos[] = { { { -1.65918, -1.8125, -1.67868 }, 90.0296 } };
+	};
+	class Land_HouseV_1L1 : Default {
+		chairPos[] = { { { 1.31738, 5.71338, -1.08368 }, 88.2198 } };
+		shoeboxPos[] = { { { 0.444336, 6.01123, -1.09442 }, 88.1007 } };
+	};
+	class Land_Misc_deerstand : Default {
+		pelicanPos[] = { { { 0.587158, -0.0830078, 1.05899 }, 94.5891 } };
+		shoeboxPos[] = { { { -0.863037, -0.864258, 1.05893 }, 355.986 } };
+	};
+	class Land_HouseBlock_A1_1 : Default {
+		shelfPos[] = { { { -2.89551, -2.44434, -4.69214 }, 92.8328 } };
+		filingPos[] = { { { -2.95898, -0.0991211, -4.69229 }, 91.0289 } };
+		lockerPos[] = { { { -4.78516, 3.52197, -4.69232 }, 270.002 } };
+		toolRackPos[] = { { { -2.58105, 3.7561, -4.69196 }, 87.6312 } };
+		shoeboxPos[] = { { { 0.851563, 4.91016, -4.69232 }, 160.104 } };
+		freezerPos[] = { { { -3.1582, -3.96021, -4.69235 }, 95.443 } };
+	};
+	class Land_HouseBlock_A3 : Default {
+		fridgePos[] = { { { -1.21191, -4.7356, -5.44312 }, 280.564 } };
+		couchPos[] = { { { -0.486328, 4.81421, -5.44312 }, 179.643 } };
+		chairPos[] = { { { -0.952148, 2.76099, -5.44305 }, 272.88 } };
+		shoeboxPos[] = { { { -7.55859, 4.49561, -5.44315 }, 146.992 } };
+		palletPos[] = { { { 1.05273, 0.597656, -5.44315 }, 358.168 } };
+	};
+	class Land_Hut06 : Default {
+		chairPos[] = { { { 0.739502, 2.50977, -1.58963 }, 76.0347 } };
+		filingPos[] = { { { -0.806396, 2.76465, -1.61115 }, 322.16 } };
+		shoeboxPos[] = { { { -1.33862, -0.988281, -1.82275 }, 312.382 } };
+		freezerPos[] = { { { -0.840332, 1.09473, -1.56879 }, 269.861 } };
+	};
+	// found missing since 0.3 - end
+
+
+
+
+
+	class Land_A_FuelStation_Build : Default {
+		shelfPos[] = {
+				{ { 0.511719, 1.52344, -1.6008 }, 359.975 }
+		};
+		filingPos[] = {
+				{ { 2.38086, 1.33203, -1.60098 }, 29.5354 }
+		};
+		pelicanPos[] = {
+				{ { 0.268066, 1.60449, 1.51318 }, 196.822 }
+		};
+		toolRackPos[] = {
+				{ { -0.960938, 1.66895, -1.60083 }, 356.645 }
+		};
+		shoeboxPos[] = {
+				{ { -1.02441, 0.646484, -1.10098 }, 30.3502 }
+		};
+		palletPos[] = {
+				{ { 2.10742, -3.15625, -1.62936 }, 89.5756 }
+		};
+		freezerPos[] = {
+				{ { 2.17334, -0.862305, -1.57938 }, 89.5756 }
+		};
+	};
+	class Land_HouseV2_04_interier : Default {
+		fridgePos[] = {
+				{ { -0.904297, 6.40674, -5.76508 }, 92.3811 }
+		};
+		couchPos[] = {
+				{ { 6.14941, -5.21924, -5.76886 }, 359.884 }
+		};
+		cookerPos[] = {
+				{ { -5.65234, 2.6543, -5.7666 }, 177.307 }
+		};
+		chairPos[] = {
+				{ { -1.33594, 2.95557, -5.76614 }, 133.309 }
+		};
+		filingPos[] = {
+				{ { 7.92773, -2.18066, -5.76855 }, 179.469 }
+		};
+		pelicanPos[] = {
+				{ { 0.557617, 1.16406, -2.91656 }, 341.166 }
+		};
+		tablePos[] = {
+				{ { -5.5498, 5.12793, -5.7655 }, 271.134 }
+		};
+		lockerPos[] = {
+				{ { 7.20996, 7.07715, -5.76495 }, 358.294 }
+		};
+		toolRackPos[] = {
+				{ { 8.28613, 2.40527, -5.76694 }, 89.7594 }
+		};
+		shoeboxPos[] = {
+				{ { -3.81738, 6.99902, -5.76492 }, 253.823 }
+		};
+		freezerPos[] = {
+				{ { 2.98145, 7.00977, -5.7554 }, 0 }
+		};
+	};
+	class Land_Ind_Garage01 : Default {
+		shelfPos[] = {
+				{ { 0.32373, 3.56543, -1.25259 }, 357.258 }
+		};
+		fridgePos[] = {
+				{ { 2.79932, 1.31641, -1.26077 }, 84.1437 }
+		};
+		filingPos[] = {
+				{ { -1.06348, 3.40234, -1.25024 }, 351.15 }
+		};
+		lockerPos[] = {
+				{ { 2.82422, -0.936523, -1.26364 }, 88.0941 }
+		};
+		toolRackPos[] = {
+				{ { -1.65771, -1.39258, -1.25879 }, 269.614 }
+		};
+		shoeboxPos[] = {
+				{ { 2.21729, 3.40918, -1.25677 }, 63.3963 }
+		};
+		palletPos[] = {
+				{ { -0.140625, -1.52051, -1.2254 }, 0 }
+		};
+	};
+	class Land_HouseV2_02_Interier : Default {
+		shelfPos[] = {
+				{ { 8.94824, 4.19141, -5.55692 }, 90.8873 }
+		};
+		fridgePos[] = {
+				{ { 6.06299, -2.00977, -5.55707 }, 176.197 }
+		};
+		bedPos[] = {
+				{ { -8.03711, -0.987305, -5.5571 }, 184.341 }
+		};
+		couchPos[] = {
+				{ { -6.82422, 5.68262, -5.5571 }, 274.703 }
+		};
+		wardrobePos[] = {
+				{ { -8.83203, 1.7168, -5.56403 }, 268.499 }
+		};
+		cookerPos[] = {
+				{ { 8.7959, -2.00684, -5.5574 }, 179.624 }
+		};
+		chairPos[] = {
+				{ { -2.62451, 0.544922, -5.5571 }, 78.4276 }
+		};
+		filingPos[] = {
+				{ { 2.37305, 5.25586, -5.55707 }, 270.596 }
+		};
+		pelicanPos[] = {
+				{ { 8.31055, 6.2832, -5.5571 }, 37.3391 }
+		};
+		tablePos[] = {
+				{ { -7.00488, 4.00879, -5.5571 }, 181.978 }
+		};
+		lockerPos[] = {
+				{ { -1.11523, 1.34863, -5.5571 }, 269.367 }
+		};
+		shoeboxPos[] = {
+				{ { -5.84717, -2.15918, -3.53598 }, 182.785 }
+		};
+		freezerPos[] = {
+				{ { 4.64551, -2.0127, -5.55713 }, 179.15 }
+		};
+	};
+	class Land_A_Office02 : Default {
+		shelfPos[] = {
+				{ { 2.7417, -4.16504, -8.17639 }, 11.0089 }
+		};
+		fridgePos[] = {
+				{ { -23.2476, 2.61133, -8.16403 }, 25.3519 }
+		};
+		couchPos[] = {
+				{ { 2.26904, 2.13965, 6.55609 }, 272.088 }
+		};
+		chairPos[] = {
+				{ { -0.574707, -6.89551, -8.17813 }, 228.796 },
+				{ { 5.17578, -6.96777, -8.17719 }, 126.645 }
+		};
+		filingPos[] = {
+				{ { 1.70068, -4.21973, -8.17661 }, 317.714 }
+		};
+		pelicanPos[] = {
+				{ { -18.9556, 3.99902, 8.6904 }, 298.688 }
+		};
+		tablePos[] = {
+				{ { 4.73682, -4.41992, -8.17679 }, 358.645 }
+		};
+		toolRackPos[] = {
+				{ { -15.4028, 0.609375, 5.35764 }, 270.07 }
+		};
+		shoeboxPos[] = {
+				{ { -0.565918, -4.23535, -8.17667 }, 351.414 }
+		};
+		palletPos[] = {
+				{ { 9.29248, 2.92676, 5.40057 }, 291 },
+				{ { -21.335, 0.68457, 5.40057 }, 0 }
+		};
+	};
+	class Land_Ind_Workshop01_04 : Default {
+		shelfPos[] = {
+				{ { 1.25928, -3.33008, -1.55847 }, 27.2942 }
+		};
+		filingPos[] = {
+				{ { -2.01318, -3.40186, -1.55435 }, 293.278 }
+		};
+		tablePos[] = {
+				{ { -1.75049, -5.65186, -1.55209 }, 271.116 }
+		};
+		lockerPos[] = {
+				{ { -1.34473, 4.75293, -1.55695 }, 352.726 }
+		};
+		toolRackPos[] = {
+				{ { 0.178711, -7.25537, -1.54968 }, 178.882 }
+		};
+		shoeboxPos[] = {
+				{ { 1.73438, 4.83691, -1.55429 }, 114.503 }
+		};
+		palletPos[] = {
+				{ { -0.102539, -4.55371, -1.51132 }, 340.1 }
+		};
+	};
+	class Land_Ind_TankBig : Default {
+		pelicanPos[] = {
+				{ { -6.74146, 1.74805, 5.31989 }, 321.741 }
+		};
+		palletPos[] = {
+				{ { 0.539551, 2.4082, 5.36334 }, 25 },
+				{ { -3.41626, -2.57813, 5.36334 }, 340.1 }
+		};
+	};
+	class Land_Ind_Workshop01_L : Default {
+		shelfPos[] = {
+				{ { 1.93359, -1.50195, -1.33884 }, 358.185 },
+				{ { -0.805664, 3.11865, -1.33203 }, 181.507 }
+		};
+		filingPos[] = {
+				{ { 0.609375, -3.6875, -1.33704 }, 245.604 }
+		};
+		tablePos[] = {
+				{ { 3.36523, -5.01807, -1.3364 }, 176.426 }
+		};
+		toolRackPos[] = {
+				{ { 3.47949, -5.65381, -1.33569 }, 177.054 }
+		};
+		shoeboxPos[] = {
+				{ { -1.85254, 1.84229, -1.30463 }, 25 }
+		};
+		palletPos[] = {
+				{ { 2.75781, -3.29102, -1.30463 }, 0.999969 },
+				{ { -2.88086, 4.13037, -1.30463 }, 0.999969 }
+		};
+	};
+	class Land_Ind_Vysypka : Default {
+		shelfPos[] = {
+				{ { 2.80957, 4.73828, -5.03592 }, 90.1414 }
+		};
+		fridgePos[] = {
+				{ { -4.28125, 27.6045, -5.04321 }, 357.803 }
+		};
+		couchPos[] = {
+				{ { -3.60791, 1.55469, 4.09406 }, 183 },
+				{ { 4.81934, -11.9766, -5.03705 }, 184.375 }
+		};
+		chairPos[] = {
+				{ { -3.03857, -1.72754, 4.09406 }, 227 },
+				{ { -2.76172, 4.26465, 4.09406 }, 320 }
+		};
+		filingPos[] = {
+				{ { 2.69727, 3.66992, -5.03604 }, 91.3547 }
+		};
+		pelicanPos[] = {
+				{ { -5.5249, 24.7314, -3.82114 }, 358.775 }
+		};
+		tablePos[] = {
+				{ { -0.984375, 1.59863, 4.09406 }, 91 }
+		};
+		lockerPos[] = {
+				{ { -2.91016, 27.584, -5.04214 }, 358.97 }
+		};
+		toolRackPos[] = {
+				{ { 3.44922, 1.56738, -5.03577 }, 270.011 },
+				{ { 3.0127, -14.3877, -5.03381 }, 89.987 }
+		};
+		shoeboxPos[] = {
+				{ { -0.418945, -22.6094, -5.034 }, 17.1755 }
+		};
+		palletPos[] = {
+				{ { 1.82861, -1.39258, -4.99594 }, 0.999969 },
+				{ { 1.86621, -8.21387, -4.99594 }, 0.999969 },
+				{ { -8.73828, 22.5186, -4.99594 }, 0.999969 },
+				{ { -11.6133, -21.9463, 4.08408 }, 183 },
+				{ { -6.85107, 26.7197, -4.99594 }, 0.999969 }
+		};
+	};
+	class Land_Tovarna2 : Default {
+		shelfPos[] = {
+				{ { -0.157715, 8.94824, -5.58551 }, 359.048 },
+				{ { -8.91797, 8.95117, -2.25754 }, 359.07 }
+		};
+		fridgePos[] = {
+				{ { -8.09814, 6.00781, -5.58575 }, 93.4298 },
+				{ { -7.38135, 7.14648, 1.59531 }, 270.644 }
+		};
+		couchPos[] = {
+				{ { -12.2339, 6.00293, 1.64017 }, 192.359 }
+		};
+		cookerPos[] = {
+				{ { -7.34717, 8.47168, 1.59488 }, 269.304 }
+		};
+		chairPos[] = {
+				{ { -11.0293, -1.44824, 0.397858 }, 288.29 },
+				{ { -12.7896, 1.66406, -2.25778 }, 256.194 },
+				{ { -10.5977, 8.53906, -5.58542 }, 7.869 }
+		};
+		filingPos[] = {
+				{ { -3.76074, 4.69141, -2.25797 }, 160.402 },
+				{ { -11.9707, 1.33887, -5.58572 }, 179.608 },
+				{ { -13.021, 1.52051, 1.64032 }, 262.552 },
+				{ { 20.0762, 16.4258, -5.52646 }, 270.183 }
+		};
+		pelicanPos[] = {
+				{ { -2.12793, 7.01758, 6.65506 }, 287.139 }
+		};
+		tablePos[] = {
+				{ { -8.11768, 0.283203, 0.376648 }, 359.581 },
+				{ { -10.749, 8.58691, 1.64017 }, 359.286 },
+				{ { -13.1201, 7.74902, -5.58578 }, 270.804 }
+		};
+		lockerPos[] = {
+				{ { 3.74805, 5.08203, -4.64661 }, 269.986 },
+				{ { -13.2759, 2.98828, -2.25778 }, 271.412 }
+		};
+		toolRackPos[] = {
+				{ { -7.78174, 8.01758, -2.25754 }, 271.288 },
+				{ { -13.6577, 2.94336, -5.5856 }, 269.477 }
+		};
+		shoeboxPos[] = {
+				{ { -3.45215, -9.0332, 3.39786 }, 287.139 }
+		};
+		palletPos[] = {
+				{ { 1.72998, 4.39844, -5.59995 }, 23.0657 },
+				{ { -5.87061, -0.490234, -5.58575 }, 0 },
+				{ { -4.23682, 5.16309, 1.59531 }, 92.0019 },
+				{ { 1.79346, 6.75781, -5.58884 }, 0.010498 },
+				{ { 9.70654, 4.69238, -4.60858 }, 183 },
+				{ { -12.2871, -5.12402, -5.58575 }, 360 },
+				{ { -11.9155, -7.37012, -5.58566 }, 197.387 }
+		};
+		freezerPos[] = {
+				{ { -4.59424, 9.50977, 1.59531 }, 6.14914 }
+		};
+	};
+	class Land_Rail_House_01 : Default {
+		lockerPos[] = {
+				{ { 1.89893, 2.44336, -1.30908 }, 183.214 }
+		};
+		palletPos[] = {
+				{ { -3.34912, 3.12109, -1.27396 }, 183 }
+		};
+	};
+	class Land_Ind_Pec_03b : Default {
+		couchPos[] = {
+				{ { 3.66602, -4.47949, 5.7338 }, 181.985 }
+		};
+		chairPos[] = {
+				{ { 4.10938, -1.91309, 5.73383 }, 287.772 },
+				{ { 3.68262, -6.99902, 5.73383 }, 247.919 }
+		};
+		pelicanPos[] = {
+				{ { 2.27197, -4.89648, 5.73383 }, 89.1607 }
+		};
+		tablePos[] = {
+				{ { 5.6416, -4.37207, 5.7338 }, 88.5916 }
+		};
+		palletPos[] = {
+				{ { -11.5796, -12.9551, 0.982483 }, 183 },
+				{ { -9.27979, -17.4229, 0.982483 }, 142 }
+		};
+	};
+	class Land_HouseV_1I4 : Default {
+		fridgePos[] = {
+				{ { -3.25146, -0.897461, -2.78665 }, 356.197 }
+		};
+		wardrobePos[] = {
+				{ { 0.701172, 3.27832, -2.746 }, 90.7671 }
+		};
+		cookerPos[] = {
+				{ { -4.33691, -3.19141, -2.79153 }, 270.746 }
+		};
+		chairPos[] = {
+				{ { 0.333496, -4.5752, -2.73975 }, 134 }
+		};
+		tablePos[] = {
+				{ { -0.798584, 4.34863, -2.76721 }, 1.79366 }
+		};
+		shoeboxPos[] = {
+				{ { -1.82813, -5.14746, -2.73975 }, 25 }
+		};
+		freezerPos[] = {
+				{ { -1.84106, -1.0957, -2.73975 }, 358 }
+		};
+	};
+	class Land_A_Hospital : Default {
+		shelfPos[] = {
+				{ { -7.99219, -2.12891, -7.36588 }, 268.792 }
+		};
+		couchPos[] = {
+				{ { -14.5566, 1.18652, 3.26541 }, 321.707 }
+		};
+		chairPos[] = {
+				{ { -12.1943, -1.09375, 3.26541 }, 77.6588 }
+		};
+		filingPos[] = {
+				{ { 1.95996, -2.26953, -7.36606 }, 269.665 },
+				{ { -1.56006, -2.83984, -7.36606 }, 1.40195 },
+				{ { -8.75537, -3.58887, -7.36606 }, 359.532 },
+				{ { -9.37305, -3.63574, -7.36606 }, 359.169 }
+		};
+		pelicanPos[] = {
+				{ { 16.3394, 4.99512, 6.13242 }, 0.903763 }
+		};
+		tablePos[] = {
+				{ { 14.708, -4.07813, -7.36609 }, 359.602 }
+		};
+		lockerPos[] = {
+				{ { -1.64258, -1.69824, -7.36609 }, 179.931 }
+		};
+		toolRackPos[] = {
+				{ { 17.5059, -2.06738, 3.26559 }, 0.163071 }
+		};
+		shoeboxPos[] = {
+				{ { -5.40625, 8.16602, -4.35254 }, 39 },
+				{ { 9.88281, -1.00098, -7.36609 }, 39.1093 },
+				{ { -14.6543, -1.31641, -7.36609 }, 321.711 }
+		};
+		palletPos[] = {
+				{ { -8.12305, -5.10742, 3.30746 }, 97 },
+				{ { 17.8105, -5.44922, 3.30746 }, 359 }
+		};
+		freezerPos[] = {
+				{ { 1.0166, -3.02148, -7.34253 }, 359 }
+		};
+	};
+	class Land_Panelak2 : Default {
+		shelfPos[] = {
+				{ { 4.20508, -4.24121, 5.41052 }, 90.3016 }
+		};
+		fridgePos[] = {
+				{ { 7.20898, 3.8916, 5.41022 }, 90.4745 }
+		};
+		couchPos[] = {
+				{ { -4.25195, 0.902344, 8.29941 }, 59.4268 }
+		};
+		wardrobePos[] = {
+				{ { 5.46094, 1.24414, 5.41287 }, 357.96 }
+		};
+		cookerPos[] = {
+				{ { 2.14648, 4.04395, 5.40982 }, 270.01 }
+		};
+		chairPos[] = {
+				{ { 4.50977, -1.51123, 5.44611 }, 241.946 },
+				{ { -6.61035, 0.507813, 8.29944 }, 165.78 }
+		};
+		filingPos[] = {
+				{ { 2.0459, -6.52539, 5.41025 }, 206.326 }
+		};
+		pelicanPos[] = {
+				{ { -0.511719, 4.30273, 4.026 }, 343.273 }
+		};
+		tablePos[] = {
+				{ { 7.14258, -0.713867, 5.41226 }, 89.9922 }
+		};
+		shoeboxPos[] = {
+				{ { -5.76758, 0.51709, 8.29941 }, 287.425 },
+				{ { 5.38281, -6.46777, 5.41022 }, 321.607 }
+		};
+		palletPos[] = {
+				{ { 0.361328, -0.00292969, -4.00781 }, 39 },
+				{ { 6.25586, 3.7168, 8.33218 }, 89 }
+		};
+		freezerPos[] = {
+				{ { 4.01465, -5.87646, 5.49216 }, 89 }
+		};
+	};
+	class Land_Panelak : Default {
+		shelfPos[] = {
+				{ { -1.18066, 4.3916, -1.39282 }, 269.731 }
+		};
+		fridgePos[] = {
+				{ { 2.1582, 3.72363, 0.0450439 }, 274.269 },
+				{ { 2.12598, 3.95801, 2.74509 }, 269.888 }
+		};
+		bedPos[] = {
+				{ { -6.67578, -0.548828, 2.745 }, 187.481 }
+		};
+		couchPos[] = {
+				{ { 6.28906, -0.512695, 0.0449829 }, 2.62378 }
+		};
+		wardrobePos[] = {
+				{ { 4.10449, -4.86182, 2.74927 }, 89.2451 },
+				{ { -7.1543, -1.32568, 0.0450439 }, 273.001 }
+		};
+		cookerPos[] = {
+				{ { 7.18164, 4.19922, 0.0445557 }, 91.3201 },
+				{ { 7.18164, 4.20068, 2.7446 }, 90.3307 }
+		};
+		chairPos[] = {
+				{ { -2.27051, 3.78564, 0.0449829 }, 56.0406 },
+				{ { -4.54883, -0.710938, 0.0562744 }, 107.616 },
+				{ { 4.48535, -1.09717, 2.74661 }, 248.12 }
+		};
+		filingPos[] = {
+				{ { 4.16895, -3.5332, 0.045105 }, 7.97333 },
+				{ { -5.07324, -6.33447, 0.0450745 }, 89.5886 }
+		};
+		pelicanPos[] = {
+				{ { -6.6582, 3.63037, 5.59943 }, 286.286 }
+		};
+		tablePos[] = {
+				{ { -3.85254, -5.19385, 2.745 }, 265.362 },
+				{ { 7.02734, -0.689453, 2.745 }, 87.3795 }
+		};
+		lockerPos[] = {
+				{ { -3.9502, -5.51221, 0.0449829 }, 268.107 }
+		};
+		shoeboxPos[] = {
+				{ { -1.06641, -6.35303, 0.0449829 }, 206.957 },
+				{ { 0.953125, -6.38086, 0.0449829 }, 141.218 }
+		};
+		palletPos[] = {
+				{ { 0.506836, -0.0400391, -1.26782 }, 89 },
+				{ { 4.21777, -3.08203, 5.59943 }, 25.7308 }
+		};
+		freezerPos[] = {
+				{ { 4.03906, -6.05371, 0.0449829 }, 90 }
+		};
+	};
+	class Land_Hangar_2 : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 1.82813, -9.84717, -2.58939 }, 181.468 },
+				{ { 5.83984, 11.563, -2.58939 }, 358.914 },
+				{ { 14.6816, -1.41748, -2.59756 }, 90.4656 },
+				{ { -14.6836, 1.73828, -2.58939 }, 269.242 }
+		};
+		fridgePos[] = {
+				{ { 12.8926, 11.2778, -2.59778 }, 3.16315 }
+		};
+		couchPos[] = {
+				{ { 1.27832, 10.2305, -2.58963 }, 273.856 }
+		};
+		filingPos[] = {
+				{ { 0.000976563, 1.3833, -2.58926 }, 178.715 }
+		};
+		lockerPos[] = {
+				{ { 13.5166, -9.63965, -2.58951 }, 180.232 }
+		};
+		toolRackPos[] = {
+				{ { -2.47949, -10.0684, -2.58923 }, 178.901 },
+				{ { 9.01953, 11.8232, -2.58923 }, 358.139 }
+		};
+		shoeboxPos[] = {
+				{ { -14.5693, -9.69629, -2.58963 }, 188.48 }
+		};
+		palletPos[] = {
+				{ { -0.00195313, -3.16113, -2.5896 }, 332.619 },
+				{ { -8.97656, -4.14404, -2.5896 }, 219.281 },
+				{ { -12.9932, 10.293, -2.5896 }, 336.817 }
+		};
+		freezerPos[] = {
+				{ { 14.3408, 9.39063, -2.58963 }, 81.9611 }
+		};
+	};
+	class Land_rail_station_big : Default {
+		shelfPos[] = {
+				{ { 2.86035, 3.51221, -5.07059 }, 91.0331 }
+		};
+		fridgePos[] = {
+				{ { 1.66113, -3.65674, -5.0708 }, 180.242 }
+		};
+		couchPos[] = {
+				{ { -4.33789, -1.56836, -5.07083 }, 178.466 },
+				{ { 8.63281, -5.13672, -5.07083 }, 1.14301 }
+		};
+		chairPos[] = {
+				{ { 1.39844, -4.95752, -5.06161 }, 4.04407 },
+				{ { -4.96484, 3.36914, -5.0708 }, 284.166 },
+				{ { -5.01758, -5.00488, -5.05701 }, 344.841 }
+		};
+		filingPos[] = {
+				{ { 2.84961, 4.2832, -5.07077 }, 90.7927 }
+		};
+		pelicanPos[] = {
+				{ { 0.118164, -6.59912, -2.26553 }, 21.0414 }
+		};
+		tablePos[] = {
+				{ { -5.13184, 1.67041, -5.07083 }, 269.411 }
+		};
+		lockerPos[] = {
+				{ { -1.66504, -3.61719, -5.07083 }, 180.479 },
+				{ { 1.15527, 5.48584, -5.07065 }, 0.986969 },
+				{ { -4.39453, 5.34619, -5.07083 }, 338.371 }
+		};
+		toolRackPos[] = {
+				{ { 6.75195, 1.96289, -5.07056 }, 271.441 }
+		};
+		shoeboxPos[] = {
+				{ { 2.43164, 0.785156, -4.09598 }, 58.5217 }
+		};
+		palletPos[] = {
+				{ { 8.96973, -2.03564, -5.02618 }, 89 },
+				{ { -9.24902, -4.90479, -5.02618 }, 89 }
+		};
+	};
+	class Land_a_stationhouse : Default {
+		shelfPos[] = {
+				{ { 0.554688, -4.70068, -9.49698 }, 272.253 },
+				{ { -0.933594, -5.35596, 4.38522 }, 27.5935 },
+				{ { -0.828125, -8.54004, -4.67508 }, 110.344 },
+				{ { -1.08203, -8.95215, 4.38516 }, 185.611 },
+				{ { -4.35742, -8.55957, -0.0701904 }, 262.53 },
+				{ { 6.66406, -7.6665, -9.49698 }, 181.745 },
+				{ { 19.1992, -5.00488, -9.49698 }, 88.5043 }
+		};
+		fridgePos[] = {
+				{ { 0.871094, -1.74805, -9.49701 }, 271.521 }
+		};
+		couchPos[] = {
+				{ { -10.5244, -3.94482, -0.534332 }, 278.335 }
+		};
+		chairPos[] = {
+				{ { -13.3145, -3.94189, -0.534332 }, 334.169 }
+		};
+		filingPos[] = {
+				{ { -0.786133, -5.13428, -9.49701 }, 78.6633 }
+		};
+		pelicanPos[] = {
+				{ { 7.44531, 0.904785, -4.53433 }, 267.207 },
+				{ { -4.35156, -7.64502, 8.46567 }, 268.821 }
+		};
+		lockerPos[] = {
+				{ { 0.811523, -3.3418, -9.49701 }, 269.999 },
+				{ { 19.0117, 0.613281, -9.49701 }, 90.2636 },
+				{ { 18.9961, -1.15674, -9.49704 }, 89.3234 },
+				{ { 19.042, -2.92578, -9.49701 }, 87.0602 }
+		};
+		toolRackPos[] = {
+				{ { 7.23633, 3.84961, -9.4968 }, 359.928 }
+		};
+		shoeboxPos[] = {
+				{ { -0.743164, -8.33643, -0.0702209 }, 110.554 },
+				{ { 19.1836, -7.53174, -9.49701 }, 105.566 }
+		};
+		palletPos[] = {
+				{ { 2.03418, 1.7915, -9.49701 }, 289.015 },
+				{ { -6.86621, 3.25586, -0.534332 }, 3.05176e-005 },
+				{ { 13.0293, -7.22363, -9.49701 }, 90 },
+				{ { 15.542, -4.56006, -4.53433 }, 6.10352e-005 }
+		};
+		freezerPos[] = {
+				{ { -4.1084, -7.31787, -9.49701 }, 268.732 }
+		};
+	};
+	class Land_Mil_ControlTower : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 1.49512, 2.23389, -1.09485 }, 268.083 },
+				{ { 9.91406, 3.22559, -9.65488 }, 179.209 }
+		};
+		fridgePos[] = {
+				{ { 9.67773, 5.91309, -9.65503 }, 83.7454 }
+		};
+		chairPos[] = {
+				{ { 7.19434, -1.62061, 2.92548 }, 262.144 }
+		};
+		filingPos[] = {
+				{ { 2.79297, 3.34814, -5.49008 }, 190.556 },
+				{ { 3.33105, 3.35791, -5.49011 }, 179.417 },
+				{ { 6.89648, -1.87549, -1.09503 }, 164.743 },
+				{ { 7.53906, -1.85107, -1.09503 }, 187.813 }
+		};
+		pelicanPos[] = {
+				{ { 2.01172, -0.6875, -1.09509 }, 259.256 },
+				{ { 9.32324, 4.80322, -1.08328 }, 41.8877 }
+		};
+		lockerPos[] = {
+				{ { 0.925781, 3.41895, -9.65512 }, 179.859 },
+				{ { 3.19531, 6.13867, -5.49011 }, 0.121094 }
+		};
+		toolRackPos[] = {
+				{ { 4.21484, 6.64795, -9.65485 }, 359.603 }
+		};
+		shoeboxPos[] = {
+				{ { 4.94727, -1.88574, -1.09512 }, 209.993 },
+				{ { 7.66211, 3.85303, -9.65512 }, 210.035 }
+		};
+	};
+	class Land_Ind_Workshop01_01 : Default {
+		shelfPos[] = {
+				{ { 0.558594, -3.47168, -1.3168 }, 176.163 }
+		};
+		fridgePos[] = {
+				{ { 2.12305, -1.83838, -1.315 }, 94.8075 }
+		};
+		filingPos[] = {
+				{ { -1.55566, -0.881836, -1.3237 }, 271.559 }
+		};
+		lockerPos[] = {
+				{ { 2.12109, -0.290039, -1.31625 }, 91.8605 }
+		};
+		toolRackPos[] = {
+				{ { 2.43848, 1.60547, -1.31653 }, 88.4978 }
+		};
+		shoeboxPos[] = {
+				{ { -0.980469, 0.617188, -1.32367 }, 120.007 }
+		};
+		palletPos[] = {
+				{ { 0.501953, 0.146484, -1.27002 }, 251 }
+		};
+	};
+	class Land_Shed_Ind02 : Default {
+		shelfPos[] = {
+				{ { -0.442383, -9.33105, -4.6525 }, 178.364 }
+		};
+		fridgePos[] = {
+				{ { 4.19824, 10.2627, -4.6524 }, 7.46219 }
+		};
+		bedPos[] = {
+				{ { 2.77148, 2.2832, -4.6524 }, 95.4069 }
+		};
+		couchPos[] = {
+				{ { 2.4043, 0.115234, -4.6524 }, 273.531 }
+		};
+		cookerPos[] = {
+				{ { 4.43066, 7.04907, -4.6528 }, 90.7597 }
+		};
+		chairPos[] = {
+				{ { 3.70605, -2.95898, -4.65234 }, 134.818 }
+		};
+		filingPos[] = {
+				{ { -4.03906, 12.3835, -1.30353 }, 1.1741 }
+		};
+		pelicanPos[] = {
+				{ { -0.52832, 11.6462, -1.30273 }, 62.2241 }
+		};
+		tablePos[] = {
+				{ { 1.62695, -1.86011, -4.6524 }, 178.21 }
+		};
+		lockerPos[] = {
+				{ { -2.87305, 12.2234, -1.30331 }, 2.62267 }
+		};
+		toolRackPos[] = {
+				{ { 2.56738, 10.6775, -4.65201 }, 0.953941 }
+		};
+		shoeboxPos[] = {
+				{ { -1.70313, -9.60083, -4.65237 }, 10.367 }
+		};
+		palletPos[] = {
+				{ { -0.0283203, 5.60742, -4.6524 }, 317.381 },
+				{ { 4.12695, -5.37256, -4.64841 }, 251 }
+		};
+	};
+	class Land_Mil_Barracks_i : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 0.902344, -2.39502, -1.12442 }, 238.007 },
+				{ { 0.548828, 2.81006, -1.12292 }, 1.08441 },
+				{ { -9.37012, 2.427, -1.12442 }, 268.201 }
+		};
+		filingPos[] = {
+				{ { 6.55566, 2.81738, -1.12451 }, 92.1856 }
+		};
+		pelicanPos[] = {
+				{ { -2.24316, -1.36963, -1.12463 }, 263.64 },
+				{ { 4.36621, -1.57617, -0.137878 }, 258.932 },
+				{ { -5.0752, -2.4939, -1.12466 }, 183.946 },
+				{ { -8.75977, -0.048584, -1.12463 }, 327.595 }
+		};
+		lockerPos[] = {
+				{ { -5.60156, -1.32373, -1.12164 }, 269.99 }
+		};
+		shoeboxPos[] = {
+				{ { -0.0751953, -1.9834, -0.699829 }, 251 },
+				{ { -3.13574, 0.466797, 0.760162 }, 360 },
+				{ { 3.02734, -1.34937, 0.0901794 }, 251 },
+				{ { 3.02441, -1.40112, -1.12466 }, 302.031 },
+				{ { -6.55371, -2.66333, -1.09982 }, 360 }
+		};
+	};
+	class Land_Ss_hangar : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 14.2314, -8.46289, -5.89868 }, 173.905 },
+				{ { 14.3447, 19.8936, -5.8782 }, 357.293 },
+				{ { 13.5869, -20.874, -5.89853 }, 180.148 }
+		};
+		fridgePos[] = {
+				{ { 14.1895, 12.5049, -5.96277 }, 147.17 }
+		};
+		bedPos[] = {
+				{ { 14.1436, 21.6729, -5.89896 }, 7.74948 }
+		};
+		couchPos[] = {
+				{ { 9.54492, 21.0107, -5.89896 }, 181.695 }
+		};
+		cookerPos[] = {
+				{ { 14.5332, 15.3887, -5.92621 }, 89.5229 }
+		};
+		chairPos[] = {
+				{ { 12.9268, 19.6445, -5.89896 }, 132.385 }
+		};
+		filingPos[] = {
+				{ { 14.2803, -9.54688, -5.89865 }, 14.8827 },
+				{ { -14.4658, -20.7148, -5.89896 }, 255.86 }
+		};
+		pelicanPos[] = {
+				{ { -14.79, 12.7246, -5.89896 }, 273.845 },
+				{ { 14.877, -20.2988, -5.89896 }, 89.7007 }
+		};
+		tablePos[] = {
+				{ { 11.3418, 23.2334, -5.89896 }, 0.0492706 }
+		};
+		lockerPos[] = {
+				{ { 6.78076, 23.332, -5.89896 }, 0.014267 }
+		};
+		toolRackPos[] = {
+				{ { 2.44531, 23.7109, -5.89871 }, 0.000427246 }
+		};
+		shoeboxPos[] = {
+				{ { -14.8848, 23.3271, -5.89899 }, 279.193 }
+		};
+		palletPos[] = {
+				{ { -8.33643, -6.57617, -5.89896 }, 135.758 },
+				{ { 6.4082, 10.2188, -5.89896 }, 115.033 },
+				{ { 14.167, 10.5293, -5.89896 }, 0 },
+				{ { -13.6675, 22.0801, -5.89896 }, 325.134 }
+		};
+	};
+	class Land_Ind_IlluminantTower : Default {
+		pelicanPos[] = {
+				{ { -0.791992, -0.320313, 10.3249 }, 212.76 }
+		};
+		palletPos[] = {
+				{ { -0.012207, 0.503906, -9.74222 }, 360 }
+		};
+	};
+	class Land_Vez : Default {
+		shoeboxPos[] = {
+				{ { 0.286133, 0.867188, 1.36356 }, 1.99998 }
+		};
+	};
+	class Land_Hlidac_budka : Default {
+		shelfPos[] = {
+				{ { 0.84668, 0.125977, -0.809265 }, 172.613 }
+		};
+		filingPos[] = {
+				{ { 0.0673828, 2.31836, -0.809967 }, 357.487 },
+				{ { 2.76709, 2.25684, -0.809967 }, 4.6116 }
+		};
+		tablePos[] = {
+				{ { -2.32129, 1.41016, -0.810028 }, 279.786 }
+		};
+		toolRackPos[] = {
+				{ { 3.00098, 0.819336, -0.809723 }, 90.0308 }
+		};
+		shoeboxPos[] = {
+				{ { 1.62109, 2.32031, -0.810089 }, 349.179 }
+		};
+	};
+	class Land_A_GeneralStore_01a : Default {
+		shelfPos[] = {
+				{ { -2.68848, -5.06152, -1.22772 }, 178.035 },
+				{ { -8.81641, 1.98926, -1.22772 }, 270.025 },
+				{ { 9.85986, 1.51953, -1.22772 }, 179.333 },
+				{ { -5.52783, 9.13184, -1.22772 }, 1.40732 }
+		};
+		fridgePos[] = {
+				{ { -5.12402, -0.585938, -1.20645 }, 180 },
+				{ { -6.17725, -0.568359, -1.20645 }, 180 },
+				{ { 5.68555, -4.8418, -1.20645 }, 188 }
+		};
+		cookerPos[] = {
+				{ { 7.99414, -0.348633, -1.22836 }, 89.6747 }
+		};
+		filingPos[] = {
+				{ { 9.23584, -1.02637, -1.22794 }, 275.138 },
+				{ { 8.26953, -4.60156, -1.22794 }, 92.7205 }
+		};
+		pelicanPos[] = {
+				{ { 14.3594, 0.392578, -1.22797 }, 91.3482 }
+		};
+		tablePos[] = {
+				{ { 14.1704, -3.93848, -1.22797 }, 103.927 }
+		};
+		lockerPos[] = {
+				{ { -7.1709, -1.75488, -1.20645 }, 1.99997 }
+		};
+		toolRackPos[] = {
+				{ { 6.76514, -1.17969, -1.22775 }, 0.137085 }
+		};
+		shoeboxPos[] = {
+				{ { -4.61426, 4.45996, -1.228 }, 198.535 },
+				{ { 12.5962, 0.78418, -1.228 }, 358.034 },
+				{ { 12.3564, 3.56055, -0.333557 }, 265.418 }
+		};
+		palletPos[] = {
+				{ { 9.37988, 5.22168, -1.22797 }, 156.892 }
+		};
+		freezerPos[] = {
+				{ { -1.78906, 0.975586, -1.228 }, 180.781 }
+		};
+	};
+	class Land_A_GeneralStore_01 : Default {
+		shelfPos[] = {
+				{ { 6.03076, -5.04883, -1.22775 }, 359.162 },
+				{ { 12.563, 5.48535, -1.22772 }, 0.496719 },
+				{ { 13.5098, -7.02734, -1.22775 }, 88.7328 }
+		};
+		fridgePos[] = {
+				{ { -9.77979, -1.54102, -1.22797 }, 269.735 },
+				{ { -9.80176, -4.10938, -1.22797 }, 269.049 },
+				{ { 8.25488, -9.58496, -1.23712 }, 0.202225 }
+		};
+		couchPos[] = {
+				{ { -0.619629, -10.2949, -1.23727 }, 270.908 }
+		};
+		cookerPos[] = {
+				{ { 6.6792, -4.2373, -1.22833 }, 91.644 }
+		};
+		chairPos[] = {
+				{ { -2.94092, -9.89746, -1.23715 }, 42.039 },
+				{ { -8.54541, 6.65723, -1.23798 }, 184.459 }
+		};
+		filingPos[] = {
+				{ { 13.4648, -2.1582, -1.22797 }, 116.635 }
+		};
+		tablePos[] = {
+				{ { -9.23096, 5.1709, -1.228 }, 357.131 }
+		};
+		lockerPos[] = {
+				{ { 7.98193, -5.34863, -1.22797 }, 269.751 }
+		};
+		toolRackPos[] = {
+				{ { -10.2231, -7.87598, -1.22778 }, 270.182 }
+		};
+		shoeboxPos[] = {
+				{ { 4.97266, -2.66504, -0.632507 }, 196.652 },
+				{ { -9.93848, -6.0293, -1.228 }, 2.52339 }
+		};
+		palletPos[] = {
+				{ { -5.75488, -1.74707, -1.228 }, 46.8216 },
+				{ { -1.28467, 7.54199, -1.23798 }, 116.447 },
+				{ { 4.78857, 8.44824, -1.23798 }, 46.8216 }
+		};
+		freezerPos[] = {
+				{ { 8.1875, -1.88184, -1.228 }, 186.869 },
+				{ { -7.14307, 5.19238, -1.228 }, 0.0339508 }
+		};
+	};
+	class Land_A_Office01 : Default {
+		shelfPos[] = {
+				{ { 4.0542, 2.89648, -4.55417 }, 178.477 },
+				{ { -1.47119, 4.86328, 0.44574 }, 0.16153 },
+				{ { 10.6284, -3.50879, -2.0542 }, 92.1019 },
+				{ { 11.4131, -4.21387, -4.5542 }, 270.255 }
+		};
+		fridgePos[] = {
+				{ { -0.907715, 5.50684, -2.05441 }, 272.559 }
+		};
+		couchPos[] = {
+				{ { -7.35449, 4.9502, -4.55441 }, 189.687 }
+		};
+		cookerPos[] = {
+				{ { -8.35449, 3.17188, -2.05484 }, 179.618 }
+		};
+		chairPos[] = {
+				{ { -3.1958, -1.04395, -4.55441 }, 275.876 }
+		};
+		filingPos[] = {
+				{ { -3.70752, 2.91797, 0.445648 }, 267.617 },
+				{ { -0.938965, 7.15039, -4.55438 }, 269.465 },
+				{ { -11.54, 0.328125, -2.05438 }, 178.477 },
+				{ { 10.4766, 6.3418, -4.55438 }, 91.5126 },
+				{ { 15.6548, -1.88672, -4.55438 }, 83.8964 }
+		};
+		pelicanPos[] = {
+				{ { 2.34717, -1.2002, 6.6937 }, 215.707 }
+		};
+		tablePos[] = {
+				{ { -3.19434, -2.84082, -4.55441 }, 267.186 },
+				{ { 5.38916, 5.21582, -2.05441 }, 89.0818 },
+				{ { 14.9824, 3.75391, -2.05441 }, 181.475 }
+		};
+		lockerPos[] = {
+				{ { -4.50391, 3.8418, -4.55444 }, 91.5282 }
+		};
+		toolRackPos[] = {
+				{ { 1.45557, 2.6543, -2.05414 }, 178.943 }
+		};
+		shoeboxPos[] = {
+				{ { 5.2876, -1.88086, -4.53268 }, 180 },
+				{ { 3.9707, -5.65234, -2.03021 }, 305.956 },
+				{ { 11.6138, 6.93262, -2.05444 }, 292.589 }
+		};
+		palletPos[] = {
+				{ { -0.465332, 4.62598, -4.53268 }, 180 },
+				{ { 14.7656, -3.98438, -4.53268 }, 176 }
+		};
+	};
+	class Land_A_Pub_01 : Default {
+		shelfPos[] = {
+				{ { 3.4812, 3.81738, -5.77243 }, 179.635 },
+				{ { 2.49878, -4.70703, -1.78745 }, 180.503 },
+				{ { -7.37402, 1.03516, -1.78741 }, 328.603 }
+		};
+		fridgePos[] = {
+				{ { -3.38892, -4.0459, -5.76288 }, 90 }
+		};
+		bedPos[] = {
+				{ { -4.10889, -0.928711, -1.78766 }, 8.16519 },
+				{ { 3.50171, 2.69043, -1.78766 }, 94.542 }
+		};
+		couchPos[] = {
+				{ { 0.366699, 5.18652, -1.78766 }, 189.226 },
+				{ { -5.29517, 0.333984, -5.77267 }, 4.48228 }
+		};
+		wardrobePos[] = {
+				{ { 4.41479, 7.93848, -1.78766 }, 2.10324 }
+		};
+		cookerPos[] = {
+				{ { -3.19385, -1.17285, -5.77298 }, 270.462 }
+		};
+		chairPos[] = {
+				{ { 2.81494, 0.904297, -5.76315 }, 61.6914 },
+				{ { -7.10742, -5.75293, -1.78763 }, 350.056 }
+		};
+		filingPos[] = {
+				{ { -0.388672, -1.30469, -5.77255 }, 181.526 },
+				{ { 4.94727, 1.18066, -1.78763 }, 42.2747 }
+		};
+		pelicanPos[] = {
+				{ { 4.75977, 5.4834, -1.78757 }, 88.4654 },
+				{ { 6.82202, -3.58984, -5.77252 }, 82.7529 }
+		};
+		tablePos[] = {
+				{ { -7.43481, -3.70313, -1.78766 }, 273.649 },
+				{ { -7.42432, -4.37988, -5.77267 }, 273.367 }
+		};
+		toolRackPos[] = {
+				{ { -0.121582, -5.72266, -5.7724 }, 89.0577 }
+		};
+		shoeboxPos[] = {
+				{ { -0.593018, 2.19141, -5.77267 }, 255.46 },
+				{ { 4.81665, -2.40332, -1.78766 }, 105.138 },
+				{ { 1.49854, -6.26074, -5.01288 }, 176 },
+				{ { 0.81665, 7.73535, -5.76288 }, 176 },
+				{ { 6.86157, 5.98242, -1.76291 }, 268 }
+		};
+		freezerPos[] = {
+				{ { 0.44751, -0.999023, -1.76291 }, 268 }
+		};
+	};
+	class Land_Barn_W_01 : Default {
+		shelfPos[] = {
+				{ { 6.55371, -5.96289, -2.65878 }, 88.4798 },
+				{ { -6.56934, -7.53516, -2.65848 }, 265.499 },
+				{ { -4.05664, 19.6274, -2.66928 }, 0.626892 }
+		};
+		couchPos[] = {
+				{ { 5.04297, -17.8179, -2.65564 }, 48.1573 }
+		};
+		chairPos[] = {
+				{ { 6.02441, -15.0469, -2.65741 }, 70.2096 },
+				{ { 2.39844, -19.5591, -2.65125 }, 198.607 }
+		};
+		tablePos[] = {
+				{ { 3.33691, -16.0547, -2.65723 }, 300.369 }
+		};
+		toolRackPos[] = {
+				{ { -6.80078, 2.37402, -2.66006 }, 268.437 },
+				{ { 6.71387, 4.32422, -2.66052 }, 89.4215 },
+				{ { -4.89844, -19.9951, -2.65622 }, 180.001 }
+		};
+		shoeboxPos[] = {
+				{ { -6.57617, 7.07227, -2.66177 }, 270.003 }
+		};
+		palletPos[] = {
+				{ { 5.37402, -1.0918, -2.65982 }, 38.9155 },
+				{ { -5.49023, -1.3877, -2.65979 }, 41.5356 },
+				{ { -0.188477, -10.4546, -2.54559 }, 268 },
+				{ { -2.21777, 10.7607, -2.54559 }, 268 },
+				{ { 2.44043, 14.0454, -2.54559 }, 268 }
+		};
+		freezerPos[] = {
+				{ { -6.14258, 14.0566, -2.66248 }, 279.4 }
+		};
+	};
+	class Land_Church_03 : Default {
+		shelfPos[] = {
+				{ { -8.8877, 4.48535, -14.3282 }, 270.093 }
+		};
+		wardrobePos[] = {
+				{ { -3.39648, 7.18848, -14.3284 }, 272.593 }
+		};
+		chairPos[] = {
+				{ { -8.41113, -4.42725, -14.3249 }, 235.951 }
+		};
+		filingPos[] = {
+				{ { 6.03809, 8.75146, -14.2857 }, 360 },
+				{ { 6.50879, 8.73438, -14.2857 }, 1.99998 }
+		};
+		pelicanPos[] = {
+				{ { 4.48633, -8.17334, -14.3448 }, 228.938 }
+		};
+		tablePos[] = {
+				{ { -8.48633, -0.0556641, -14.3285 }, 269.825 }
+		};
+		shoeboxPos[] = {
+				{ { 6.20117, 1.86279, -14.3285 }, 261.718 }
+		};
+	};
+	class Land_Stodola_open : Default {
+		shelfPos[] = {
+				{ { 2.08691, -3.11719, -4.14142 }, 89.2592 }
+		};
+		fridgePos[] = {
+				{ { 1.88672, -6.01465, -4.13022 }, 119.538 }
+		};
+		couchPos[] = {
+				{ { -3.63477, -0.27832, -4.13867 }, 190.732 }
+		};
+		pelicanPos[] = {
+				{ { -1.71191, -5.06445, 0.297943 }, 33.203 }
+		};
+		tablePos[] = {
+				{ { -3.62988, 6.59473, -4.12482 }, 357.638 }
+		};
+		toolRackPos[] = {
+				{ { -0.110352, 7.2373, -3.92352 }, 357.513 }
+		};
+		shoeboxPos[] = {
+				{ { -4.24609, -6.01855, -3.31656 }, 2.337 }
+		};
+		palletPos[] = {
+				{ { 1.70313, 3.39844, -4.13934 }, 92.8326 },
+				{ { -3.4292, 3.97168, -4.14297 }, 255.32 }
+		};
+	};
+	class Land_Stodola_old_open : Default {
+		shelfPos[] = {
+				{ { 4.81055, 2.76367, -5.10672 }, 359.199 },
+				{ { -3.54004, 7.10645, -1.01971 }, 270.013 }
+		};
+		bedPos[] = {
+				{ { -2.76221, 8.75, -5.10696 }, 179.304 }
+		};
+		couchPos[] = {
+				{ { 0.687988, -8.36133, -5.10693 }, 4.33433 }
+		};
+		chairPos[] = {
+				{ { -1.06543, -5.28418, -1.04489 }, 195.228 }
+		};
+		filingPos[] = {
+				{ { 5.25879, -8.9043, -5.1069 }, 9.83752 }
+		};
+		pelicanPos[] = {
+				{ { -1.12891, -10.7266, 2.93811 }, 260.837 }
+		};
+		tablePos[] = {
+				{ { -2.98828, 10.5127, -1.0199 }, 281.235 }
+		};
+		toolRackPos[] = {
+				{ { 5.54102, 6.0625, -5.10693 }, 90.7204 },
+				{ { -1.74854, -11.5723, -5.10672 }, 178.397 }
+		};
+		shoeboxPos[] = {
+				{ { 5.70313, 11.2139, 0.829681 }, 91 }
+		};
+		palletPos[] = {
+				{ { 2.06445, 3.54102, -5.10696 }, 337.003 },
+				{ { -2.80469, 5.27539, -1.01993 }, 264.974 },
+				{ { 2.44336, 10.6328, 2.93802 }, 86.0969 },
+				{ { 4.44727, 10.3809, -5.07031 }, 91 }
+		};
+		freezerPos[] = {
+				{ { -3.05957, -4.44141, -5.10699 }, 359.822 }
+		};
+	};
+	class Land_Farm_Cowshed_a : Default {
+		shelfPos[] = {
+				{ { 9.65234, 4.23633, -3.1138 }, 359.83 }
+		};
+		bedPos[] = {
+				{ { 7.85938, -5.01221, -3.11404 }, 274.383 }
+		};
+		wardrobePos[] = {
+				{ { 10.21, -4.81543, -3.11401 }, 71.5758 }
+		};
+		filingPos[] = {
+				{ { -2.51855, -6.98389, -3.08054 }, 202 }
+		};
+		tablePos[] = {
+				{ { 0.553711, -6.81055, -3.11404 }, 179.415 }
+		};
+		toolRackPos[] = {
+				{ { 3.21973, -3.75781, -3.11374 }, 182.361 }
+		};
+		shoeboxPos[] = {
+				{ { -3.04883, -0.802246, -3.11407 }, 198.771 }
+		};
+		palletPos[] = {
+				{ { -2.58008, 3.32617, -3.11404 }, 6.63828 },
+				{ { -6.11816, -2.65527, -3.11404 }, 314.264 }
+		};
+		freezerPos[] = {
+				{ { 10.4248, -2.19531, -3.11407 }, 92.3688 }
+		};
+	};
+	class Land_Farm_Cowshed_b : Default {
+		shelfPos[] = {
+				{ { -9.05615, -3.76465, -3.12729 }, 178.703 }
+		};
+		couchPos[] = {
+				{ { -9.82813, 2.41113, -3.1275 }, 296.045 }
+		};
+		chairPos[] = {
+				{ { 8.15674, -3.28955, -3.1275 }, 182.81 }
+		};
+		toolRackPos[] = {
+				{ { 3.27734, -3.9209, -3.12726 }, 178.757 }
+		};
+		palletPos[] = {
+				{ { -4.85059, -2.72949, -3.12753 }, 360 },
+				{ { 5.6416, -2.8208, -3.12753 }, 0.0419464 }
+		};
+	};
+	class Land_Farm_Cowshed_c : Default {
+		shelfPos[] = {
+				{ { -2.79785, 3.14209, -3.15216 }, 295.449 }
+		};
+		filingPos[] = {
+				{ { -2.97998, -4.05469, -3.15222 }, 179.983 }
+		};
+		palletPos[] = {
+				{ { 2.44531, 2.57227, -3.14557 }, 0.00511169 }
+		};
+	};
+	class Land_Barn_Metal : Default {
+		shelfPos[] = {
+				{ { 0.265625, -14.7627, -5.46832 }, 89.0258 }
+		};
+		fridgePos[] = {
+				{ { 10.749, -15.9448, -5.46857 }, 39.8741 }
+		};
+		bedPos[] = {
+				{ { -9.12109, 21.7661, -5.46857 }, 175.889 }
+		};
+		couchPos[] = {
+				{ { 0.614746, -24.1509, -5.46857 }, 96.7518 }
+		};
+		cookerPos[] = {
+				{ { 10.8979, -14.4751, -5.46899 }, 90.5039 }
+		};
+		chairPos[] = {
+				{ { 10.4609, 17.1284, -5.46857 }, 39.8881 }
+		};
+		filingPos[] = {
+				{ { -9.91895, -9.12695, -5.46854 }, 268.516 }
+		};
+		pelicanPos[] = {
+				{ { 10.5273, 22.7065, 5.56143 }, 37.3798 }
+		};
+		tablePos[] = {
+				{ { 10.7944, 13.7324, -5.46857 }, 87.4971 }
+		};
+		lockerPos[] = {
+				{ { 11.0068, 6.05371, -5.46857 }, 89.1326 }
+		};
+		toolRackPos[] = {
+				{ { -10.2012, 9.57422, -5.46823 }, 272.816 }
+		};
+		shoeboxPos[] = {
+				{ { -9.87012, -20.9902, -5.4425 }, 298.809 }
+		};
+		palletPos[] = {
+				{ { 3.24316, 2.2959, -5.44501 }, 135.351 },
+				{ { -4.39014, -9.8374, -5.44501 }, 141.049 },
+				{ { 4.74316, 12.5591, -5.44501 }, 199 },
+				{ { -8.59766, 15.3638, -5.44501 }, 335.835 }
+		};
+		freezerPos[] = {
+				{ { 10.8242, -10.4863, -5.44501 }, 87.998 }
+		};
+	};
+	class Land_A_TVTower_Base : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -0.569824, -2.05078, -2.25824 }, 91.0596 }
+		};
+		filingPos[] = {
+				{ { -2.6377, -4.59082, -2.26028 }, 188.775 }
+		};
+		pelicanPos[] = {
+				{ { 4.38086, -0.0766602, 1.15182 }, 274.734 }
+		};
+		toolRackPos[] = {
+				{ { -0.304199, 2.37402, -2.25836 }, 85.8353 }
+		};
+		shoeboxPos[] = {
+				{ { -1.10449, 6.4873, -2.25751 }, 134.996 }
+		};
+	};
+	class Land_Mil_Guardhouse : Default {
+		shelfPos[] = {
+				{ { -2.07324, 2.70703, -1.69559 }, 89.8322 }
+		};
+		pelicanPos[] = {
+				{ { -2.46729, 0.0585938, -1.69565 }, 113.271 }
+		};
+		shoeboxPos[] = {
+				{ { 5.75781, 1.48828, -1.99878 }, 202 }
+		};
+	};
+	class Land_Mil_House : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 12.2617, 0.839355, -5.67438 }, 181.266 }
+		};
+		fridgePos[] = {
+				{ { 15.2783, 4.78369, -5.67462 }, 52.5764 }
+		};
+		pelicanPos[] = {
+				{ { -1.56152, 6.73584, -1.87274 }, 179.391 }
+		};
+		lockerPos[] = {
+				{ { 10.856, 4.21094, -5.67462 }, 268.532 }
+		};
+		toolRackPos[] = {
+				{ { 10.5098, 1.98584, -5.67413 }, 270.325 }
+		};
+		shoeboxPos[] = {
+				{ { 5.61719, 6.21191, -5.24442 }, 152.873 }
+		};
+		palletPos[] = {
+				{ { 13.2031, 2.67773, -5.67609 }, 202 }
+		};
+	};
+	class Land_houseV_2T2 : Default {
+		chairPos[] = {
+				{ { -3.60352, 1.40381, -4.05566 }, 60.0702 }
+		};
+		pelicanPos[] = {
+				{ { 2.81348, -4.28516, -1.45676 }, 39.1084 }
+		};
+		shoeboxPos[] = {
+				{ { -3.23193, 0.272461, -4.06165 }, 128.845 }
+		};
+	};
+
+	//Cherno End//
+
+	//// Taki Start ///
+
+	class Land_Mil_hangar_EP1 : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 13.5303, -21.2351, -5.37346 }, 179 },
+				{ { -14.126, -21.2678, -5.37346 }, 179 }
+		};
+		bedPos[] = {
+				{ { 14.0625, 22.0742, -5.3957 }, 7 }
+		};
+		couchPos[] = {
+				{ { 11.0576, 22.5518, -5.3945 }, 277 }
+		};
+		filingPos[] = {
+				{ { 14.5166, 19.895, -5.39077 }, 90 }
+		};
+		pelicanPos[] = {
+				{ { -14.792, 12.4836, -5.37346 }, 270 }
+		};
+		tablePos[] = {
+				{ { 10.9473, 20.2598, -5.38906 }, 200 }
+		};
+		lockerPos[] = {
+				{ { 5.74609, 23.2769, -5.38878 }, 1.99997 }
+		};
+		toolRackPos[] = {
+				{ { -1.87793, 23.8599, -5.38821 }, 0 },
+				{ { 3.83594, 23.8569, -5.38976 }, 0 },
+				{ { -4.74805, 23.8733, -5.38261 }, 0 }
+		};
+		palletPos[] = {
+				{ { -13.4512, -7.20068, -5.37346 }, 270 },
+				{ { -13.5586, 9.13232, -5.37346 }, 270 },
+				{ { 14.1084, -11.3062, -5.37346 }, 270 },
+				{ { 14.123, 12.8301, -5.37681 }, 270 }
+		};
+		freezerPos[] = {
+				{ { 7.4668, 23.3438, -5.39095 }, 1.99997 }
+		};
+	};
+	class Land_Mil_ControlTower_EP1 : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { 3.03467, 2.64453, -1.16014 }, 0 },
+				{ { 3.44043, 3.19238, -5.52014 }, 180 },
+				{ { 10.312, 5.80469, -9.65014 }, 90.8 }
+		};
+		chairPos[] = {
+				{ { 7.3042, -1.58887, 2.98985 }, 270 }
+		};
+		filingPos[] = {
+				{ { 1.48389, 2.47754, -1.08014 }, 350 },
+				{ { 2.1123, 2.47852, -1.08014 }, 0 },
+				{ { 4.2417, 3.28027, -5.52014 }, 180 },
+				{ { -0.137207, 6.33496, -9.65014 }, 270.8 }
+		};
+		pelicanPos[] = {
+				{ { 3.23828, 0.329102, 2.73985 }, 270 },
+				{ { 6.84424, -1.66113, -1.08014 }, 180 },
+				{ { 6.38379, 5.06836, -1.11015 }, 180 }
+		};
+		tablePos[] = {
+				{ { 7.49512, 0.382813, -1.16014 }, 90 }
+		};
+		lockerPos[] = {
+				{ { 4.08936, 6.23438, -9.65014 }, 360 }
+		};
+		shoeboxPos[] = {
+				{ { 6.70996, 3.96582, -1.08014 }, 180 }
+		};
+		palletPos[] = {
+				{ { 8.53418, 4.22461, -9.65014 }, 90.8 }
+		};
+	};
+	class Land_House_K_1_EP1 : Default {
+		fridgePos[] = {
+				{ { 1.96729, 5.96875, 1.53209 }, 360 }
+		};
+		bedPos[] = {
+				{ { -3.29297, 2.14844, 1.53209 }, 187 }
+		};
+		cookerPos[] = {
+				{ { 4.28125, 4.49707, 1.53209 }, 90 }
+		};
+		chairPos[] = {
+				{ { 1.74805, 1.2832, 1.45209 }, 7.99997 }
+		};
+		tablePos[] = {
+				{ { 1.18652, 2.95605, 1.54209 }, 179 }
+		};
+		shoeboxPos[] = {
+				{ { -3.5874, 5.99609, 1.93209 }, 265 }
+		};
+	};
+	class Land_House_K_7_EP1 : Default {
+		shelfPos[] = {
+				{ { -4.74609, 2.42871, 3.25069 }, 0 }
+		};
+		bedPos[] = {
+				{ { -2.32715, 3.7876, 3.23069 }, 273 }
+		};
+		couchPos[] = {
+				{ { -7.00684, 2.23486, 6.33068 }, 180 }
+		};
+		cookerPos[] = {
+				{ { -3.0957, 4.38916, -0.309311 }, 0 }
+		};
+		chairPos[] = {
+				{ { -5.16113, 0.416992, 3.25069 }, 220 }
+		};
+		filingPos[] = {
+				{ { -6.82422, 0.91748, -0.279312 }, 180 }
+		};
+		pelicanPos[] = {
+				{ { -3.28223, -1.62939, 3.13068 }, 180 }
+		};
+		toolRackPos[] = {
+				{ { -7.16309, 3.74512, -0.27932 }, 0.600037 }
+		};
+		shoeboxPos[] = {
+				{ { 0.185547, 2.03467, -0.26931 }, 90 }
+		};
+		palletPos[] = {
+				{ { 0.12793, 3.51074, 6.31068 }, 180 },
+				{ { -6.24219, -1.57227, -0.479317 }, 0.600037 }
+		};
+		freezerPos[] = {
+				{ { -4.16406, 4.40723, -0.309311 }, 0 }
+		};
+	};
+	class Land_House_K_8_EP1 : Default {
+		shelfPos[] = {
+				{ { -2.79102, -1.12793, 0.246132 }, 180 },
+				{ { 3.0918, -1.73926, 0.246132 }, 180 }
+		};
+		fridgePos[] = {
+				{ { 0.333008, 3.72217, 3.35612 }, 90 }
+		};
+		bedPos[] = {
+				{ { -2.45508, -0.746582, -2.63387 }, 183.6 }
+		};
+		couchPos[] = {
+				{ { 0.733398, -0.199707, 0.246132 }, 90 }
+		};
+		wardrobePos[] = {
+				{ { 2.45313, -3.42871, -2.63387 }, 178.6 }
+		};
+		cookerPos[] = {
+				{ { -2.93066, -0.961914, 3.35612 }, 270.9 }
+		};
+		chairPos[] = {
+				{ { 0.383789, -3.7002, -2.70387 }, 14.6 }
+		};
+		filingPos[] = {
+				{ { -3.12109, 2.6582, 0.246132 }, 270 }
+		};
+		pelicanPos[] = {
+				{ { -4.36133, 1.78027, 0.0861282 }, 89.9 }
+		};
+		tablePos[] = {
+				{ { 0.0976563, 3.64941, 0.246132 }, 0 }
+		};
+		shoeboxPos[] = {
+				{ { 3.11035, -0.191406, -1.81387 }, 90 },
+				{ { -2.76172, 4.86816, 0.276131 }, 90 }
+		};
+		palletPos[] = {
+				{ { -0.880859, 2.4375, 5.95612 }, 90 },
+				{ { 2.48926, -3.10742, 0.336128 }, 90 }
+		};
+		freezerPos[] = {
+				{ { 2.14258, 1.05176, -2.62387 }, 90 },
+				{ { -2.89063, 0.299316, 3.35612 }, 270.9 }
+		};
+	};
+	class Land_House_K_3_EP1 : Default {
+		shelfPos[] = {
+				{ { -6.08594, -1.3042, -0.833199 }, 270 }
+		};
+		fridgePos[] = {
+				{ { 2.39648, 3.49902, -0.523193 }, 93 }
+		};
+		bedPos[] = {
+				{ { -0.318359, 4.32129, 2.7168 }, 180 }
+		};
+		couchPos[] = {
+				{ { 1.73438, -0.850586, -0.833199 }, 0 }
+		};
+		cookerPos[] = {
+				{ { -2.74316, 5.83447, -0.523193 }, 270 }
+		};
+		filingPos[] = {
+				{ { -0.918945, 4.41162, -0.5532 }, 273 }
+		};
+		shoeboxPos[] = {
+				{ { 1.85449, -3.16846, 0.0268021 }, 90 }
+		};
+		palletPos[] = {
+				{ { -1.43359, -2.14697, 3.0168 }, 180 }
+		};
+	};
+	class Land_House_K_5_EP1 : Default {
+		bedPos[] = {
+				{ { 3.78613, 0.527832, 2.3232 }, 184 }
+		};
+		cookerPos[] = {
+				{ { -2.3457, 1.2207, 1.5132 }, 188 }
+		};
+		chairPos[] = {
+				{ { 5.12695, 3.30615, 2.3232 }, 64 }
+		};
+		filingPos[] = {
+				{ { -7.18164, -0.956055, 0.493202 }, 198 }
+		};
+		tablePos[] = {
+				{ { 0.817383, 1.22754, 1.5232 }, 181 }
+		};
+		toolRackPos[] = {
+				{ { -7.52832, 1.88867, 0.5532 }, 271 }
+		};
+		shoeboxPos[] = {
+				{ { -5.31934, 2.41211, 0.823196 }, 47 }
+		};
+	};
+	class Land_House_K_6_EP1 : Default {
+		shelfPos[] = {
+				{ { -0.28418, 5.72656, 1.47783 }, 0 },
+				{ { 5.16602, -2.67188, 4.40265 }, 98 }
+		};
+		fridgePos[] = {
+				{ { -0.412109, 5.45508, -1.62217 }, 0 }
+		};
+		bedPos[] = {
+				{ { 0.00683594, 0.544434, 1.47783 }, 186 },
+				{ { 0.00390625, 0.682617, -1.62217 }, 181 },
+				{ { 1.93848, -2.26953, 4.40265 }, 98 },
+				{ { 0.609375, 4.77393, 4.40783 }, 276 }
+		};
+		couchPos[] = {
+				{ { -2.81738, 0.666504, 1.47783 }, 0 },
+				{ { -4.17773, 5.6582, -1.63217 }, 95 }
+		};
+		wardrobePos[] = {
+				{ { 3.08594, 3.81934, -1.63217 }, 91 }
+		};
+		cookerPos[] = {
+				{ { 0.499023, 5.37695, -1.62217 }, 0 }
+		};
+		chairPos[] = {
+				{ { -2.29883, 3.50537, 4.40265 }, 48 }
+		};
+		filingPos[] = {
+				{ { 4.90918, 5.43604, 4.40265 }, 48 }
+		};
+		tablePos[] = {
+				{ { 2.20313, -2.98926, -1.62217 }, 180 },
+				{ { -4.65527, 0.294922, 1.49783 }, 276 }
+		};
+		shoeboxPos[] = {
+				{ { 2.70117, 2.21973, 1.47783 }, 180 }
+		};
+		palletPos[] = {
+				{ { -2.96582, -2.4082, -1.60735 }, 48 }
+		};
+		freezerPos[] = {
+				{ { 1.55273, 5.40479, -1.62217 }, 0 }
+		};
+	};
+	class Land_A_Mosque_small_1_EP1 : Default {
+		couchPos[] = {
+				{ { 6.44385, 2.74609, -2.0341 }, 2 }
+		};
+		chairPos[] = {
+				{ { 1.59668, -1.55762, -1.7141 }, 357 },
+				{ { 3.95605, -1.53418, -1.7141 }, 2 }
+		};
+		pelicanPos[] = {
+				{ { 1.90137, 4.70996, -1.2041 }, 357 }
+		};
+		toolRackPos[] = {
+				{ { -0.34375, -1.05762, -1.7041 }, 357 }
+		};
+		shoeboxPos[] = {
+				{ { 2.07715, 0.0810547, -1.1841 }, 180 }
+		};
+	};
+	class Land_A_Mosque_small_2_EP1 : Default {
+		tablePos[] = {
+				{ { -0.952637, 2.83398, -2.48343 }, 1.52588e-005 }
+		};
+		toolRackPos[] = {
+				{ { -0.880371, -3.7002, -2.37343 }, 1.52588e-005 }
+		};
+		shoeboxPos[] = {
+				{ { 0.770508, -3.40137, -1.69343 }, 1.52588e-005 },
+				{ { -2.49805, -3.33887, -1.69343 }, 1.52588e-005 }
+		};
+	};
+	class Land_House_L_1_EP1 : Default {
+		cookerPos[] = {
+				{ { 0.925293, 0.850586, -0.629326 }, 355 }
+		};
+		tablePos[] = {
+				{ { 0.134766, -1.86328, -0.629326 }, 185 }
+		};
+		shoeboxPos[] = {
+				{ { 2.03125, -1.08301, -0.689323 }, 180 }
+		};
+	};
+	class Land_House_L_3_EP1 : Default {
+		couchPos[] = {
+				{ { -4.16162, 1.71777, 2.56979 }, 278 }
+		};
+		pelicanPos[] = {
+				{ { -5.42822, 0.783203, -0.190216 }, 296 }
+		};
+		tablePos[] = {
+				{ { -3.4624, 2.94727, -0.27021 }, 0 }
+		};
+		shoeboxPos[] = {
+				{ { -0.625488, 3.21289, 0.559784 }, 323 }
+		};
+		palletPos[] = {
+				{ { 0.00488281, 0.0605469, -0.27021 }, 267 },
+				{ { 2.24951, 2.31934, -0.27021 }, 323 }
+		};
+	};
+	class Land_Misc_Well_L_EP1 : Default {
+		shoeboxPos[] = {
+				{ { 0.611328, -0.0751953, 0.135979 }, 278 }
+		};
+	};
+	class Land_House_L_4_EP1 : Default {
+		shelfPos[] = {
+				{ { -6.77441, -0.744141, -1.33775 }, 270 }
+		};
+		fridgePos[] = {
+				{ { 4.71045, -0.0517578, -1.10775 }, 180 }
+		};
+		bedPos[] = {
+				{ { 0.240723, 2.24023, -1.33775 }, 0 }
+		};
+		cookerPos[] = {
+				{ { 6.61816, 1.50781, -1.10775 }, 90 }
+		};
+		chairPos[] = {
+				{ { 0.864258, -0.628906, -1.27775 }, 120 }
+		};
+		tablePos[] = {
+				{ { -3.76367, 1.76953, -1.33775 }, 0 }
+		};
+		shoeboxPos[] = {
+				{ { 4.05518, 3.79492, -0.687752 }, 180 }
+		};
+		palletPos[] = {
+				{ { -4.59863, 0.929688, 1.72225 }, 0 },
+				{ { 4.38574, 2.06152, 1.44225 }, 0 }
+		};
+		freezerPos[] = {
+				{ { 2.82861, 2.85352, -1.03775 }, 270 }
+		};
+	};
+	class Land_House_L_8_EP1 : Default {
+		fridgePos[] = {
+				{ { -1.23047, 0.239258, -1.10685 }, 173 }
+		};
+		couchPos[] = {
+				{ { 4.48242, 0.755859, 1.59315 }, 270 }
+		};
+		cookerPos[] = {
+				{ { 1.48291, 5.91797, -1.10685 }, 90 }
+		};
+		chairPos[] = {
+				{ { 1.6333, 3.23535, 1.59315 }, 109 }
+		};
+		pelicanPos[] = {
+				{ { 5.0166, 3.38574, 1.79315 }, 170 }
+		};
+		tablePos[] = {
+				{ { -3.3335, 2.53906, -1.10685 }, 267 }
+		};
+		toolRackPos[] = {
+				{ { 3.69043, 1.83301, -1.10685 }, 357 }
+		};
+		shoeboxPos[] = {
+				{ { 4.80859, 4.89551, -0.726845 }, 180 }
+		};
+		palletPos[] = {
+				{ { -2.80469, 3.25488, 1.59315 }, 109 },
+				{ { -2.77539, -3.80273, 1.29315 }, 109 }
+		};
+		freezerPos[] = {
+				{ { 1.80957, 2.24609, -1.10685 }, 83 }
+		};
+	};
+	class Land_House_L_6_EP1 : Default {
+		shelfPos[] = {
+				{ { 7.42285, 2.09961, 1.28293 }, 90 }
+		};
+		fridgePos[] = {
+				{ { 1.31592, -2.31445, -1.50707 }, 180 }
+		};
+		bedPos[] = {
+				{ { 6.03418, 1.56738, -1.54707 }, 274 }
+		};
+		couchPos[] = {
+				{ { -3.1333, -1.29785, -1.54707 }, 184 }
+		};
+		wardrobePos[] = {
+				{ { 3.08252, 0.0664063, -1.54707 }, 274 }
+		};
+		chairPos[] = {
+				{ { -3.73486, 2.18555, -1.54707 }, 330 }
+		};
+		filingPos[] = {
+				{ { 7.31006, 1.4082, 1.28293 }, 90 }
+		};
+		lockerPos[] = {
+				{ { 7.1626, 0.398438, 1.28293 }, 90 }
+		};
+		shoeboxPos[] = {
+				{ { 1.97461, 1.40332, -0.347069 }, 79 }
+		};
+		palletPos[] = {
+				{ { 0.405762, 1.2666, 1.18293 }, 90 },
+				{ { -3.02832, -1.68848, 1.18293 }, 90 }
+		};
+	};
+	class Land_House_L_7_EP1 : Default {
+		shelfPos[] = {
+				{ { -2.28809, -2.28809, -0.939278 }, 87.8875 }
+		};
+		fridgePos[] = {
+				{ { -5.4751, 3.71777, -0.639275 }, 1.52588e-005 }
+		};
+		bedPos[] = {
+				{ { 3.79248, 2.42188, -0.269272 }, 270 }
+		};
+		wardrobePos[] = {
+				{ { 5.65527, 0.724609, -0.269272 }, 90 }
+		};
+		cookerPos[] = {
+				{ { -4.51172, 3.5752, -0.609276 }, 1.52588e-005 }
+		};
+		chairPos[] = {
+				{ { -0.388672, 3.22656, -0.269272 }, 290 }
+		};
+		filingPos[] = {
+				{ { -5.10986, -2.08984, -0.939278 }, 350 }
+		};
+		tablePos[] = {
+				{ { 1.99707, -1.52637, -0.239273 }, 180 }
+		};
+		shoeboxPos[] = {
+				{ { -5.84033, -7.11621, 0.0807266 }, 180 }
+		};
+		palletPos[] = {
+				{ { -0.26416, -6.77637, -1.06927 }, 180 }
+		};
+		freezerPos[] = {
+				{ { -2.71533, -6.71777, -0.909271 }, 180 }
+		};
+	};
+	class Land_House_C_1_EP1 : Default {
+		shelfPos[] = {
+				{ { -4.81006, 1.55078, -0.937073 }, 0 }
+		};
+		fridgePos[] = {
+				{ { -8.3042, -0.625977, -0.937073 }, 270 }
+		};
+		cookerPos[] = {
+				{ { -8.23047, 0.216797, -0.937073 }, 270 }
+		};
+		chairPos[] = {
+				{ { -2.89355, -0.30957, -0.937073 }, 80 }
+		};
+		filingPos[] = {
+				{ { -5.55566, 1.49609, -0.937073 }, 0 },
+				{ { -6.02295, 1.42188, -0.937073 }, 350 }
+		};
+		pelicanPos[] = {
+				{ { 4.00537, -1.68848, 3.61293 }, 90 }
+		};
+		tablePos[] = {
+				{ { 8.03906, -0.804688, -0.937073 }, 90 }
+		};
+		lockerPos[] = {
+				{ { -1.39746, -0.689453, -1.01707 }, 270 }
+		};
+		toolRackPos[] = {
+				{ { 5.82568, 1.81348, -0.937073 }, 0 }
+		};
+		shoeboxPos[] = {
+				{ { 7.98096, -4.17969, -0.247078 }, 0 }
+		};
+		palletPos[] = {
+				{ { 1.82275, -3.33984, 3.51293 }, 90 },
+				{ { -6.08203, 0.685547, 3.61293 }, 90 },
+				{ { 6.38721, 0.277344, -0.947075 }, 90 }
+		};
+	};
+	class Land_A_Minaret_Porto_EP1 : Default {
+		chairPos[] = {
+				{ { 1.95752, 0.271484, 9.38245 }, 349 }
+		};
+		pelicanPos[] = {
+				{ { 2.31885, 1.9375, 6.58245 }, 229 },
+				{ { 3.31494, -0.510742, -3.39755 }, 113 }
+		};
+		shoeboxPos[] = {
+				{ { 2.27197, -2.94531, 9.39245 }, 349 }
+		};
+		palletPos[] = {
+				{ { 1.25195, 0.685547, -20.5975 }, 100 }
+		};
+	};
+	class Land_House_C_2_EP1 : Default {
+		shelfPos[] = {
+				{ { -4.66211, -0.90918, -2.31334 }, 266.8 }
+		};
+		bedPos[] = {
+				{ { 1.67383, -0.817383, 0.763084 }, 90 },
+				{ { -3.94336, 1.43262, 0.603081 }, 180 }
+		};
+		couchPos[] = {
+				{ { 3.84131, -3.79883, 0.786659 }, 271.8 }
+		};
+		wardrobePos[] = {
+				{ { 5.73877, 2.78125, 0.786659 }, 90 }
+		};
+		chairPos[] = {
+				{ { -2.6167, -1.23535, 0.603081 }, 170 },
+				{ { -4.00146, -1.05664, 0.603081 }, 220 }
+		};
+		filingPos[] = {
+				{ { 5.95459, -0.53418, -2.31334 }, 104 },
+				{ { 6.07373, 2.55176, -2.31334 }, 80 }
+		};
+		tablePos[] = {
+				{ { 2.34131, 2.0918, 0.763084 }, 1.52588e-005 }
+		};
+		lockerPos[] = {
+				{ { -1.74707, 2.39355, -2.31334 }, 89.8 }
+		};
+		toolRackPos[] = {
+				{ { -1.31445, 0.375, -2.31334 }, 89.8 }
+		};
+		shoeboxPos[] = {
+				{ { -0.350098, 2.94238, -2.28334 }, 180 },
+				{ { 5.8457, -1.50684, 1.57308 }, 180 }
+		};
+		palletPos[] = {
+				{ { -2.55518, -0.786133, -2.31334 }, 110.8 },
+				{ { -3.98584, 1.06055, -2.31334 }, 89.8 }
+		};
+		freezerPos[] = {
+				{ { 2.02002, 0.932617, -2.31334 }, 270 }
+		};
+	};
+	class Land_Misc_Well_C_EP1 : Default {
+		shoeboxPos[] = {
+				{ { 0.225098, 0.355469, 0.0676193 }, 170 }
+		};
+	};
+	class Land_House_C_1_v2_EP1 : Default {
+		fridgePos[] = {
+				{ { -8.29102, 0.782227, -1.35837 }, 270 }
+		};
+		couchPos[] = {
+				{ { 4.57666, -1.00195, 3.16163 }, 270 },
+				{ { -5.02783, 2.07422, -1.38837 }, 270 }
+		};
+		cookerPos[] = {
+				{ { -2.74414, 1.1748, -1.38837 }, 90 }
+		};
+		chairPos[] = {
+				{ { 1.5708, -1.56641, 3.16163 }, 320 },
+				{ { 3.31299, -1.47852, -1.35837 }, 40 },
+				{ { 6.86572, -0.859375, 3.16163 }, 30 }
+		};
+		filingPos[] = {
+				{ { 6.08252, 2.9502, -1.35837 }, 1.52588e-005 }
+		};
+		tablePos[] = {
+				{ { 4.56787, -2.96875, 3.16163 }, 180 },
+				{ { 7.90332, 1.12305, -1.35837 }, 90 }
+		};
+		shoeboxPos[] = {
+				{ { -8.38037, -0.833008, -0.93837 }, 270 }
+		};
+		freezerPos[] = {
+				{ { 5.20557, 2.90332, -1.35837 }, 1.52588e-005 }
+		};
+	};
+	class Land_House_C_5_V1_EP1 : Default {
+		bedPos[] = {
+				{ { -1.40234, -0.685547, -1.48295 }, 0 }
+		};
+		couchPos[] = {
+				{ { -3.50098, -0.208008, -1.48295 }, 180 }
+		};
+		wardrobePos[] = {
+				{ { 0.958984, 1.01758, 1.42705 }, 180 }
+		};
+		chairPos[] = {
+				{ { -2.87109, 3.71875, -1.48295 }, 0 },
+				{ { 2.50684, 4.94287, -1.48295 }, 0 }
+		};
+		filingPos[] = {
+				{ { 3.30664, 1.04639, -1.48295 }, 180 }
+		};
+		tablePos[] = {
+				{ { -2.96191, -4.95508, -1.48295 }, 180 }
+		};
+		palletPos[] = {
+				{ { -3.12891, 0.384277, 1.15705 }, 250 },
+				{ { -1.73926, -4.29883, 1.15705 }, 180 }
+		};
+	};
+	class Land_House_C_3_EP1 : Default {
+		shelfPos[] = {
+				{ { -2.11621, -1.87402, 1.91428 }, 265 }
+		};
+		fridgePos[] = {
+				{ { 5.75879, 3.84033, -3.87573 }, 0 }
+		};
+		bedPos[] = {
+				{ { -7.35449, -1.97705, 0.514275 }, 178 }
+		};
+		couchPos[] = {
+				{ { 4.89453, -0.709961, -0.485725 }, 178 },
+				{ { 4.43359, -2.51904, 4.04428 }, 178 }
+		};
+		wardrobePos[] = {
+				{ { -8.02637, 1.10938, 0.514275 }, 268 }
+		};
+		cookerPos[] = {
+				{ { 4.0166, -1.12305, -3.87573 }, 270 }
+		};
+		chairPos[] = {
+				{ { -4.25684, 3.56934, 0.514275 }, 40 },
+				{ { 6.07715, 0.655273, -3.88573 }, 339 }
+		};
+		filingPos[] = {
+				{ { 8.17969, -2.87256, -3.87573 }, 130 }
+		};
+		pelicanPos[] = {
+				{ { 2.22266, -1.98633, 4.91428 }, 88 }
+		};
+		tablePos[] = {
+				{ { -4.11621, -1.11475, 0.514275 }, 90 },
+				{ { 4.09082, 1.39307, -3.87573 }, 270 }
+		};
+		lockerPos[] = {
+				{ { -8.1543, 2.12549, -3.81573 }, 269 }
+		};
+		toolRackPos[] = {
+				{ { -8.44434, 0.203125, -3.81573 }, 269 }
+		};
+		shoeboxPos[] = {
+				{ { -3.90039, -3.58984, -3.83573 }, 270 },
+				{ { -8.08105, 3.58545, 1.58427 }, 0 }
+		};
+		palletPos[] = {
+				{ { 6.55469, -1.37988, 4.11428 }, 178 },
+				{ { -6.69141, 2.4873, 4.31428 }, 178 }
+		};
+		freezerPos[] = {
+				{ { 4.03809, -2.17529, -3.87573 }, 270 }
+		};
+	};
+	class Land_House_C_12_EP1 : Default {
+		shelfPos[] = {
+				{ { 3.25391, -5.70898, 0.10778 }, 91 },
+				{ { 7.60742, 2.00098, -3.52222 }, 90 }
+		};
+		fridgePos[] = {
+				{ { -3.77637, -8.55176, 0.137779 }, 271 }
+		};
+		couchPos[] = {
+				{ { 3.50244, -2.55566, 0.137779 }, 110 },
+				{ { -0.932129, -7.9502, -3.52222 }, 90 }
+		};
+		cookerPos[] = {
+				{ { 3.01709, -8.48047, -3.52222 }, 91 }
+		};
+		chairPos[] = {
+				{ { 2.45898, -8.51855, 0.10778 }, 141 },
+				{ { -3.55371, -8.52734, -3.52222 }, 221 }
+		};
+		filingPos[] = {
+				{ { 4.61963, -4.02148, -3.52222 }, 180 }
+		};
+		pelicanPos[] = {
+				{ { -4.01465, -5.48633, 3.38778 }, 269 }
+		};
+		lockerPos[] = {
+				{ { -0.582031, -3.9248, -3.52222 }, 180 }
+		};
+		toolRackPos[] = {
+				{ { 1.37695, -4.39941, -3.52222 }, 180 },
+				{ { 3.40576, -4.42773, -3.52222 }, 180 }
+		};
+		shoeboxPos[] = {
+				{ { 1.32813, 0.0283203, -3.52222 }, 88 }
+		};
+		palletPos[] = {
+				{ { -1.87988, 1.13379, 0.137779 }, 230 },
+				{ { -2.9873, 0.550781, -3.48222 }, 88 },
+				{ { 5.65186, 0.637695, 0.137779 }, 110 }
+		};
+		freezerPos[] = {
+				{ { 3.05566, -5.78223, -3.50222 }, 88 }
+		};
+	};
+	class Land_House_C_5_V3_EP1 : Default {
+		shelfPos[] = {
+				{ { 0.881836, 0.848633, 1.42705 }, 180 }
+		};
+		fridgePos[] = {
+				{ { 2.83398, 5.10059, -1.48295 }, 0.999985 }
+		};
+		bedPos[] = {
+				{ { -1.33838, -0.811523, -1.48295 }, 0.999985 }
+		};
+		wardrobePos[] = {
+				{ { 4.41846, 1.59082, 1.42705 }, 90 },
+				{ { -2.9209, 3.87598, -1.48295 }, 1.00179e-005 }
+		};
+		chairPos[] = {
+				{ { -3.97998, 1.91797, -1.48295 }, 289 },
+				{ { -1.0791, -4.83398, 1.12706 }, 130 }
+		};
+		filingPos[] = {
+				{ { -4.45752, -5.25781, -1.49295 }, 269 }
+		};
+		shoeboxPos[] = {
+				{ { 4.48828, 1.01172, -1.40295 }, 1.00179e-005 },
+				{ { 0.500977, 4.9834, -0.672951 }, 30 }
+		};
+		palletPos[] = {
+				{ { -1.72168, -1.9375, 1.17705 }, 1.00179e-005 }
+		};
+	};
+	class Land_House_C_5_V2_EP1 : Default {
+		shelfPos[] = {
+				{ { 0.882813, 0.848145, 1.42705 }, 180 }
+		};
+		fridgePos[] = {
+				{ { 2.83398, 5.1001, -1.48295 }, 1 }
+		};
+		bedPos[] = {
+				{ { -1.33789, -0.811523, -1.48295 }, 1 }
+		};
+		wardrobePos[] = {
+				{ { -2.85547, 3.81055, -1.38295 }, 360 }
+		};
+		chairPos[] = {
+				{ { -3.86914, 1.87793, -1.39295 }, 320 },
+				{ { -1.0752, -4.77148, 1.32706 }, 130 }
+		};
+		filingPos[] = {
+				{ { -4.45703, -5.25781, -1.49295 }, 269 }
+		};
+		shoeboxPos[] = {
+				{ { 4.48828, 1.01172, -1.40295 }, 0 },
+				{ { 0.507813, 5.10352, -0.572945 }, 130 }
+		};
+		palletPos[] = {
+				{ { -1.72168, -1.9375, 1.17705 }, 0 }
+		};
+	};
+	class Land_Mil_Guardhouse_EP1 : Default {
+		lootType = "mil";
+		shelfPos[] = {
+				{ { -2.9021, -0.569336, -1.69236 }, 180 }
+		};
+		shoeboxPos[] = {
+				{ { -2.94189, 3.83496, -1.69236 }, 180 }
+		};
+	};
+	class Land_Mil_House_EP1 : Default {
+		lootType = "mil";
+		toolRackPos[] = {
+				{ { 10.4844, 2.79492, -5.67967 }, 270 }
+		};
+		palletPos[] = {
+				{ { 12.9678, 2.14258, -5.67967 }, 180 },
+				{ { 13.7964, 4.08594, -5.67967 }, 220 }
+		};
+	};
+	class Land_Mil_Repair_center_EP1 : Default {
+		lootType = "mil";
+		toolRackPos[] = {
+				{ { 3.14893, 3.31836, -2.52654 }, 270 }
+		};
+		palletPos[] = {
+				{ { -2.75195, 0.446289, -1.52654 }, 180 },
+				{ { -2.63037, 3.28516, -1.52654 }, 270 }
+		};
+	};
+	class Land_Misc_Cargo1Ao_EP1 : Default {
+		palletPos[] = {
+				{ { 0.109863, -0.517578, -1.05425 }, 180 },
+				{ { 0.0776367, 2.01465, -1.05425 }, 180 }
+		};
+	};
+	class Land_Mil_Barracks_i_EP1 : Default {
+		pelicanPos[] = {
+				{ { -0.560303, -1.47656, 0.0566254 }, 88 },
+				{ { 2.65845, -1.47168, -0.743378 }, 108 },
+				{ { -3.68701, -1.2627, 0.0566254 }, 100 },
+				{ { -7.04419, -1.5918, -0.743378 }, 100 }
+		};
+		lockerPos[] = {
+				{ { 0.652588, -1.77637, -1.10338 }, 270 },
+				{ { -5.80908, -1.9834, -1.10338 }, 270 },
+				{ { -9.33643, -2.05957, -1.10338 }, 270 }
+		};
+		toolRackPos[] = {
+				{ { -9.41846, 2.02148, -1.10338 }, 270 }
+		};
+		shoeboxPos[] = {
+				{ { -3.3457, -0.0791016, -1.11338 }, 108 },
+				{ { 2.53003, -2.44336, -1.08338 }, 108 },
+				{ { 4.00269, -1.91797, -0.083374 }, 290 },
+				{ { -10.4619, 1.45313, 0.296623 }, 90 }
+		};
+	};
+	class Land_Ind_Oil_Pump_EP1 : Default {
+		shoeboxPos[] = {
+				{ { 1.8894, -2.20605, 2.53654 }, 90 }
+		};
+	};
+	class Land_Ind_Oil_Tower_EP1 : Default {
+		couchPos[] = {
+				{ { 3.39233, -5.47461, 3.74753 }, 270 }
+		};
+		pelicanPos[] = {
+				{ { 3.07495, -2.38086, 16.4475 }, 270 }
+		};
+		toolRackPos[] = {
+				{ { -1.66919, 0.807617, -13.0425 }, 270 }
+		};
+		palletPos[] = {
+				{ { 3.64136, 0.131836, -8.04248 }, 30 },
+				{ { -1.99316, -6.2207, -13.0225 }, 270 },
+				{ { 4.45459, -5.50098, -12.9225 }, 270 },
+				{ { 5.33862, -6.28711, -8.04248 }, 30 }
+		};
+	};
+	class Land_House_L_9_EP1 : Default {
+		bedPos[] = {
+				{ { 1.09448, 1.40039, -0.611443 }, 275 }
+		};
+	};
+	class Land_fortified_nest_big_EP1 : Default {
+		shelfPos[] = {
+				{ { -0.359375, -0.337891, -0.721558 }, 95 }
+		};
+		couchPos[] = {
+				{ { 1.25781, -0.414063, -0.721558 }, 185 }
+		};
+		pelicanPos[] = {
+				{ { -1.12036, 0.962891, -0.721558 }, 85 }
+		};
+	};
+	class Land_A_Mosque_big_wall_EP1 : Default {
+		shelfPos[] = {
+				{ { 4.53857, 4.61914, -2.04966 }, 350 }
+		};
+		couchPos[] = {
+				{ { 0.288574, 3.51367, -2.04966 }, 270 }
+		};
+	};
+	class Land_A_Mosque_big_wall_corner_EP1 : Default {
+		couchPos[] = {
+				{ { -1.2417, 3.2666, -2.03097 }, 270 }
+		};
+		chairPos[] = {
+				{ { 1.03223, 3.80176, -2.03097 }, 20 }
+		};
+		palletPos[] = {
+				{ { 3.88086, -2.67383, -2.03097 }, 20 }
+		};
+	};
+	class Land_A_Mosque_big_minaret_1_EP1 : Default {
+		pelicanPos[] = {
+				{ { -0.907227, -1.69336, 7.16152 }, 280 }
+		};
+		shoeboxPos[] = {
+				{ { 0.124023, -0.527344, 7.16152 }, 4.89996 }
+		};
+	};
+	class Land_A_Mosque_big_addon_EP1 : Default {
+		tablePos[] = {
+				{ { 5.93506, -2.10156, -6.46501 }, 90 },
+				{ { 5.84326, 2.80566, -6.46501 }, 90 }
+		};
+		shoeboxPos[] = {
+				{ { 2.78027, -5.18164, -5.965 }, 360 },
+				{ { 2.78857, 5.36914, -5.965 }, 360 }
+		};
+	};
+	class Land_A_Mosque_big_hq_EP1 : Default {
+		couchPos[] = {
+				{ { 5.84619, -2.89648, -9.35999 }, 41 },
+				{ { 8.43701, 8.60352, 0.520004 }, 150 }
+		};
+		chairPos[] = {
+				{ { -6.1582, -2.72852, -9.33999 }, 249 },
+				{ { -5.3042, -4.38379, -9.33999 }, 239 },
+				{ { 9.77832, -7.29395, 0.520004 }, 320 },
+				{ { 8.51709, -9.12793, 0.520004 }, 300 },
+				{ { 9.05273, -10.5879, -4.05999 }, 340 }
+		};
+		pelicanPos[] = {
+				{ { 2.45459, 9.20313, 0.520004 }, 340 }
+		};
+		tablePos[] = {
+				{ { 5.96045, 3.45703, -9.36999 }, 49 }
+		};
+		shoeboxPos[] = {
+				{ { -9.6084, -10.9482, -4.05999 }, 41 }
+		};
+		palletPos[] = {
+				{ { -5.69385, -8.7002, 0.520004 }, 320 },
+				{ { -8.27783, 7.9248, 0.520004 }, 70 }
+		};
+	};
+	class Land_A_Mosque_big_minaret_2_EP1 : Default {
+		pelicanPos[] = {
+				{ { -1.92383, -1.92188, 3.61214 }, 80 }
+		};
+		shoeboxPos[] = {
+				{ { -0.0258789, -0.837891, 3.60214 }, 9.99997 }
+		};
+	};
+	class Land_House_C_11_EP1 : Default {
+		shelfPos[] = {
+				{ { 4.04883, -0.507813, -2.07991 }, 180 }
+		};
+		couchPos[] = {
+				{ { 0.967773, -3.72266, -2.07991 }, 89.9999 },
+				{ { 5.21924, -3.53809, 0.940094 }, 93 }
+		};
+		wardrobePos[] = {
+				{ { 3.57471, 2.49023, 1.04009 }, 265.8 }
+		};
+		cookerPos[] = {
+				{ { 4.06836, -1.44141, -2.07991 }, 0 }
+		};
+		chairPos[] = {
+				{ { 7.38379, -1.58594, 0.940094 }, 43 }
+		};
+		filingPos[] = {
+				{ { 2.47314, -4.53809, 0.940094 }, 179 },
+				{ { 2.99756, -4.46582, 0.940094 }, 173 }
+		};
+		toolRackPos[] = {
+				{ { 6.90332, -0.748047, -2.07991 }, 180 }
+		};
+		palletPos[] = {
+				{ { 0.493652, -1.9248, 0.960091 }, 179 },
+				{ { 1.26855, 4.18555, -2.45991 }, 185.8 }
+		};
+		freezerPos[] = {
+				{ { -1.55713, -2.66016, -2.07991 }, 272 }
+		};
+	};
+	class Land_House_C_9_EP1 : Default {
+		shelfPos[] = {
+				{ { 0.668945, -4.66406, -3.85899 }, 180 }
+		};
+		fridgePos[] = {
+				{ { 3.50098, 2.56934, -3.82899 }, 356 }
+		};
+		bedPos[] = {
+				{ { 1.76123, 4.54102, -0.278992 }, 92.1 }
+		};
+		couchPos[] = {
+				{ { 1.69629, -4.80762, -0.278992 }, 90 }
+		};
+		cookerPos[] = {
+				{ { 2.63574, 2.44336, -3.82899 }, 356 }
+		};
+		chairPos[] = {
+				{ { -4.98926, 0.46582, -0.278992 }, 262.1 }
+		};
+		filingPos[] = {
+				{ { 5.79541, -5.37012, -0.278992 }, 94 },
+				{ { 5.7334, -5.89941, -0.278992 }, 100 }
+		};
+		pelicanPos[] = {
+				{ { -5.09424, -1.93359, -0.278992 }, 94 }
+		};
+		lockerPos[] = {
+				{ { 2.80615, 2.62598, -0.278992 }, 0 }
+		};
+		toolRackPos[] = {
+				{ { 1.40625, 6.66602, -3.85899 }, 0 }
+		};
+		shoeboxPos[] = {
+				{ { -2.22803, 0.438477, -2.99899 }, 360 }
+		};
+		palletPos[] = {
+				{ { 0.508301, -2.66895, 2.53101 }, 144 },
+				{ { -3.3208, 1.07031, 2.53101 }, 144 },
+				{ { -1.89307, 4.89941, 2.53101 }, 94 }
+		};
+		freezerPos[] = {
+				{ { 5.66406, 0.332031, -3.83899 }, 89 }
+		};
+	};
+	class Land_Fort_Watchtower_EP1 : Default {
+		pelicanPos[] = {
+				{ { -0.390137, 0.0634766, 0.56675 }, 144 }
+		};
+		shoeboxPos[] = {
+				{ { -1.28271, 2.67969, -2.21325 }, 360 }
+		};
+		palletPos[] = {
+				{ { -0.77832, -2.00293, -2.20325 }, 144 }
+		};
+	};
+	class Land_Misc_Cargo1Bo_EP1 : Default {
+		palletPos[] = {
+				{ { 0.0737305, -0.0507813, -1.06068 }, 184 },
+				{ { 0.0981445, -1.95898, -1.06068 }, 184 },
+				{ { -0.0107422, 1.99707, -1.06068 }, 184 }
+		};
+	};
+	class Land_A_Minaret_EP1 : Default {
+		pelicanPos[] = {
+				{ { -1.50244, -0.742188, 5.06781 }, 123.5 }
+		};
+	};
+	//Taki End//
+
+
+
+	// BornHolm Start //
+	class Land_bo_HouseV2_04_Interier : Land_HouseV2_04_interier {};
+	class Land_bo_HouseV2_02_Interier : Land_HouseV2_02_Interier {};
+	class Land_bo_HouseV_1I4 : Land_HouseV_1I4 {};
+	class Land_bo_Panelak2 : Land_Panelak2 {};
+	class Land_bo_Panelak : Land_Panelak {};
+	class Land_bo_Shed_Ind02 : Land_Shed_Ind02 {};
+	class Land_bo_Hlidac_budka : Land_Hlidac_budka {};
+	class Land_bo_A_GeneralStore_01 : Land_A_GeneralStore_01 {};
+	class Land_bo_A_Office01 : Land_A_Office01 {};
+	class Land_bo_Barn_W_01 : Land_Barn_W_01 {};
+	class Land_bo_stodola_open : Land_Stodola_open {};
+	class Land_bo_stodola_old_open : Land_Stodola_old_open {};
+	class Land_bo_Farm_Cowshed_a : Land_Farm_Cowshed_a {};
+	class Land_bo_Farm_Cowshed_b : Land_Farm_Cowshed_b {};
+	class Land_bo_Farm_Cowshed_c : Land_Farm_Cowshed_c {};
+	class Land_bo_A_TVTower_Base : Land_A_TVTower_Base {};
+	class Land_bo_HouseV_2T2 : Land_houseV_2T2 {};
+	class Land_bo_HouseB_Tenement : Land_HouseB_Tenement {};
+	class Land_bo_HouseV2_01A : Land_HouseV2_01A {};
+	class Land_bo_Barn_W_02 : Land_Barn_W_02 {};
+	class Land_bo_Shed_W01 : Land_Shed_W01 {};
+	class Land_bo_A_Castle_Bergfrit : Land_A_Castle_Bergfrit {};
+	class Land_bo_Shed_wooden : Land_Shed_wooden {};
+	class Land_bo_HouseV_1L2 : Land_HouseV_1L2 {};
+	class Land_bo_HouseV_1I1 : Land_HouseV_1I1 {};
+	class Land_bo_HouseV_1L1 : Land_HouseV_1L1 {};
+	class Land_bo_misc_deerstand : Land_Misc_deerstand {};
+	class Land_bo_HouseBlock_A1_1 : Land_HouseBlock_A1_1 {};
+	class Land_bo_HouseBlock_A3 : Land_HouseBlock_A3 {};
+	class Land_bo_hut06 : Land_Hut06 {};
+
+
+	class Land_bo_dum_rasovna : Default {
+		shelfPos[] = {
+				{ { -1.56299, -4.67383, -2.69722 }, 185 },
+				{ { 3.88159, -3.92139, -2.69722 }, 90 }
+		};
+		fridgePos[] = {
+				{ { -1.78345, 1.78662, -2.70722 }, 275.214 }
+		};
+		bedPos[] = {
+				{ { 2.92578, 2.73438, -2.70722 }, 1.99997 }
+		};
+		wardrobePos[] = {
+				{ { 0.281494, -0.0371094, 0.232777 }, 261 }
+		};
+		cookerPos[] = {
+				{ { -1.76978, -1.74072, -2.69722 }, 356 }
+		};
+		chairPos[] = {
+				{ { 3.47705, -0.510742, -2.69722 }, 101 }
+		};
+		filingPos[] = {
+				{ { 0.485352, -1.99902, 0.232777 }, 0 },
+				{ { -0.095459, -2.09375, 0.232777 }, 354 }
+		};
+		pelicanPos[] = {
+				{ { -0.732422, 3.80762, 0.202777 }, 329 }
+		};
+		tablePos[] = {
+				{ { 0.746094, 4.27881, -2.67722 }, 360 }
+		};
+		toolRackPos[] = {
+				{ { -0.824707, -1.08301, -2.69722 }, 182 }
+		};
+		shoeboxPos[] = {
+				{ { -1.76807, -4.44385, 0.232777 }, 44 }
+		};
+		freezerPos[] = {
+				{ { 0.0822754, -1.64063, -2.69722 }, 0.999985 }
+		};
+	};
+	class Land_bo_Nasypka : Default {
+		shoeboxPos[] = {
+				{ { -1.81689, -0.786621, 6.56536 }, 213 }
+		};
+	};
+	class Land_TTowerBig_2_F : Default {
+		pelicanPos[] = {
+				{ { 1.03027, 1.00903, 0.757229 }, 41 }
+		};
+	};
+	class Land_bo_Sara_domek_sedy : Default {
+		fridgePos[] = {
+				{ { 0.498047, 1.38989, -2.14814 }, 270 }
+		};
+		bedPos[] = {
+				{ { -1.31494, 0.212891, -2.13814 }, 4 },
+				{ { -5.03027, 3.2019, -2.14814 }, 183 }
+		};
+		couchPos[] = {
+				{ { 4.9209, 1.72778, -2.12814 }, 5 }
+		};
+		wardrobePos[] = {
+				{ { -4.1416, -0.853271, -2.14814 }, 182 }
+		};
+		cookerPos[] = {
+				{ { 1.86279, 4.21411, -2.13814 }, 0 }
+		};
+		chairPos[] = {
+				{ { -2.82324, -0.666504, -2.13814 }, 170 },
+				{ { -5.49121, -0.572998, -2.13814 }, 200 }
+		};
+		filingPos[] = {
+				{ { 6.03662, -1.00024, -2.13814 }, 97 }
+		};
+		tablePos[] = {
+				{ { -3.23926, 4.17554, -2.11814 }, 0 }
+		};
+		shoeboxPos[] = {
+				{ { 0.770508, -0.9021, -2.14814 }, 161 }
+		};
+		freezerPos[] = {
+				{ { 0.799805, 4.23804, -2.11814 }, 0 }
+		};
+	};
+	class Land_bo_Ind_SawMill : Default {
+		toolRackPos[] = {
+				{ { 9.0686, 5.12646, -5.83683 }, 270 }
+		};
+		palletPos[] = {
+				{ { -7.46265, -5.69971, -5.83683 }, 142 },
+				{ { 7.71216, -10.9746, -5.83683 }, 290 },
+				{ { -3.50049, 16.1409, -5.83683 }, 142 },
+				{ { 5.49878, 16.4333, -5.83683 }, 40 }
+		};
+	};
+	class Land_bo_Dum_mesto2 : Default {
+		shelfPos[] = {
+				{ { -0.732422, 3.7876, -4.33935 }, 179 },
+				{ { 3.63647, -6.50879, -4.34044 }, 91 },
+				{ { 2.20728, -8.04883, -4.34044 }, 175 }
+		};
+		fridgePos[] = {
+				{ { -3.60547, 6.22559, -4.34935 }, 269 }
+		};
+		bedPos[] = {
+				{ { -2.60498, -1.65088, -0.949348 }, 270 },
+				{ { -2.51392, -7.01904, -0.970444 }, 90 },
+				{ { -2.25952, 7.11084, -0.949348 }, 272 }
+		};
+		couchPos[] = {
+				{ { -2.60498, -2.21436, -4.33935 }, 187 },
+				{ { -0.407959, -6.73047, -4.35044 }, 94 }
+		};
+		wardrobePos[] = {
+				{ { 0.216064, 3.9668, -0.949348 }, 180 }
+		};
+		cookerPos[] = {
+				{ { -3.6228, 7.73877, -4.34935 }, 269 }
+		};
+		chairPos[] = {
+				{ { 3.14673, 4.25439, -4.33935 }, 119 },
+				{ { 3.18701, 7.4751, -0.970444 }, 20 },
+				{ { 3.13452, 7.49902, -4.33935 }, 60 },
+				{ { 3.17065, -7.49072, -0.970444 }, 119 },
+				{ { -3.35864, -7.4707, -4.34044 }, 220 }
+		};
+		filingPos[] = {
+				{ { 3.42749, -7.72949, -4.34044 }, 121 }
+		};
+		tablePos[] = {
+				{ { -0.68042, -0.654297, -4.35044 }, 0 },
+				{ { -3.53101, -3.4668, -0.970444 }, 270 },
+				{ { -3.49878, -5.11523, -4.34044 }, 271 },
+				{ { -0.356445, 7.76953, -4.33935 }, 359 }
+		};
+		toolRackPos[] = {
+				{ { -2.99243, 0.0361328, -4.34044 }, 181 }
+		};
+		shoeboxPos[] = {
+				{ { 1.50415, -0.472656, -4.35044 }, 207 },
+				{ { -3.74072, 0.48584, -4.34935 }, 90 }
+		};
+		freezerPos[] = {
+				{ { -2.49731, 0.472168, -4.34935 }, 178 }
+		};
+	};
+	class Land_bo_HouseV_3I4 : Default {
+		shoeboxPos[] = {
+				{ { 5.35449, 3.04028, -2.60763 }, 280 }
+		};
+	};
+	class Land_bo_Dum_mesto3 : Default {
+		palletPos[] = {
+				{ { 5.67578, -0.140137, -5.89724 }, 250 }
+		};
+	};
+	class Land_bo_HouseV_1T : Default {
+		shoeboxPos[] = {
+				{ { -4.95142, 6.46045, -3.09068 }, 130 }
+		};
+	};
+	class Land_bo_Ind_Stack_Big : Default {
+		shelfPos[] = {
+				{ { 0.137939, 0.352539, -28.5862 }, 270 }
+		};
+		pelicanPos[] = {
+				{ { -2.85278, 0.0292969, -23.3062 }, 290 }
+		};
+		shoeboxPos[] = {
+				{ { -1.56006, -2.29297, -28.5862 }, 260 }
+		};
+	};
+
+	class Land_bo_Deutshe_mini : Default {
+		shelfPos[] = {
+				{ { 3.57666, -1.0437, -2.74453 }, 0 }
+		};
+		fridgePos[] = {
+				{ { -0.42334, -0.501709, -2.75453 }, 90 }
+		};
+		bedPos[] = {
+				{ { -3.61914, 2.44849, -2.74453 }, 89 }
+		};
+		couchPos[] = {
+				{ { 3.54443, 0.63208, -2.75453 }, 90 }
+		};
+		wardrobePos[] = {
+				{ { -0.290039, 2.47217, -2.74453 }, 89 }
+		};
+		chairPos[] = {
+				{ { 0.885254, 2.83154, -2.75453 }, 310 }
+		};
+		filingPos[] = {
+				{ { 0.569336, -3.05029, -2.74453 }, 259 }
+		};
+		shoeboxPos[] = {
+				{ { -3.51465, -1.95581, -1.98453 }, 209 }
+		};
+		freezerPos[] = {
+				{ { -4.28613, 0.733887, -2.74453 }, 359 }
+		};
+	};
+	class Land_u_House_Small_01_V1_F : Default {
+		fridgePos[] = { { { 0.867188, 4.2207, -0.931187 }, 81.6768 } };
+		shelfPos[] = { { { 1.16602, -1.34766, -0.929543 }, 86.5322 }, { { -2.17383, 1.22656, -0.930841 }, 1.30102 } };
+		wardrobePos[] = { { { 0.298828, -3.75586, -0.929366 }, 164.186 } };
+		bedPos[] = { { { -3.52344, -2.83594, -0.931075 }, 183.389 } };
+		couchPos[] = { { { 2.90234, -1.70313, -0.929456 }, 179.44 } };
+		cookerPos[] = { { { -1.41602, 2.18359, -0.931049 }, 180.625 } };
+		chairPos[] = { { { 2.46484, 1.8125, -0.930511 }, 270.894 }, { { -4.09766, -0.542969, -0.925941 }, 270.019 }, { { 2.42383, 3.95703, -0.931158 }, 329.895 } };
+	};
+
+
+	// Bornholm End
+
+
+	class Land_TTowerBig_1_F : Default {
+		pelicanPos[] = {
+				{ { 0.920898, 0.00317383, 2.89364 }, 270 }
+		};
+	};
+	class Land_u_Addon_01_V1_F : Default {
+		shoeboxPos[] = {
+				{ { 5.52515, 0.521484, -0.014668 }, 270 }
+		};
+		palletPos[] = {
+				{ { -0.227783, 0.866699, -0.0246677 }, 270 }
+		};
+	};
+	class Land_SlideCastle_F : Default {
+		shoeboxPos[] = {
+				{ { 0.0908203, 2.18799, -1.94149 }, 193 }
+		};
+		palletPos[] = {
+				{ { -0.841309, -3.05469, -1.94149 }, 193 }
+		};
+	};
+	class Land_Shed_Big_F : Default {
+		palletPos[] = {
+				{ { 6.66553, 4.03369, -10 }, 177 },
+				{ { -7.18652, -10.4651, -10 }, 177 },
+				{ { -5.99561, 11.2573, -10 }, 177 }
+		};
+	};
+	class Land_Radar_F : Default {
+		pelicanPos[] = {
+				{ { 9.36694, 4.97949, -4.86368 }, 247 }
+		};
+		shoeboxPos[] = {
+				{ { -9.92847, -8.72461, -4.86368 }, 44 }
+		};
+	};
+	class Land_ReservoirTower_F : Default {
+		pelicanPos[] = {
+				{ { -1.41553, 1.15063, -1.81942 }, 130 }
+		};
+		shoeboxPos[] = {
+				{ { -3.8374, 1.81421, -1.81942 }, 320 }
+		};
+		palletPos[] = {
+				{ { 0.737793, -3.30981, -1.80942 }, 250 }
+		};
+	};
+	class Land_i_Stone_Shed_V2_dam_F : Default {
+		tablePos[] = {
+				{ { 3.07764, 2.17139, 0.0703425 }, 11 }
+		};
+		shoeboxPos[] = {
+				{ { -2.41211, 0.305664, 0.660343 }, 11 }
+		};
+	};
+	class Land_Cargo_HQ_V1_ruins_F : Default {
+		lootType = "mil";
+		filingPos[] = {
+				{ { 0.798828, -3.87012, -1.28783 }, 250 }
+		};
+		pelicanPos[] = {
+				{ { -4.96582, -4.61182, 1.13217 }, 250 }
+		};
+		lockerPos[] = {
+				{ { -5.49561, -3.40796, -1.28783 }, 270 }
+		};
+	};
+	class Land_cmp_Tower_F : Default {
+		palletPos[] = {
+				{ { 0.0551758, 1.01294, -7.64906 }, 250 }
+		};
+	};
+	class Land_Crane_F : Default {
+		pelicanPos[] = {
+				{ { -7.11475, -1.31665, 5.74081 }, 90 }
+		};
+		shoeboxPos[] = {
+				{ { -6.86523, 1.83765, 5.74081 }, 230 }
+		};
+	};
+
+	class Land_WIP_ruins_F : Default {
+		shelfPos[] = {
+				{ { 2.08887, -10.4897, -1.32377 }, 180.2 },
+				{ { -17.3643, -11.3889, 2.47623 }, 2 }
+		};
+		pelicanPos[] = {
+				{ { 16.8608, -8.03442, 0.106226 }, 266 }
+		};
+		shoeboxPos[] = {
+				{ { 3.64453, 7.38965, 0.126227 }, 350 }
+		};
+		palletPos[] = {
+				{ { 15.3921, 8.71045, 2.38622 }, 350 },
+				{ { -21.7849, 11.9456, 2.45622 }, 350 }
+		};
+	};
+	class Land_i_Stone_HouseBig_V2_dam_F : Default {
+		fridgePos[] = {
+				{ { -3.23608, -3.12744, -1.70506 }, 220 }
+		};
+		bedPos[] = {
+				{ { -2.40332, 0.169434, 1.17494 }, 180 }
+		};
+		wardrobePos[] = {
+				{ { 2.80518, -3.146, 1.17494 }, 170 }
+		};
+		shoeboxPos[] = {
+				{ { -0.989014, 1.07861, -0.595063 }, 20 }
+		};
+		freezerPos[] = {
+				{ { 2.25952, 0.718262, -1.63506 }, 20 }
+		};
+	};
+	class Land_u_House_Big_02_V1_dam_F : Default {
+		shelfPos[] = {
+				{ { 3.21387, -1.56421, -2.44914 }, 90 }
+		};
+		fridgePos[] = {
+				{ { -3.24658, -3.44507, -2.44914 }, 260 }
+		};
+		couchPos[] = {
+				{ { 3.49512, 0.929443, -2.44234 }, 90.0046 }
+		};
+		cookerPos[] = {
+				{ { -3.28662, -1.65698, -2.44914 }, 273 }
+		};
+		chairPos[] = {
+				{ { 4.87988, 4.62207, -2.43914 }, 66 }
+		};
+		filingPos[] = {
+				{ { -3.41895, -0.0644531, 0.940861 }, 260 },
+				{ { -3.42285, 0.570068, 0.940861 }, 270 }
+		};
+		pelicanPos[] = {
+				{ { 4.70313, 4.34985, 0.870861 }, 55 }
+		};
+		tablePos[] = {
+				{ { -3.23535, 2.41211, -2.43914 }, 270 },
+				{ { -3.29639, 2.79614, 0.940861 }, 270 }
+		};
+		lockerPos[] = {
+				{ { 2.81396, 0.00561523, 0.940861 }, 181 }
+		};
+		toolRackPos[] = {
+				{ { -3.71387, 2.81152, 0.940861 }, 270 }
+		};
+		shoeboxPos[] = {
+				{ { 5.19873, -0.0888672, 0.940861 }, 160 }
+		};
+	};
+	class Land_Cargo_HQ_V2_ruins_F : Default {
+		shelfPos[] = {
+				{ { -5.3606, -5.2832, -1.15784 }, 270 }
+		};
+		pelicanPos[] = {
+				{ { 3.18213, -3.80811, -1.30784 }, 160 }
+		};
+	};
+	class Land_i_Stone_HouseSmall_V2_dam_F : Default {
+		couchPos[] = {
+				{ { -6.25098, 1.86572, -0.594842 }, 270 }
+		};
+		filingPos[] = {
+				{ { 8.93213, -2.67773, -0.594842 }, 176 }
+		};
+		pelicanPos[] = {
+				{ { -8.89941, -2.2583, -0.644842 }, 260 }
+		};
+		tablePos[] = {
+				{ { -3.21045, -2.5564, -0.594842 }, 180 }
+		};
+		shoeboxPos[] = {
+				{ { 1.99609, 2.25684, -0.274842 }, 320 }
+		};
+	};
+	class Land_cmp_Shed_dam_F : Default {
+		palletPos[] = {
+				{ { 0.482422, -2.13721, -3.10882 }, 329 },
+				{ { -3.94287, 1.30518, -3.10882 }, 260 }
+		};
+	};
+	class Land_Cargo_Tower_V1_ruins_F : Default {
+		pelicanPos[] = {
+				{ { 3.30298, 3.41357, -1.80046 }, 329 }
+		};
+	};
+
+	class Land_dp_mainFactory_F : Default {
+		lootBias = 100;
+		shelfPos[] = { { { 12.6934, -1.76367, -4.24943 }, 359.454 }, { { 13.9473, 5.0918, -4.24943 }, 88.7339 }, { { 19.9414, 1.03125, -7.43432 }, 269.954 }, { { 19.9355, -2.62891, -7.43443 }, 271.718 } };
+		pelicanPos[] = { { { -2.5332, 1.13477, 1.24025 }, 333.827 }, { { -11.8926, -2.99023, -4.39045 }, 104.689 }, { { 7.52734, -16.5059, -0.257848 }, 243.576 } };
+		toolRackPos[] = { { { 14.1328, 0.527344, -4.33314 }, 90.5367 } };
+	};
+
+
+
+
+	class Land_i_Windmill01_F : Default { pelicanPos[] = { { { -0.910156, 1.28125, -3.54742 }, 289.979 } };                                   tablePos[] = { { { 0.111328, -1.23047, -3.50127 }, 181.768 } }; };
+	class Land_d_House_Big_01_V1_F : Default { fridgePos[] = { { { -5.62891, 6.35156, -2.19721 }, 269.825 } };     shelfPos[] = { { { -0.492188, 1.82031, 1.22279 }, 163.832 }, { { 0.046875, 5.28125, -2.19721 }, 359.157 }, { { 2.66797, 8.1875, -2.19721 }, 122.214 } };     pelicanPos[] = { { { 0.826172, 0.384766, 4.54293 }, 312.651 } };     wardrobePos[] = { { { -4.50781, 7.97852, 1.22285 }, 181.679 } };     bedPos[] = { { { 1.57227, 8.62891, 1.22277 }, 96.399 } };    couchPos[] = { { { -5.1582, 3.05859, -2.28722 }, 352.256 }, { { -4.88086, 4.26758, 1.2291 }, 179.049 }, { { -4.02344, 10.252, -2.19723 }, 254.622 } };     cookerPos[] = { { { -1.58008, 6.79102, -2.19759 }, 3.80292 } };     chairPos[] = { { { -0.201172, 0.566406, 1.23667 }, 21.1155 }, { { -5.29688, 8.16406, -2.19722 }, 259.465 }, { { 2.27734, 11.0098, 1.22278 }, 48.7725 }, { { -5.27539, 10.9121, 1.22279 }, 288.988 } };           tablePos[] = { { { -0.554688, 11.1719, -2.19723 }, 0.0478998 }, { { -1.80469, 11.25, 1.22271 }, 358.754 } }; };
+	class Land_Sara_hasic_zbroj : Default {
+		lootBiasPos = 20;
+		fridgePos[] = { { { 2.29688, -2.56055, -2.67931 }, 181.134 } };
+		shelfPos[] = { { { -3.21436, 2.60376, -2.67923 }, 3.40094 } };
+		couchPos[] = { { { 0.914063, 1.76172, -2.67932 }, 266.479 } };
+		chairPos[] = { { { 6.72656, 1.56543, -2.67931 }, 88.7719 } };
+		filingPos[] = { { { 7.02832, 2.61157, -2.67928 }, 59.8147 } };
+		tablePos[] = { { { 6.98535, -0.0251465, -2.67932 }, 92.6988 } };
+		toolRackPos[] = { { { 7.46631, -1.93945, -2.67892 }, 89.8589 } };
+		shoeboxPos[] = { { { -3.5083, -2.67578, -2.67946 }, 263.378 } };
+		palletPos[] = { { { 3.6665, -0.135742, -2.67929 }, 42.8049 } };
+		freezerPos[] = { { { 1.00684, -2.44629, -2.67932 }, 171.556 } };
+	};
+	class Land_Sara_domek_zluty : Default {
+		lootBiasPos = 20;
+		fridgePos[] = { { { 2.29858, 3.92969, -2.46014 }, 0.136902 } };
+		shelfPos[] = { { { 7.44385, 1.58789, -2.45992 }, 97.2181 } };
+		wardrobePos[] = { { { -2.31445, -3.1958, -2.4576 }, 90.1546 } };
+		bedPos[] = { { { -5.8606, -2.66113, -2.46011 }, 272.354 } };
+		couchPos[] = { { { -6.51392, 1.26953, -2.45143 }, 180.056 } };
+		cookerPos[] = { { { 4.7998, 3.88477, -2.46054 }, 0.147873 } };
+		chairPos[] = { { { 6.95801, -0.59668, -2.45824 }, 129.461 } };
+		filingPos[] = { { { -1.29907, 4.00537, -2.45987 }, 359.897 } };
+		tablePos[] = { { { -2.49561, 3.39307, -2.45992 }, 90.2344 } };
+		toolRackPos[] = { { { 0.131104, 4.27246, -2.45967 }, 2.38939 } };
+		shoeboxPos[] = { { { 4.66724, -0.834961, -2.46021 }, 90.2617 } };
+		freezerPos[] = { { { 2.16602, 1.70801, -2.44791 }, 269.45 } };
+	};
+	class Land_Misc_PowerStation : Default {
+		lootBiasPos = 20;
+		shelfPos[] = { { { 3.3877, -0.776367, -1.29452 }, 268.884 } };
+		lockerPos[] = { { { 3.64697, -4.25073, -1.29468 }, 270.752 } };
+		toolRackPos[] = { { { 3.20557, 5.78613, -1.33632 }, 269.596 } };
+		shoeboxPos[] = { { { -5.59229, 1.92871, -0.186357 }, 90.4582 } };
+	};
+	class Land_A_CraneCon : Default {
+		pelicanPos[] = { { { -7.55273, 1.50537, 5.76163 }, 28.6669 } };
+	};
+	class Land_A_BuildingWIP : Default {
+		lootBiasPos = 20;
+		fridgePos[] = { { { 2.03711, 11.1831, -6.49124 }, 305.356 }, { { -3.37573, 12.2183, 1.48757 }, 305.492 }, { { -24.4121, 7.55957, -2.55824 }, 228.742 } };
+		shelfPos[] = { { { -5.80249, -5.0332, -2.55812 }, 356.037 }, { { 12.1133, -4.3042, -6.54318 }, 179.13 }, { { -21.0371, 1.47168, -6.46016 }, 178.598 } };
+		pelicanPos[] = { { { -7.40625, -9.28955, -6.50476 }, 210.36 }, { { -20.8247, -11.5005, 5.44375 }, 191.623 }, { { -24.6577, 11.9438, 1.45556 }, 293.971 } };
+		bedPos[] = { { { -13.3503, 2.34912, 1.47256 }, 91.8744 } };
+		couchPos[] = { { { -4.99951, -3.28564, -6.43938 }, 90.7108 }, { { 0.594727, 9.11768, 5.4154 }, 148.196 }, { { 7.01099, 10.9351, -6.42719 }, 225.517 } };
+		cookerPos[] = { { { 3.14282, -10.0327, 1.45506 }, 208.035 }, { { -3.64502, 12.6104, -2.55869 }, 359.938 } };
+		chairPos[] = { { { 2.41162, 6.28955, 5.44231 }, 221.29 }, { { 3.62207, -9.91504, -2.56085 }, 177.773 }, { { -0.971924, 11.6572, 5.4412 }, 260.795 } };
+		filingPos[] = { { { -10.2527, -10.1606, -6.51096 }, 127.045 }, { { -16.0825, -6.36279, -6.53292 }, 282.391 }, { { -19.718, -6.7124, -6.54208 }, 87.3864 } };
+		tablePos[] = { { { 1.70239, -3.92871, -6.48336 }, 180.567 }, { { 3.09351, 9.49219, 5.38227 }, 48.7191 }, { { 22.1931, 1.64893, -6.55645 }, 90.8045 } };
+		lockerPos[] = { { { 13.3423, -1.85791, -6.54581 }, 268.445 }, { { 18.0183, -1.11426, -6.55955 }, 89.6342 }, { { -20.8777, 0.597656, -6.46828 }, 359.864 } };
+		toolRackPos[] = { { { -9.04224, 6.72314, -6.52912 }, 357.538 }, { { -15.5667, -10.561, -2.5611 }, 180.234 }, { { -15.666, -14.3716, 1.45426 }, 179.938 } };
+		shoeboxPos[] = { { { 3.03394, 7.42188, -2.55826 }, 176.357 }, { { -7.98242, -10.2471, -2.56173 }, 129.945 }, { { -16.2637, -5.07031, -6.53438 }, 305.636 } };
+		palletPos[] = { { { 5.09961, 0.015625, -6.47375 }, 270.66 }, { { -10.7837, -0.00585938, 5.26227 }, 0.0174896 }, { { -14.4409, 11.9912, -6.45401 }, 359.986 }, { { -24.2383, -9.16699, -2.55823 }, 359.993 } };
+		freezerPos[] = { { { -15.1563, -5.16211, -2.55844 }, 359.99 }, { { -24.2168, 12.3633, -6.50327 }, 0.00699433 } };
+	};
+	class Land_A_MunicipalOffice : Default {
+		lootBiasPos = 20;
+		pelicanPos[] = { { { 3.55029, -5.18726, -0.195372 }, 283.895 } };
+		chairPos[] = { { { 4.35693, -5.59839, -18.4382 }, 65.3313 }, { { -4.6333, -5.60132, -18.4559 }, 295.001 } };
+		tablePos[] = { { { -3.02002, -7.53784, -18.4554 }, 177.998 } };
+		shoeboxPos[] = { { { 4.2832, -8.80542, -18.4554 }, 136.305 }, { { 6.69727, -9.01831, -3.18546 }, 279.196 } };
+	};
+	class Land_Ind_Pec_01 : Default {
+		lootBiasPos = 20;
+		fridgePos[] = { { { -2.46875, -0.492676, -3.98998 }, 359.611 } };
+		shelfPos[] = { { { 2.24902, -2.04834, -17.6013 }, 177.989 }, { { -7.08789, -0.345215, -3.98975 }, 359.84 }, { { 5.65918, -7.55225, -17.6013 }, 271.929 }, { { -7.43311, -7.98193, 6.06143 }, 181.606 }, { { -4.87256, -11.603, -17.6013 }, 87.6155 } };
+		pelicanPos[] = { { { -4.57373, -7.05566, 15.777 }, 275.886 } };
+		couchPos[] = { { { -0.665527, -7.01172, -22.6032 }, 319.689 }, { { 5.09033, -8.23584, 1.00238 }, 267.37 } };
+		chairPos[] = { { { -0.938965, -1.05859, 6.05994 }, 148.592 } };
+		filingPos[] = { { { -1.88159, -15.1143, 6.06143 }, 274.02 } };
+		lockerPos[] = { { { -1.07202, -15.2251, -17.6016 }, 178.66 }, { { 2.56372, -15.189, -4.00633 }, 179.789 } };
+		toolRackPos[] = { { { -4.23755, -9.12012, -22.6029 }, 181.414 }, { { 0.57666, -15.625, 1.00255 }, 179.995 } };
+		shoeboxPos[] = { { { -4.63867, -7.03174, 10.4987 }, 320.834 }, { { 0.787842, -9.5957, -14.8937 }, 85.2537 } };
+	};
+	class Land_Ind_SiloVelke_01 : Default {
+		lootBiasPos = 20;
+		shelfPos[] = { { { 0.950195, -2.19312, -5.47868 }, 269.46 }, { { 1.01416, 5.71167, 11.865 }, 268.444 }, { { -11.5195, 2.36597, 11.865 }, 357.551 } };
+		pelicanPos[] = { { { 1.54297, -3.29761, 24.2848 }, 69.2792 } };
+		filingPos[] = { { { 0.224121, -0.709229, 11.8648 }, 98.1473 } };
+		lockerPos[] = { { { -5.82715, 3.93359, 11.8648 }, 359.3 } };
+		toolRackPos[] = { { { 0.811523, 5.54248, -5.47879 }, 272.599 } };
+		shoeboxPos[] = { { { -5.8335, 9.96069, 21.9903 }, 115.472 } };
+	};
+	class Land_Ind_Expedice_1 : Default {
+		lootBiasPos = 20;
+		shelfPos[] = { { { 1.92285, -0.303223, -5.71009 }, 178.232 }, { { 0.276855, 12.5483, -3.51293 }, 203.944 } };
+		pelicanPos[] = { { { -0.432617, 10.3877, 9.68375 }, 299.993 }, { { 5.82813, -18.415, -5.78366 }, 292.046 } };
+		chairPos[] = { { { 9.70801, -12.1228, -5.78367 }, 325.934 } };
+		filingPos[] = { { { 5.53857, -16.9548, -5.78365 }, 270.844 } };
+		tablePos[] = { { { 13.5181, -12.0178, -5.78368 }, 4.86572 } };
+		lockerPos[] = { { { 3.24707, -0.168457, -12.9662 }, 180.509 } };
+		toolRackPos[] = { { { -2.06592, -0.598389, -12.9654 }, 180.956 } };
+		shoeboxPos[] = { { { -8.35303, -7.32837, -3.51317 }, 29.0581 } };
+		palletPos[] = { { { 2.58789, -6.03125, -3.51315 }, 1.43912e-005 }, { { 0.29248, -18.708, -3.51315 }, 1.43571e-005 } };
+	};
+	class Land_Repair_center : Default {
+		lootBiasPos = 20;
+		shelfPos[] = { { { -2.13965, 2.55408, -1.55488 }, 88.8949 } };
+	};
+	class Land_Church_01 : Default {
+		lootBiasPos = 20;
+		chairPos[] = { { { -6.88208, 0.851563, -4.24353 }, 6.01111 } };
+	};
+	class Land_NAV_Lighthouse2 : Default {
+		lootBiasPos = 20;
+		shoeboxPos[] = { { { -0.019043, 1.00098, 2.39268 }, 0.452068 } };
+	};
+	class Land_A_Castle_Gate : Default {
+		fridgePos[] = { { { 7.84082, 4.31982, -2.93036 }, 21.3658 } };
+		bedPos[] = { { { 7.11719, 0.841797, -2.94727 }, 93.608 } };
+		cookerPos[] = { { { 4.0166, 4.29834, -2.97534 }, 0.0874265 } };
+		chairPos[] = { { { 4.8916, -3.36182, -2.96021 }, 149.004 } };
+		filingPos[] = { { { 3.54395, -3.67285, -2.97525 }, 177.859 } };
+		tablePos[] = { { { 5.6875, 4.36865, -2.9592 }, 0 } };
+		toolRackPos[] = { { { 3.1123, 0.0239258, -2.97498 }, 273.35 } };
+	};
+	class Land_A_Castle_Donjon : Default {
+		pelicanPos[] = { { { 2.69727, -2.01953, 12.4763 }, 148.853 } };
+		tablePos[] = { { { -1.54395, 5.67383, 12.4445 }, 0.407131 } };
+	};
+	class Land_A_Castle_Stairs_A : Default {
+		shelfPos[] = { { { 6.63232, 2.28223, 0.998962 }, 356.642 } };
+	};
+	class Land_Ind_Quarry : Default {
+		fridgePos[] = { { { -3.92188, -16.6865, -7.50379 }, 222.564 } };
+		shelfPos[] = { { { -5.69775, -2.96875, -7.5061 }, 89.9894 }, { { -0.436035, -9.52051, -7.50609 }, 358.647 }, { { -4.89355, -8.41016, 5.26832 }, 178.014 } };
+		pelicanPos[] = { { { -5.05762, -3.05859, 5.26825 }, 270.152 } };
+		bedPos[] = { { { 1.77393, -13.1133, -7.49595 }, 252.259 } };
+		couchPos[] = { { { 1.89893, -10.4531, -7.45486 }, 270.008 } };
+		cookerPos[] = { { { -1.72217, -16.9697, -7.50663 }, 173.088 } };
+		chairPos[] = { { { -6.30176, -7.69727, -7.50632 }, 152.678 } };
+		filingPos[] = { { { -3.79004, -8.23633, 5.26811 }, 173.701 } };
+		tablePos[] = { { { -3.25244, -9.7373, -7.48993 }, 0.00806808 } };
+		lockerPos[] = { { { -6.44287, 11.9883, -7.50632 }, 358.505 } };
+		toolRackPos[] = { { { -5.43994, -4.68066, -7.50566 }, 88.4867 } };
+		shoeboxPos[] = { { { -3.70996, 15.8828, 5.26807 }, 106.559 } };
+	};
+	class Land_Barrack2 : Default {
+		shelfPos[] = { { { -1.14063, 5.67578, -0.714874 }, 359.937 } };
+		pelicanPos[] = { { { 0.917969, 1.40234, 0.507111 }, 62.1253 } };
+		chairPos[] = { { { 1.83398, -2.77344, -0.720093 }, 121.012 } };
+		filingPos[] = { { { -0.944336, -3.10254, -0.720245 }, 197.347 } };
+		shoeboxPos[] = { { { 0.393555, 5.52734, -0.71524 }, 0.0174853 } };
+		freezerPos[] = { { { -2.02344, -3.13086, -0.720337 }, 181.294 } };
+	};
+	class Land_Ind_Mlyn_01 : Default {
+		shelfPos[] = { { { -1.02686, 8.37012, -7.73514 }, 84.5525 } };
+		pelicanPos[] = { { { -6.94873, 2.21094, 9.19446 }, 187.886 } };
+		couchPos[] = { { { -7.31543, 7.59277, -20.8141 }, 0 } };
+		lockerPos[] = { { { -8.98828, 1.93848, -2.84921 }, 179.976 } };
+		toolRackPos[] = { { { -9.06104, 1.60645, -8.90338 }, 180.213 } };
+		shoeboxPos[] = { { { -1.24268, 7.89355, 0.201904 }, 123.572 } };
+		palletPos[] = { { { -9.00732, 8.61621, -20.8141 }, 23.6484 } };
+	};
+	class Land_wagon_box : Default {
+		palletPos[] = { { { 0.342285, -0.0380859, -0.80304 }, 0 }, { { 0.271973, 2.38574, -0.80304 }, 354.097 }, { { 0.399414, -2.41504, -0.803101 }, 0 } };
+	};
+	class CampEast_EP1 : Default {
+		shelfPos[] = { { { -1.97852, 2.84473, -1.35312 }, 308.338 }, { { -1.98438, -3.44043, -1.35312 }, 205.956 } };
+		pelicanPos[] = { { { -2.32422, 0.889648, -1.35324 }, 274.054 } };
+		bedPos[] = { { { 1.50342, 1.74512, -1.35336 }, 6.52914 }, { { 1.42969, -2.29004, -1.35336 }, 0.577024 } };
+		shoeboxPos[] = { { { 2.55762, -0.0185547, -1.35336 }, 78.3071 } };
+	};
+	class Land_Barrack2_EP1 : Default {
+		fridgePos[] = { { { 1.29102, -3.15332, -0.720306 }, 178.283 } };
+		shelfPos[] = { { { -0.526855, 5.5127, -0.714996 }, 357.779 } };
+		pelicanPos[] = { { { 1.26465, 3.2002, -0.319916 }, 355.705 } };
+		chairPos[] = { { { -2.10107, -2.88477, -0.720154 }, 187.124 } };
+		filingPos[] = { { { -1.48193, 5.50488, -0.715179 }, 359.955 } };
+		tablePos[] = { { { -2.27539, 3.22363, -0.716583 }, 274.247 } };
+		shoeboxPos[] = { { { 2.25586, 0.362305, -0.0941467 }, 29.4483 } };
+	};
+	class Land_Hlidac_Budka_EP1 : Default {
+		shelfPos[] = { { { 1.20264, 0.0732422, -0.809814 }, 179.636 } };
+		filingPos[] = { { { 2.69629, 2.27246, -0.809937 }, 21.4708 } };
+		tablePos[] = { { { -2.5376, 1.29883, -0.810028 }, 271.096 } };
+		toolRackPos[] = { { { 2.77783, 0.842773, -0.799316 }, 82.7987 } };
+		shoeboxPos[] = { { { 1.61768, 2.3584, -0.810059 }, 356.431 } };
+	};
+	class Land_HouseV_2L : Default {
+		shelfPos[] = { { { 4.29199, 2.60913, -3.95322 }, 7.9453 } };
+		filingPos[] = { { { 4.08594, 1.27734, -3.9534 }, 243.743 } };
+		shoeboxPos[] = { { { -3.90137, -3.46313, -4.01404 }, 65.2182 } };
+		freezerPos[] = { { { -5.02051, -0.899658, -4.10532 }, 87.6906 } };
+	};
+	class Land_Shed_M02 : Default {
+		shoeboxPos[] = { { { 1.05371, 0.0397949, -1.03354 }, 83.8544 } };
+		palletPos[] = { { { -0.0595703, -2.57153, -1.53342 }, 122.62 } };
+	};
+	class ferrisWheel_EPOCH : Default {
+		lootBiasPos = 50;
+		palletPos[] = { { { 2, 2, -10 }, 322.16 }, { { -2, -2, -10 }, 232.16 } };
+		shoeboxPos[] = { { { -3.33862, -0.988281, -10 }, 312.382 } };
+		pelicanPos[] = { { { -0.840332, 1.09473, -10 }, 269.861 } };
+	};
+	class Carnival_Tent : Default {
+		lootBiasPos = 50;
+		shelfPos[] = { { { -1, -1, -10 }, 178.014 } };
+		chairPos[] = { { { 1, 1, -10 }, 152.678 } };
+		filingPos[] = { { { -1, 0, -10 }, 173.701 } };
+		tablePos[] = { { { 0, 0, -10 }, 120.893 } };
+		shoeboxPos[] = { { { 0, 1, -10 }, 106.559 } };
+		palletPos[] = { { { 1, 0, -10 }, 322.16 } };
+	};
+};
